### 2024-04-02 #6569 @editorialbot
[PRE REVIEW]: BESMARTS: A toolkit for data-driven force field design using binary-encoded SMARTS
**Submitting author:** @trevorgokey (<a href="http://orcid.org/0000-0001-7856-1156">Trevor Gokey</a>)
**Repository:** https://github.com/trevorgokey/besmarts
**Branch with paper.md** (empty if default branch): joss-manuscript
**Version:** 0.1.0
