### 2024-04-02 #6569 @Kevin-Mattheus-Moerman
Dear author, @trevorgokey, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again. It may be that no DOI exists, if so perhaps a URL can be added if needed.
