### 2024-04-02 #6569 @trevorgokey
Thanks for the feedback @Kevin-Mattheus-Moerman. I believe I have completed the requests:

- The bib entry without a DOI does already include a URL the corrsponding github repository, please let me know how I should proceed with any fixes.
- I have added some contributing guidelines as CONTRIBUTING.md, adapting from the provided link. Thanks for the suggestion.
