### 2024-04-18 #6569 @Sulstice
Sure, seems up my alley.
