### 2024-04-19 #6569 @Ruibin-Liu
The software cannot be installed following its instructions. Since it is a Python package, I suggest @trevorgokey to make it available in PyPI.
