### 2024-04-20 #6569 @trevorgokey
Thanks for checking this @Ruibin-Liu; I'll work on making it available on PyPI. I believe I have fixed the issue that you may have encountered. I assumed JOSS reviewers would install the branch considered for submission, in this case `joss-manuscript`. In this case, the clone command should be followed by `git checkout joss-manuscript` before the install command.
