### 2024-04-22 #6569 @lucydot
> I assumed JOSS reviewers would install the branch considered for submission, in this case joss-manuscript. In this case, the clone command should be followed by git checkout joss-manuscript before the install command.

@trevorgokey The JOSS review is usually based on the `main` / `master` branch, with any additional `joss-manuscript` branch usually containing the JOSS paper. If you want ` joss-manuscript` to be reviewed that is fine also, and then we'd typically expect it to be merged into `main` once review is complete.
