### 2024-04-22 #6653 @editorialbot
[REVIEW]: BESMARTS: A toolkit for data-driven force field design using binary-encoded SMARTS
**Submitting author:** @trevorgokey (<a href="http://orcid.org/0000-0001-7856-1156">Trevor Gokey</a>)
**Repository:** https://github.com/trevorgokey/besmarts
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0
