### 2024-04-23 #6653 @trevorgokey
> The JOSS review is usually based on the main / master branch, with any additional joss-manuscript branch usually containing the JOSS paper. If you want  joss-manuscript to be reviewed that is fine also, and then we'd typically expect it to be merged into main once review is complete.

@lucydot thanks for the explanation. I am fine with merging the paper into main after review. I am still actively contributing to the codebase and pushing to main, so I intended the `joss-manuscript` branch to be a frozen, versioned codebase to review. I've already made some commits to `main` since submitting this, but while this is under review I'll develop on a different branch to avoid inintentionally breaking something.
