### 2024-04-26 #6653 @Ruibin-Liu
@trevorgokey I can install through the method in the README.md page but cannot run the examples in the RTD. Also, the RTD installation page says the packages can be directly installed through pip. Please make them consistent and ensure the examples can run.
