### 2024-04-29 #6653 @Ruibin-Liu
The error is:
...besmarts-git/besmarts-core/python/setup.py", line 5, in <module>
          with open("README.md", "r") as fh:
               ^^^^^^^^^^^^^^^^^^^^^^
      FileNotFoundError: [Errno 2] No such file or directory: 'README.md'
      [end of output]
