### 2024-04-29 #6653 @trevorgokey
Hi, thanks for indicating the issue. I fixed that error in the pre-review stage when @Ruibin-Liu first reported issues with installing. I just installed the code using a fresh environment and I did not have any issue.

> I can install through the method in the README.md page but cannot run the examples in the RTD
I am noticing that some of the RTD is out of date and I will need to go over it and make it consisent with the package. I worked on the examples in `besmarts/besmarts-core/python/examples` but did not update the docs examples. Thank you @Ruibin-Liu for pointing this out.
