### 2024-05-07 #6653 @lucydot
>  I just installed the code using a fresh environment and I did not have any issue.

Could this confusion result from using different branches? @Ruibin-Liu  - I believe the branch to review is `joss-manuscript` - can you confirm that is correct @trevorgokey ?
