### 2024-05-13 #6653 @Ruibin-Liu
I don't have issue to install manually now but I don't think the RTD claim for pip installation method is correct. Specifically, this line `pip install besmarts-hierarchy besmarts-splitter besmarts-resolve` is incorrect in that none of them is in PyPI.

Other issues in the RTD examples I have checked.

- `from besmarts.codecs import native` doesn't work in the [Graphs](https://besmarts.readthedocs.io/en/latest/graph_io.html#) section.
- The following doesn't work in the [Graphs](https://besmarts.readthedocs.io/en/latest/graph_io.html#) section:
	```python
	for indices in graphs.graph_dihedrals(g):
        print(indices)
	for indices in graphs.graph_impropers(g):
    	print(indices)
	```
- `print(geometry.bond(idx[::-1]))` incorrect because `idx` is `int`

There are more in the RTD doc. I stopped checking. Please make sure those examples can run without any issue.
