### 2024-05-17 #6653 @trevorgokey
I'm out for a conference right now and will work on the documentation
immediately after it ends. I should have it done in the next few days.

On Wed, May 15, 2024, 10:17 AM Lucy Whalley ***@***.***>
wrote:

> Thanks for update @Ruibin-Liu <https://github.com/Ruibin-Liu> -
> @trevorgokey <https://github.com/trevorgokey> it sounds like there is
> some more updates to make to the documentation so examples can be ran
> locally. If you could update here with expected timescales for fixing that
> would be useful -
>
