### 2024-05-21 #6653 @AntObi
Hi @trevorgokey , I've opened up an issue [trevorgokey/besmarts#1](https://github.com/trevorgokey/besmarts/issues/1#issue-2309094892) which contains my detailed breakdown of my current problems that I have found in this repository. A few may overlap @Ruibin-Liu 's comments.

Here is a checklist summary of what needs to be done before I can complete the review on my end, but also look at the issue for full details:

- [ ] Installation instructions in Readthedocs (RTD) should be consistent with README in the main repo
- [ ] Examples/Usage
    - [ ] Make sure all the examples in the repo are functional
    - [ ] Fill out the README in [besmarts/besmarts-core/python/examples](https://github.com/trevorgokey/besmarts/tree/joss-manuscript/besmarts-core/python/examples) to ensure that a new user can fully understand how each example works.
    - [ ] Make sure all the usage examples in the RTD are functional
- [ ] Testing
    - [ ] Apply my suggestion in the linked issue to ensure that `pytest` runs with no issues.
    - [ ] Make sure that the codebase is sufficiently covered with tests
    - [ ]  I would recommend that you set up some form of continous integration (Github Actions, Circle Ci etc) in your repository in order to automatically test your code. This will help when you have new contributors to the package in terms of making sure new commits don't break your package
    - [ ]  A lot of the code is undocumented (many classes and functions lack docstrings). This makes using many functions inaccessible to a someone unfamiliar with the codebase.
