### 2024-05-24 3410e2d
update docs and ff fitting
