+++ besmarts-core/python/besmarts/mechanics/smirnoff_models.py
@@ -36,6 +36,7 @@ def chemical_model_bond_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     proc = mm.chemical_model_procedure_smarts_assignment(
         pcp, cm.topology_terms
     )
+    pid = len(cm.procedures)
     proc.name = f"{cm.name} SMARTS assignment"
 
     proc.unit_hierarchy = hierarchies.structure_hierarchy(
@@ -43,23 +44,29 @@ def chemical_model_bond_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     )
     u = proc.unit_hierarchy.index.node_add_below(None)
     u.name = "u1"
+    uid = u.index
 
     proc.unit_hierarchy.smarts[u.index] = "[*:1]"
 
     h = hierarchies.structure_hierarchy(
         trees.tree_index(), {}, {}, topology.bond
     )
-    proc.smarts_hierarchies = {u.index: h}
+    proc.smarts_hierarchies = {uid: h}
 
     root = proc.smarts_hierarchies[u.index].index.node_add_below(None)
     root.name = "Bonds"
+    root.category = [-1, pid, uid]
+    root.type = "hierarchy"
     proc.smarts_hierarchies[u.index].smarts[root.index] = None
 
+    uid = u.index
     for param in d["parameters"]:
         node = proc.smarts_hierarchies[u.index].index.node_add_below(
             root.index
         )
         node.name = param.get("id", "")
+        node.type = "parameter"
+        node.category = [-1, pid, uid]
         h.smarts[node.index] = param.get("smirks", None)
         # print(f"Loading {node.name} {h.smarts[node.index]}")
 
@@ -76,6 +83,8 @@ def chemical_model_bond_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     cm.procedures.append(proc)
 
     return cm
+
+
 
 def chemical_model_angle_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     cm = mm.chemical_model("A", "Angles", topology.angle)
@@ -97,6 +106,7 @@ def chemical_model_angle_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     proc = mm.chemical_model_procedure_smarts_assignment(
         pcp, cm.topology_terms
     )
+    pid = len(cm.procedures)
     proc.name = f"{cm.name} SMARTS assignment"
 
     proc.unit_hierarchy = hierarchies.structure_hierarchy(
@@ -104,24 +114,30 @@ def chemical_model_angle_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     )
     u = proc.unit_hierarchy.index.node_add_below(None)
     u.name = "u1"
+    uid = 0
 
-    proc.unit_hierarchy.smarts[u.index] = "[*:1]"
+    proc.unit_hierarchy.smarts[uid] = "[*:1]"
 
     h = hierarchies.structure_hierarchy(
         trees.tree_index(), {}, {}, topology.angle
     )
-    proc.smarts_hierarchies = {u.index: h}
+    proc.smarts_hierarchies = {uid: h}
 
     root = proc.smarts_hierarchies[u.index].index.node_add_below(None)
     root.name = "Angles"
+    root.category = [-1, pid, uid]
+    root.type = "hierarchy"
     proc.smarts_hierarchies[u.index].smarts[root.index] = None
 
+    uid = u.index
     for param in d["parameters"]:
 
         node = proc.smarts_hierarchies[u.index].index.node_add_below(
             root.index
         )
         node.name = param.get("id", "")
[…]
