+++ besmarts-core/python/besmarts/core/assignments.py
@@ -41,26 +41,135 @@ ASSN_NAMES = {
     RADII : "radii", # atom, 1x1
 }
 
-gid_t = int
-sid_t = Sequence[int]
-aid_t = int
-did_t = Sequence
+eid_t = int # entry
+tid_t = int # table id
+gid_t = int # graph id
+rid_t = int # row id
+cid_t = int # col id
+sid_t = int # selection id
 
-class graph_topology_db_table:
+pid_t = int # topology id
+xid_t = int
+
+class graph_db_column:
+    def __init__(self):
+        self.selections: Dict[sid_t, List] = {}
+
+    def __getitem__(self, sid: sid_t) -> List:
+        return self.selections[sid]
+
+    def __setitem(self, sid: sid_t, values):
+        self.selections[sid].clear()
+        self.selections[sid].extend(values)
+
+
+class graph_db_row:
+
+    def __init__(self):
+        self.columns: Dict[cid_t, graph_db_column] = {}
+
+    def __getitem__(self, cid: cid_t) -> graph_db_column:
+        return self.columns[cid]
+
+    def __setitem(self, cid: cid_t, columns):
+        self.columns[cid] = columns
+
+class graph_db_graph:
+    def __init__(self):
+        self.rows: Dict[rid_t, graph_db_row] = {}
+
+    def __getitem__(self, rid: cid_t) -> graph_db_row:
+        return self.rows[rid]
+
+    def __setitem(self, rid: rid_t, rows):
+        self.columns[rid] = rows
+
+class graph_db_table:
     """
     Holds the data for a db of graphs that all share the same topology.
     """
-    def __init__(self, topo, sel):
-        self.topology: topology.structure_topology = topo
-        self.selection: Dict[gid_t, Dict[sid_t, did_t]] = sel
+    def __init__(self, top):
+        self.topology: topology.structure_topology = top
+        self.graphs: Dict[gid_t, graph_db_graph] = {}
+        self.values = []
 
-class graph_topology_db:
+    def __getitem__(self, gid: gid_t) -> graph_db_graph:
+        return self.graphs[gid]
+
+    def __setitem(self, gid: rid_t, graph: graph_db_graph):
+        self.graphs[gid] = graph
+
+class graph_db_entry:
     """
-    Aggregation of data of various topology with the graphs
+    Data associated with a set of graphs
     """
     def __init__(self):
+        self.tables: Dict[tid_t, graph_db_table] = {}
+
+    def __getitem__(self, tid: tid_t) -> graph_db_table:
+        return self.tables[tid]
+
+    def __setitem(self, gid: rid_t, table: graph_db_table):
+        self.tables[tid] = table
+
+class graph_db:
+
+    """
+    The global data definitions
+    """
+
+    def __init__(self):
+        self.smiles: Dict[gid_t, str] = {}
         self.graphs: Dict[gid_t, graphs.graph] = {}
-        self.gfuncs: Dict[aid_t, graph_topology_db_table] = {}
+        self.selections: Dict[pid_t, graph_db_selection] = {}
+        self.entries: Dict[eid_t, graph_db_entry] = {}
+
+    def __getitem__(self, eid: eid_t) -> graph_db_entry:
+        return self.entries[eid]
+
+    def __setitem(self, eid: eid_t, entry: graph_db_entry):
+        self.entries[eid] = entry
+
+def graph_db_get_entries(GDB: graph_db, eids: List[eid_t]) -> graph_db:
+    gdb = graph_db()
+
+    gdb.entries = {i: GDB.entries[i] for i in eids}
+    gids = set([gid for t in gdb.entries.values() for r in t.tables.values() for gid in r.graphs])
[…]
