+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -2,6 +2,7 @@
 besmarts.mechanics.fits
 """
 
+import os
 import copy
 import heapq
 import datetime
@@ -9,6 +10,7 @@ import pickle
 import multiprocessing
 import collections
 import sys
+import math
 from typing import List, Dict, Tuple
 from besmarts.core import configs
 from besmarts.core import arrays
@@ -23,6 +25,8 @@ from besmarts.core import splits
 from besmarts.core import trees
 from besmarts.core import tree_iterators
 from besmarts.core import optimization
+from besmarts.core import primitives
+from besmarts.core import logs
 from besmarts.mechanics import optimizers_scipy 
 from besmarts.mechanics import objectives
 from besmarts.mechanics import molecular_models as mm
@@ -68,21 +72,26 @@ class compute_config:
 
         results: List[Dict[assignments.tid_t, assignments.graph_db_table]] = []
 
-        for k, v in self.keys.items():
+        # for k, v in self.keys.items():
             # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.6g} to {v:.6g} d={v-mm.chemical_system_get_value(csys, k):.6g}")
-            mm.chemical_system_set_value(csys, k, v)
+            # mm.chemical_system_set_value(csys, k, v)
             # mm.physical_system_set_value(self.psys, k, v)
 
         for eid, gdb in GDB.entries.items():
             tbl = assignments.graph_db_table(topology.null)
             for gid in gdb.graphs:
                 # pos0 = assignments.graph_db_entry_to_graph_assignment(gdb, tbl_idx, gid)
-                pos0 = assignments.graph_db_graph_to_graph_assignment(gdb, eid, tid, gid)
+                rid = 0
+                pos0 = assignments.graph_db_graph_to_graph_assignment(gdb, eid, tid, gid, rid)
 
                 # this means fold all graphs into a single graph
                 # pos0 = assignments.graph_db_entry_to_graph_assignment(gdb, eid, tid)
 
                 psys = mm.chemical_system_to_physical_system(csys, [pos0], ref=self.psys, reuse=self.reuse)
+                for k, v in self.keys.items():
+                    # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.10g} to {v:.10g} d={v-mm.chemical_system_get_value(csys, k):.10g}")
+                    # mm.chemical_system_set_value(csys, k, v)
+                    mm.physical_system_set_value(psys, k, v)
                 ene = objectives.physical_system_energy(psys, csys)
                 
                 # ene_ga = assignments.graph_assignment("", {(0,): [[energy]]}, gdb.graphs[gid])
@@ -98,7 +107,6 @@ class compute_config:
 class compute_config_energy(compute_config):
 
     """
-    Default is single point energy
     """
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_entry]]:
@@ -108,9 +116,9 @@ class compute_config_energy(compute_config):
         tbl_idx = assignments.ENERGY
         tid = assignments.POSITIONS
 
-        for k, v in self.keys.items():
+        # for k, v in self.keys.items():
             # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.6g} to {v:.6g} d={v-mm.chemical_system_get_value(csys, k):.6g}")
-            mm.chemical_system_set_value(csys, k, v)
+            # mm.chemical_system_set_value(csys, k, v)
 
         results: List[Dict[assignments.tid_t, assignments.graph_db_table]] = []
         for eid, gde in gdb.entries.items():
@@ -121,7 +129,7 @@ class compute_config_energy(compute_config):
                 for gid in gdb.graphs:
[…]
