+++ besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py
@@ -11,6 +11,7 @@ from besmarts.core import geometry
 from besmarts.core import assignments
 from besmarts.core import compute
 from besmarts.core import arrays
+from besmarts.core import logs
 from besmarts.mechanics import molecular_models as mm
 from besmarts.mechanics import objectives
 from besmarts.mechanics import fits
@@ -31,7 +32,7 @@ def optimize_positions_scipy(csys, psys: mm.physical_system):
         args,
         jac=jac,
         args=(keys, csys, psys),
-        options={'disp': False, 'gtol': 1e-3, 'ftol': 1e-4, 'maxiter': 1000},
+        options={'disp': False, 'gtol': 1e-3, 'ftol': 1e-4, 'maxiter': 2},
         method='L-BFGS-B',
     )
 
@@ -113,94 +114,130 @@ def objective_gdb(csys, gdb, obj, psysref=None, reuse=None, ws=None):
         X += xi*x.scale
     return X
 
-def objective_gradient_gdb(args, keys, csys, gdb, obj, psysref=None, reuse=None, ws=None, verbose=False):
+def objective_gradient_gdb(args, keys, csys, gdb, obj, history=None, psysref=None, reuse=None, ws=None, verbose=False):
 
+    if history is None:
+        history = []
+    n = len(history)
     # csys = copy.deepcopy(csys)
     tasks = {}
     h = []
     # print("REUSE IS", reuse)
+    logs.dprint(f"{logs.timestamp()} Generating {len(obj)} objectives", on=verbose)
     for i, x in obj.items():
         psys = None
         if psysref:
             psys = psysref[x.addr.eid[0]]
-        tasks[(i, 0)] = x.get_task(gdb, csys, keys={}, psys=psys, reuse=reuse)
-        if psys:
+            # logs.dprint(f"{logs.timestamp()} Reparameterizing with reuse={reuse}", on=verbose)
+            
+            # get a copy
+            # psys = mm.chemical_system_to_physical_system(
+            #     csys,
+            #     psys.models[0].positions,
+            #     ref=psys,
+            #     reuse=list(range(len(psys.models)))
+            # )
+            # now change values
+            for k, v in zip(keys, args):
+                mm.physical_system_set_value(psys, k, v)
+        else:
+            print("WARNING: No parameterized system given. This will recharge the molecules and is likely not intended.")
+            csys = copy.deepcopy(csys)
+            for k, v in zip(keys, args):
+                mm.chemical_system_set_value(csys, k, v)
             psys = mm.chemical_system_to_physical_system(
                 csys,
                 psys.models[0].positions,
-                ref=psys,
-                reuse=reuse
+                ref=None,
+                # reuse=[0,2,3,4,5]
             )
-        # print("Pos")
-        # for ic, vals in psys.models[0].positions[0].selections.items():
-        #     print(ic, vals)
+            reuse=list(range(len(psys.models)))
+
+            #     logs.dprint(f"PSYS: setting {k}={v}", on=verbose)
+        # dcsys = copy.deepcopy(csys)
+        dcsys = csys
+        dreuse=list(range(len(psys.models)))
+        # dreuse=[2,3,4,5]
+        tasks[(n, (i, 0))] = x.get_task(gdb, dcsys, keys={}, psys=psys, reuse=dreuse)
         for j, (k, v) in enumerate(zip(keys, args), 1):
 
             # hi = v*1e-4
             hi = 1e-6
             h.append(hi)
[…]
