### 2024-06-02 302cb33
rename test function to work with pytest
