### 2024-06-02 3420481
updates for simple ff fit example
