+++ doc/graph_io.rst
@@ -21,7 +21,7 @@ coordinate:
 >>> gcd = graph_codec_rdkit()
 >>> smarts = '[#6:1]~[*:2]=[#8]'
 >>> g = gcd.smarts_decode(smarts)
->>> print(gcd.smarts_encode(smarts))
+>>> print(gcd.smarts_encode(g))
 [#6:1]~[*:2]=[#8:3]
 
 We have just created a graph, but specifically a subgraph because the SMARTS
@@ -44,8 +44,8 @@ decoding twice:
 .. code-block:: python
 
 >>> from besmarts.codecs import codec_native
->>> codec_native.graph_codec_native_save("g.bes", [g])
->>> g = native.graph_codec_native_load("g.bes")[0]
+>>> success = codec_native.graph_codec_native_save("g.bes", [g])
+>>> g = codec_native.graph_codec_native_load("g.bes")[0]
 >>> print(gcd.smarts_encode(g))
 [#6:1]~[*:2]=[#8:3]
 
@@ -117,6 +117,7 @@ and then only print SMARTS patterns using the element and bond order primitives:
 >>> print(gcd.smarts_encode(g))
 [#6](-[#1])(-[#1])(-[#1])-[#6](-[#1])(-[#1])-[#1]
 >>> graphs.graph_set_primitives_atom(g, atom_primitives[::-1])
+>>> print(gcd.smarts_encode(g))
 [+0A!rx0X4H3#6](-[+0A!rx0X1H0#1])(-[+0A!rx0X1H0#1])(-[+0A!rx0X1H0#1])-[+0A!rx0X4H3#6](-[+0A!rx0X1H0#1])(-[+0A!rx0X1H0#1])-[+0A!rx0X1H0#1]
 
 This will not delete the other primitives as show above. Subsequent operations,
@@ -134,16 +135,16 @@ atoms define the topology. For example, if have the SMARTS
 
 .. code-block:: python
 
-    from besmarts.codecs.codec_rdkit import graph_codec_rdkit
-    from besmarts.core import graphs
-    from besmarts.core import topology
-
-    gcd = graph_codec_rdkit()
-
-    g = "[#6:1]([#6])[#6:2]"
-    g = gcd.smarts_decode(g)
-    topo = topology.bond_topology()
-    bond = graphs.subgraph_to_structure(g, topo)
+>>> from besmarts.codecs.codec_rdkit import graph_codec_rdkit
+>>> from besmarts.core import graphs
+>>> from besmarts.core import topology
+>>> gcd = graph_codec_rdkit()
+>>> g = "[#6:1]([#6])[#6:2]"
+>>> g = gcd.smarts_decode(g)
+>>> topo = topology.bond_topology()
+>>> bond = graphs.subgraph_to_structure(g, topo)
+>>> print(gcd.smarts_encode(bond))
+[#6:1]([#6:2])[#6]
 
 Note that loading a SMARTS always produces a subgraph data type, and so the
 untagged atom in the pattern will also be part of the subgraph. This would
@@ -233,14 +234,57 @@ used to get the canonical ordering:
 
 .. code-block:: python
 
+>>> from besmarts.codecs.codec_rdkit import graph_codec_rdkit
+>>> from besmarts.core import graphs 
+>>> from besmarts.core import geometry
+>>> gcd = graph_codec_rdkit()
+>>> g = 'CC'
+>>> g = gcd.smiles_decode(g)
 >>> for indices in graphs.graph_bonds(g):
->>>     geometry.bond(indices[::-1])
+>>>     print(indices, indices == geometry.bond(indices[::-1]))
 >>> for indices in graphs.graph_angles(g):
->>>     geometry.angle(indices[::-1])
+>>>     print(indices, indices == geometry.angle(indices[::-1]))
[…]
