+++ doc/clustering.rst
@@ -24,6 +24,10 @@ same labels. Here is how to do it:
 >>> from besmarts.codecs.codec_rdkit import graph_codec_rdkit
 >>> from besmarts.assign.hierarchy_assign_rdkit import smarts_hierarchy_assignment_rdkit
 >>> from besmarts.core import hierarchies
+>>> from besmarts.core import configs
+>>> 
+>>> configs.workqueue_port = 54321 # make sure this port is open/unused
+>>> configs.remote_compute_enable = False # port is only open to localhost
 >>> 
 >>> gcd = graph_codec_rdkit()
 >>> labeler = smarts_hierarchy_assignment_rdkit()
@@ -37,17 +41,19 @@ same labels. Here is how to do it:
 >>> 
 >>> hierarchies.smarts_hierarchy_print(cst.hierarchy)
 
+
 .. code-block::
     
     Assigning molecule     1/1 at depth 0
+    
     Labels per unique structure that need more depth
     There are 2/3 unique structures at depth 0
     There are 0 problems:
     Max depth is set to 0
-    2024-05-23 17:26:26.587125 Labeling subgraphs
-    2024-05-23 17:26:26.587688 Checking consistency...
+    2024-06-02 10:59:16.684934 Labeling subgraphs
+    2024-06-02 10:59:16.685488 Checking consistency...
     Optimization strategy is building steps...
-    2024-05-23 17:26:26.587762 The optimization strategy has the following iterations:
+    2024-06-02 10:59:16.685556 The optimization strategy has the following iterations:
     ->   1. op= 1 a=[0] b=1->1 d=0->0 n=0->3
          2. op=-1 a=[0] b=0->0 d=0->0 n=0->0
          3. op= 1 a=[0] b=2->2 d=0->0 n=0->3
@@ -61,15 +67,15 @@ same labels. Here is how to do it:
     
     
     *******************
-     2024-05-23 17:26:26.587850 iteration=   1 macro=  1/6 X=        0 params=(1|2) G=Y S=Y bits=1->3 depth=0->0 branch=0->3
+     2024-06-02 10:59:16.685609 iteration=   1 macro=  1/6 X=        0 params=(1|2) G=Y S=Y bits=1->3 depth=0->0 branch=0->3
     *******************
     
     Tree:
     **  0   0 p0   {'a', 'b'} [*:1]~[*:2]
     =====
     
-    2024-05-23 17:26:26.587901 Saving checkpoint to chk.cst.p
-    2024-05-23 17:26:26.588329 Collecting SMARTS for p0 N=3/3 and setting to depth=0
+    2024-06-02 10:59:16.685637 Saving checkpoint to chk.cst.p
+    2024-06-02 10:59:16.686071 Collecting SMARTS for p0 N=3/3 and setting to depth=0
      == iteration=   2 macro=  1/6 micro=  1/1 operation=1 params=(1|2) cluster=p0   N= 3 overlap=[0] bits=1->1 depth=0->0 branch=0->3
     
     Attempting to split p0:
@@ -79,13 +85,13 @@ same labels. Here is how to do it:
     000002 (0, (1, 2))              {'a'} [#6H1X2x0!rA+0:1]!@;#[#6H1X2x0!rA+0:2]
     000003 (0, (2, 4))              {'b'} [#6H1X2x0!rA+0:2]!@;-[#1H0X1x0!rA+0:4]
     
-    2024-05-23 17:26:26.591432 Union merging=3
-    2024-05-23 17:26:26.604654 Union merging=2
-    2024-05-23 17:26:26.613429 Union merging=1
[…]
