### 2024-06-03 552e08d
force local workqueue if remote compute is disabled
