+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -176,6 +176,9 @@ class compute_config_gradient(compute_config):
     """
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_entry]]:
+        """
+        compute_config_gradient::run
+        """
         # print("Starting calculation", self)
         csys = self.csys
         gdb = self.GDB
@@ -198,10 +201,7 @@ class compute_config_gradient(compute_config):
                 # pos0 = assignments.graph_db_entry_to_graph_assignment(gdb, assignments.POSITIONS, gid)
 
                     pos0 = assignments.graph_db_graph_to_graph_assignment(gdb, eid, tid, gid, rid)
-                    for k, v in self.keys.items():
-                        # dprint(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.10g} to {v:.10g} d={v-mm.chemical_system_get_value(csys, k):.10g}", on=verbose)
-                        # mm.chemical_system_set_value(csys, k, v)
-                        mm.physical_system_set_value(psys, k, v)
+                    # psys = mm.chemical_system_to_physical_system(csys, [pos0], ref=self.psys, reuse=self.reuse)
                     system.append(pos0)
 
 
@@ -214,6 +214,10 @@ class compute_config_gradient(compute_config):
                 # such as (gid, sid, rid)
 
                 psys = mm.chemical_system_to_physical_system(csys, system, ref=self.psys, reuse=self.reuse)
+                for k, v in self.keys.items():
+                    # dprint(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.10g} to {v:.10g} d={v-mm.chemical_system_get_value(csys, k):.10g}", on=verbose)
+                    # mm.chemical_system_set_value(csys, k, v)
+                    mm.physical_system_set_value(psys, k, v)
 
                 # this must put everything in the psys into a single system
                 # this must mean that we flatten
@@ -411,7 +415,7 @@ class objective_config_gradient(objective_config):
         dx = arrays.array_scale(dx, 1.0/(2*h))
         DX = 2 * arrays.array_inner_product(X0, dx)
 
-        return dx
+        return DX
 
     def compute_diff(self, GDB: assignments.graph_db, D: List[Dict[int, assignments.graph_db_table]]):
         """
@@ -575,6 +579,7 @@ def gdb_to_physical_systems(gdb, csys, ref=None, reuse=None):
             print("Exception:")
             print(e)
             failures += 1
+            raise e
     if len(gdb.entries) > 100:
         print()
     if failures:
@@ -716,12 +721,27 @@ def forcefield_optimization_strategy_default(csys, models=None):
     determines how to step the optimization forward, choosing which hyperparameter
     to try next
     """
+    if models is None:
+        models = {}
+    elif type(models) is not dict:
+        models = {i: None for i in models}
 
     bounds = {}
+    strat = forcefield_optimization_strategy(bounds)
     for m, cm in enumerate(csys.models):
-        if models is not None and m not in models:
+        if m not in models:
             continue
+
+
         for h, hidx in enumerate(mm.chemical_model_iter_smarts_hierarchies(cm)):
+
+            nodes = [n.name for n in hidx.index.nodes.values() if n.type == "parameter"]
+            strat.reference_list.extend(nodes)
+
+            assert type(models[m]) is not str
+            if models[m] is not None:
+                strat.target_list.extend(models[m])
+
             splitter = configs.smarts_splitter_config(
                 1, 1, 0, 0, 0, 0, True, True, 0, True, True, True, True
[…]
