+++ doc/fffit.rst
@@ -6,7 +6,17 @@ this package, was force field fitting with automatic chemical perception. What
 this means is that we want the code to fit the force field parameters, but also
 find the best parameters that give the best fit as well.
 
+This example tries to split the parameter b4 from OpenFF 2.1.0 using a single
+molecule. The physical objectives to fit to are the geometry and forces from
+a DFT calculation using B3LYP-D3BJ/DZVP in Psi4. This example will first fit
+the physical parameters before splitting, and then try to find new bond parameters
+that are able to further improve the fit (only targeting b4). The physical
+parameter fits will only adjust the bonds, and so angles, torsions, non-bonded,
+etc. are all unchanged throughout the run. Furthermore, only the equilibrium
+lengths of the bonds will be modified.
+
 .. code-block:: python
+
 >>> from typing import Dict, List, Tuple
 >>> from besmarts.mechanics import fits
 >>> from besmarts.mechanics import smirnoff_xml
@@ -24,9 +34,13 @@ find the best parameters that give the best fit as well.
 >>> from besmarts.codecs import codec_rdkit
 >>> from besmarts.core import compute
 >>> from besmarts.core import configs
->>> # configs.processors = 1
+>>> 
+>>> configs.processors = 4
+>>> configs.remote_compute_enable = False
+>>> configs.workqueue_port = 54321
+>>> 
 >>> xyz_positions = """11
->>> Energy =  -802.9904298568 (+0.195 kcal/mol)
+>>> 
 >>>   C -1.44819400 -0.84940800  0.16848900
 >>>   C -1.59401300  0.50318700 -0.01678100
 >>>   C -0.27397600  1.02622600 -0.13503500
@@ -62,7 +76,7 @@ find the best parameters that give the best fit as well.
 >>>         if N is None:
 >>>             N = int(lines[0].split()[0])
 >>>         if indices is None:
->>>             indices = list(range(N))
+>>>             indices = [*range(1,N+1)]
 >>>         assert N == int(lines[0].split()[0])
 >>>         for chunk in arrays.batched(lines, N+2):
 >>>             if chunk and chunk[0]:
@@ -70,16 +84,17 @@ find the best parameters that give the best fit as well.
 >>>                 for i, line in enumerate(chunk, -2):
 >>>                     if i >= 0:
 >>>                         sym, x, y, z = line.split()[:4]
->>>                         sel[indices[i]] = list(map(float, (x, y, z)))
+>>>                         sel[indices[i],] = list(map(float, (x, y, z)))
 >>> 
 >>>                 gdc = assignments.graph_db_column()
 >>>                 gdc.selections.update(sel)
 >>>                 gdr.columns[s] = gdc
 >>>                 s += 1
 >>>     return gdr
+>>> 
 >>> def make():
 >>>     smi = "[C:1]1([H:9])=[C:2]([H:10])[C:3]([H:11])=[C:4]([C:5](=[O:6])[Cl:7])[O:8]1"
->>>     s = xyz_positions 
+>>>     s = xyz_positions
 >>>     g = xyz_grad
 >>>     d  = {
 >>>         smi: [
@@ -90,6 +105,7 @@ find the best parameters that give the best fit as well.
 >>>         ],
 >>>     }
 >>>     return d
+>>> 
[…]
