+++ besmarts-core/python/examples/ff_fit.py
@@ -1,7 +1,3 @@
-
-"""
-"""
-
 from typing import Dict, List, Tuple
 from besmarts.mechanics import fits
 from besmarts.mechanics import smirnoff_xml
@@ -19,9 +15,13 @@ from besmarts.assign import hierarchy_assign_rdkit
 from besmarts.codecs import codec_rdkit
 from besmarts.core import compute
 from besmarts.core import configs
-# configs.processors = 1
+
+configs.processors = 4
+configs.remote_compute_enable = False
+configs.workqueue_port = 54321
+
 xyz_positions = """11
-Energy =  -802.9904298568 (+0.195 kcal/mol)
+
   C -1.44819400 -0.84940800  0.16848900
   C -1.59401300  0.50318700 -0.01678100
   C -0.27397600  1.02622600 -0.13503500
@@ -57,7 +57,7 @@ def load_xyz(flist, indices=None) -> assignments.graph_db_row:
         if N is None:
             N = int(lines[0].split()[0])
         if indices is None:
-            indices = list(range(N))
+            indices = [*range(1,N+1)]
         assert N == int(lines[0].split()[0])
         for chunk in arrays.batched(lines, N+2):
             if chunk and chunk[0]:
@@ -65,16 +65,17 @@ def load_xyz(flist, indices=None) -> assignments.graph_db_row:
                 for i, line in enumerate(chunk, -2):
                     if i >= 0:
                         sym, x, y, z = line.split()[:4]
-                        sel[indices[i]] = list(map(float, (x, y, z)))
+                        sel[indices[i],] = list(map(float, (x, y, z)))
 
                 gdc = assignments.graph_db_column()
                 gdc.selections.update(sel)
                 gdr.columns[s] = gdc
                 s += 1
     return gdr
+
 def make():
     smi = "[C:1]1([H:9])=[C:2]([H:10])[C:3]([H:11])=[C:4]([C:5](=[O:6])[Cl:7])[O:8]1"
-    s = xyz_positions 
+    s = xyz_positions
     g = xyz_grad
     d  = {
         smi: [
@@ -85,6 +86,7 @@ def make():
         ],
     }
     return d
+
 def new_gdb(f: Dict[str, List[str]]) -> assignments.graph_db:
     gcd = codec_rdkit.graph_codec_rdkit()
     gdb = assignments.graph_db()
@@ -93,7 +95,7 @@ def new_gdb(f: Dict[str, List[str]]) -> assignments.graph_db:
     for smi, fn_dict in f.items():
 
         g = gcd.smiles_decode(smi)
-        gid = assignments.graph_db_add_graph(gdb, g)
+        gid = assignments.graph_db_add_graph(gdb, smi, g)
 
         gdb.graphs[gid] = g
         gdb.smiles[gid] = smi
@@ -108,8 +110,9 @@ def new_gdb(f: Dict[str, List[str]]) -> assignments.graph_db:
             gdg = assignments.graph_db_graph()
             gdt.graphs[gid] = gdg
             fn = rdata[tid]
-            indices=dict(sorted([(j, x) for j, x in enumerate(g.nodes)], key=lambda x: x[1]))
-            gdr = load_xyz([fn], indices=list(indices))
+            # indices=dict(sorted([(j, x) for j, x in enumerate(g.nodes, 1)], key=lambda x: x[1]))
+            indices = None
+            gdr = load_xyz([fn], indices=indices)
             gdg.rows[0] = gdr
             gde.tables[tid] = gdt
             tid = assignments.GRADIENTS
@@ -118,8 +121,8 @@ def new_gdb(f: Dict[str, List[str]]) -> assignments.graph_db:
                 gdg = assignments.graph_db_graph()
                 gdt.graphs[gid] = gdg
[…]
