### 2024-06-03 c071ea7
remove smiles enumeration doc from main branch
