### 2024-06-03 trevorgokey/besmarts#1 @trevorgokey
Hi @AntObi, thanks for taking such a detailed look. I have since given the documentation a nearly complete overhaul, and most what didn't work or was out of date was removed. The current documentation [https://besmarts.readthedocs.io/en/latest/index.html]() should be up to date and all examples and code snippets should be working.

I'll give a broken down response below:

### Installation

The PyPI instructions were outdated and based on old software that this package has superseded. I have removed this reference for now to reflect the current state of the package, and will add it again once the package is up on PyPI.

### Testing

Thanks for catching this! I built the tests using the `unittest` module in the standard python library, which didn't find an issue with this. I made a name change though, so I hope this will make the tests pytest friendly!

### Examples

Thanks for giving these a spin. I've fixed the `bond_union.py` example to work properly. As for the installation issue, there was a typo in the `setup.py` and was pointing at the wrong folder. This means installing was a success but the wrong files were installed. I've fixed this and the examples should be working.

### Usage

> [Graphs](https://besmarts.readthedocs.io/en/latest/graph_io.html#graphs)

Import errors and other bugs have been fixed. Some code which has been indicated here (in particular `mapper.structure_mapper`) has not been updated/tested thoroughly and needs to be fixed. It is not essential to any critical code in the examples and therefore I have removed it from the documentation until it is better developed and tested.

> [Labeling](https://besmarts.readthedocs.io/en/latest/labeling.html)

There is now an example showing how to label a molecule.

> [Splitting](https://besmarts.readthedocs.io/en/latest/splitting.html)

Should be fixed and working now.

> [Clustering](https://besmarts.readthedocs.io/en/latest/clustering.html)

Should be fixed and working now.

> [Dataset](https://besmarts.readthedocs.io/en/latest/dataset.html)

Ah, good catch. I've since moved this functionally to a difference branch and forgot to remove this page. I plan to further work on this later since it is not essential for the main purpose of this package
