### 2024-06-03 #6653 @trevorgokey
Thanks for going through the documentation @Ruibin-Liu and @AntObi. I've since given the RTD documentation an overhaul and should be all working as far as I can tell, including installation, tests, and examples. Also, I've made the `main` and `joss-manuscript` branches the same just to reduce any confusion.

I'll continue to work on code quality such as docstrings; lots of the newer stuff is in flux and so I haven't quite fixed their behavior. Unfortunately, a lof of the newer stuff is also the most complicated so I apologize if the interesting bits of the code have been opaque. I'll continue working on the second checklist provided by @AntObi.
