### 2024-06-24 #6653 @lucydot
@trevorgokey; an update on how the amendments are going and timeframe would be great.
