+++ besmarts-openmm/python/besmarts/mechanics/optimizers_openmm.py
@@ -1,0 +1,486 @@
+#!/usr/bin/env python3 
+
+from typing import List, Dict, Tuple
+import math
+import copy
+import sys
+
+from besmarts.core import graphs
+from besmarts.core import perception
+from besmarts.core import assignments
+from besmarts.core import arrays
+from besmarts.codecs import codec_rdkit
+from besmarts.mechanics import smirnoff_models
+from besmarts.mechanics import molecular_models as mm
+from besmarts.assign import hierarchy_assign_rdkit
+from besmarts.mechanics import optimizers_scipy
+from besmarts.mechanics import objectives
+
+import openmm.openmm
+import openmm.app.element
+import openmm.app.topology
+import openmm.app.simulation
+import openmm.app
+
+def physical_system_to_openmm_system(psys):
+    pos = psys.models[0].positions[0]
+    system = openmm.openmm.System()
+    atom_map = {}
+
+    for i, node in sorted(pos.graph.nodes.items(), key=lambda x: x[0]):
+        n = node.primitives['element'].on()[0]
+        m = openmm.app.element.Element.getByAtomicNumber(n)
+        atom_map[i] = system.addParticle(m.mass)
+    
+
+    frc, b_map = assign_bonds(psys, atom_map)
+    system.addForce(frc)
+
+    frc, a_map = assign_angles(psys, atom_map)
+    system.addForce(frc)
+
+    frc, t_map = assign_torsions(psys, atom_map)
+    system.addForce(frc)
+
+    frc, i_map = assign_outofplanes(psys, atom_map)
+    system.addForce(frc)
+
+    # frc, nb_map = assign_nonbonded(psys, atom_map)
+    # e_map = assign_scales_nonbonded(psys, atom_map, frc)
+    # frc.setNonbondedMethod(frc.NoCutoff)
+
+    frc, q_map = assign_elec(psys, atom_map)
+    frc.setNonbondedMethod(frc.NoCutoff)
+    eq_map = assign_scales_elec(psys, atom_map, frc)
+    system.addForce(frc)
+
+    frc, lj_map = assign_lj(psys, atom_map)
+    frc.setNonbondedMethod(frc.NoCutoff)
+    elj_map = assign_scales_lj(psys, atom_map, frc)
+    system.addForce(frc)
+
+    for i, f in enumerate(system.getForces()):
+        f.setForceGroup(i)
+
+    topo = openmm.app.Topology()
+    chain = topo.addChain()
+    res = topo.addResidue("MOL", chain)
+
+    for i, j in sorted(atom_map.items(), key=lambda x: x[1]):
+        n = pos.graph.nodes[i].primitives['element'].on()[0]
+        elem = openmm.app.element.Element.getByAtomicNumber(n)
+        topo.addAtom(str(n), elem, res, id=j)
+
+    for i, j in sorted(b_map.items(), key=lambda x: x[1]):
+        ic = atom_map[i[0]], atom_map[i[1]]
+        topo.addBond(*ic)
+
+    integ = openmm.openmm.VerletIntegrator(1.0)
+    # for p in range(openmm.openmm.Platform.getNumPlatforms()):
+    #     plat = openmm.openmm.Platform.getPlatform(p)
+    #     print(plat.getName())
+    platform = openmm.openmm.Platform.getPlatform(0)
+    sim = openmm.app.simulation.Simulation(topo, system, integ, platform)
+    ctx = sim.context
+
+    xyz = []
+    for i, j in sorted(atom_map.items(), key=lambda x: x[1]):
+        xyz.append(arrays.array_scale(pos.selections[i,][0], 0.1))
+    ctx.setPositions(xyz)
+
[…]
