### 2024-07-20 c54096d
initial impl of b-matrix second derivs
