+++ besmarts-core/python/besmarts/core/geometry.py
@@ -39,6 +39,9 @@ def array_multiply(a, b):
 def array_inner_product(a, b):
     return sum((i*j for i,j in zip(a,b)))
 
+def array_outer_product(a, b):
+    return [x for ai in a for x in array_scale(b,  ai)]
+
 def array_cross(a, b):
     return (
         (a[1]*b[2] - a[2]*b[1]),
@@ -64,7 +67,10 @@ def array_magnitude(a) -> float:
 
 
 def array_distance(a,b) -> float:
-    return sum([x*x for x in array_difference(b,a)])**.5
+    return math.sqrt(sum([x*x for x in array_difference(b,a)]))
+
+def array_round(a, b):
+    return [round(x, b) for x in a]
 
 def measure_distance(xyz1, xyz2):
 
@@ -86,7 +92,43 @@ def jacobian_distance(xyz1, xyz2):
 
     return result
 
-def jacobian_angle_geometric(xyz1, xyz2, xyz3):
+def jacobian2_distance_terms(u, lu):
+
+    x = [(u[i]*u[j] - int(i == j))/lu for i in range(3) for j in range(3)]
+
+    return x
+
+def jacobian2_distance(xyz1, xyz2):
+
+    """
+    from Bakken & Helgaker 2012 10.1063/1.1515483
+    gives 2x2 1x9 arrays per position, (4 3x3)
+    """
+
+    result = []
+
+    for a, b in zip(xyz1, xyz2):
+
+        mat = []
+
+        u, lu = array_basis(a, b)
+        t = jacobian2_distance_terms(u, lu)
+
+        row = []
+        row.append(array_scale(t, -1))
+        row.append(array_scale(t,  1))
+        mat.append(row)
+        
+        row = []
+        row.append(array_scale(t,  1))
+        row.append(array_scale(t, -1))
+        mat.append(row)
+
+        result.append(mat)
+
+    return result
+
+def jacobian_angle(xyz1, xyz2, xyz3):
     result = []
     basis = array_basis
     for a, b, c in zip(xyz1, xyz2, xyz3):
@@ -102,36 +144,122 @@ def jacobian_angle_geometric(xyz1, xyz2, xyz3):
         result.append([term1, array_scale(array_add(term1, term2), -1), term2])
     return result
 
-def jacobian_angle(xyz1, xyz2, xyz3):
+def jacobian2_angle_terms(uu, uv, vv, lu, lv, cosq, sinq):
 
-    return jacobian_angle_geometric(xyz1, xyz2, xyz3)
+    scale = array_scale
+    lus = 1/(lu*lu*sinq)
+    lvs = 1/(lv*lv*sinq)
+
+    luvs = 1/(lu*lv*sinq)
+
+
+    t1 = [
+        (uv[3*i+j] + uv[3*j+i] - 3*uu[3*i+j]*cosq + int(i==j)*cosq) * lus
+            for i in range(3) for j in range(3)
+    ]
+
+    t2 = [
+        (uv[3*j+i] + uv[3*i+j] - 3*vv[3*i+j]*cosq + int(i==j)*cosq) * lvs
+            for i in range(3) for j in range(3)
+    ]
+
+    t3 = [
+        (uu[3*i+j] + vv[3*j+i] - uv[3*i+j]*cosq - int(i==j)) * luvs
+            for i in range(3) for j in range(3)
+    ]
+
+    t4 = [
+        (uu[3*j+i] + vv[3*i+j] - uv[3*j+i]*cosq - int(i==j)) * luvs
[…]
