+++ besmarts-core/python/besmarts/core/assignments.py
@@ -51,7 +51,20 @@ sid_t = int # selection id
 pid_t = int # topology id
 xid_t = int
 
+class graph_db_address:
+    def __init__(self, *, eid=None, tid=None, gid=None, rid=None, cid=None, sid=None, xid=None):
+        self.eid: list = [] if not eid else eid
+        self.tid: list = [] if not tid else tid
+        self.gid: list = [] if not gid else gid
+        self.rid: list = [] if not rid else rid
+        self.cid: list = [] if not cid else cid
+        self.sid: list = [] if not sid else sid
+        self.xid: list = [] if not xid else xid
+
 class graph_db_column:
+
+    __slots__ = "selections",
+
     def __init__(self):
         self.selections: Dict[sid_t, List] = {}
 
@@ -65,6 +78,8 @@ class graph_db_column:
 
 class graph_db_row:
 
+    __slots__ = "columns",
+
     def __init__(self):
         self.columns: Dict[cid_t, graph_db_column] = {}
 
@@ -75,6 +90,9 @@ class graph_db_row:
         self.columns[cid] = columns
 
 class graph_db_graph:
+
+    __slots__ = "rows",
+
     def __init__(self):
         self.rows: Dict[rid_t, graph_db_row] = {}
 
@@ -88,6 +106,8 @@ class graph_db_table:
     """
     Holds the data for a db of graphs that all share the same topology.
     """
+    __slots__ = "topology", "graphs", "values"
+
     def __init__(self, top):
         self.topology: topology.structure_topology = top
         self.graphs: Dict[gid_t, graph_db_graph] = {}
@@ -103,6 +123,8 @@ class graph_db_entry:
     """
     Data associated with a set of graphs
     """
+    __slots__ = "tables",
+
     def __init__(self):
         self.tables: Dict[tid_t, graph_db_table] = {}
 
@@ -117,6 +139,7 @@ class graph_db:
     """
     The global data definitions
     """
+    __slots__ = "smiles", "graphs", "selections", "entries"
 
     def __init__(self):
         self.smiles: Dict[gid_t, str] = {}
@@ -136,6 +159,47 @@ def graph_db_get_entries(GDB: graph_db, eids: List[eid_t]) -> graph_db:
     gdb.entries = {i: GDB.entries[i] for i in eids}
     gids = set([gid for t in gdb.entries.values() for r in t.tables.values() for gid in r.graphs])
     gdb.graphs.update({i: GDB.graphs[i] for i in gids})
+
+    gdb.smiles.update({i: GDB.smiles[i] for i in gids})
+
+    for topo, sel in GDB.selections.items():
+        gdb.selections[topo] = {i: sel[i] for i in gids if i in sel}
+
+    return gdb
+
+
+def graph_db_get(GDB: graph_db, addr: graph_db_address) -> graph_db:
+
+    gdb = graph_db()
+
+    assert addr.eid
+
+    gdb.entries = {i: graph_db_entry() for i in addr.eid}
+
+    tids = None
+    if addr.tid:
+        tids = addr.tid
+
+    if addr.gid:
+        gids = addr.gid
+    else:
+        gids = set([gid for gde, e in GDB.entries.items() 
+            for tid, t in e.tables.items() 
+                for gid in t.graphs 
+                    if gde in addr.eid and (tids is None or tid in tids)
+        ])
+
+    for eid, e in gdb.entries.items():
+        GDE = GDB.entries[eid]
+        if tids is None:
+            e.tables.update(GDE.tables)
+        else:
+            for tid in tids:
+                if tid in GDE.tables:
+                    e.tables[tid] = GDE.tables[tid]
+
[…]
