### 2024-07-20 347e338
remove excess sleepiness and try to get around hangs
