+++ besmarts-core/python/besmarts/core/compute.py
@@ -9,6 +9,7 @@ Architecture-wise, it interfaces the multiprocessing module
 
 
 import traceback
+import signal
 from typing import Tuple, Callable, Sequence, Mapping, Dict
 import itertools
 import io
@@ -53,7 +54,8 @@ distributed_function = Tuple[Callable, Sequence, Mapping]
 # search/replace this if using a single file
 # remote_compute_enable = configs.remote_compute_enable
 
-TIMEOUT = 20
+TIMEOUT = 3
+
 
 SHM_GLOBAL = {}
 
@@ -83,7 +85,9 @@ class workspace_status:
 
 
 def dprint(*args, **kwds):
-    if False:
+    
+    if kwds.get("on", False):
+        kwds.pop("on")
         print(*args, **kwds)
 
 
@@ -152,20 +156,17 @@ class Connection(connection.Connection):
         # int which is about 2GB. This should remove the problem
         header = n.to_bytes(8, byteorder="big", signed=False)
         # print("SHIP IT:", n,  header)
-        try:
-            if n > 16384:
-                # The payload is large so Nagle's algorithm won't be triggered
-                # and we'd better avoid the cost of concatenation.
-                self._send(header)
-                self._send(buf)
-            else:
-                # Issue #20540: concatenate before sending, to avoid delays due
-                # to Nagle's algorithm on a TCP socket.
-                # Also note we want to avoid sending a 0-length buffer separately,
-                # to avoid "broken pipe" errors if the other end closed the pipe.
-                self._send(header + buf)
-        except BrokenPipeError:
-            pass
+        if n > 16384:
+            # The payload is large so Nagle's algorithm won't be triggered
+            # and we'd better avoid the cost of concatenation.
+            self._send(header)
+            self._send(buf)
+        else:
+            # Issue #20540: concatenate before sending, to avoid delays due
+            # to Nagle's algorithm on a TCP socket.
+            # Also note we want to avoid sending a 0-length buffer separately,
+            # to avoid "broken pipe" errors if the other end closed the pipe.
+            self._send(header + buf)
 
     def _recv_bytes(self, maxsize=None):
         # buf = bytes(self._recv(8).getbuffer())
@@ -238,7 +239,7 @@ def Client(address, family=None, authkey=None, timeout=TIMEOUT):
 
 class Listener(connection.Listener):
     def __init__(self, *args, **kwargs):
-        kwargs["backlog"] = 64000
+        kwargs["backlog"] = 8000
         super().__init__(*args, **kwargs)
         dprint("INIT LISTENER WITH LARGE BACKLOG")
 
@@ -288,7 +289,13 @@ class Server(managers.Server):
             try:
                 methodname = obj = None
                 thread_name_set("recv")
+                t0 = time.perf_counter()
                 request = recv()
+                t = time.perf_counter()
+                if t-t0 > 10:
+                    dprint(
+                        f"T{name} RECV DONE t={t-t0:.6f} {request[:2]}", on=True
+                    )
                 ident, methodname, args, kwds = request
 
                 try:
@@ -315,14 +322,15 @@ class Server(managers.Server):
                     thread_name_set(name + methodname)
                     t0 = time.perf_counter()
                     # these are always readonly so we can fork and skip GIL
-                    if False and name.startswith("shm"):
-                        res = self.shm_pool.apply(function, args, kwds)
-                    else:
-                        res = function(*args, **kwds)
+                    res = function(*args, **kwds)
                     t = time.perf_counter()
-                    dprint(
[…]
