### 2024-07-20 1113a99
quick util to get mapped SMILES
