### 2024-07-20 fba85df
initial frequency calc impl
