+++ besmarts-core/python/besmarts/mechanics/vibration.py
@@ -1,0 +1,692 @@
+#!/usr/bin/env python3
+import pickle
+import numpy as np
+# import matplotlib.pyplot as plt
+# from matplotlib.ticker import MultipleLocator, AutoMinorLocator
+# from geometric.internal import *
+# from geometric.molecule import Molecule as geometric_mol
+from scipy.linalg import svd
+import sys
+
+gpermol2au = 1836
+au2amu = 1/1836.
+amu2au = 1822.888184038791
+kcal2hartree = 1/627.5096080306
+ang2bohr = 1/0.529177249
+bohr2angstrom = 0.529177249
+hartree2kcalmol = 627.5096080306
+# eig2wavenumber = 1.0 /(2*np.pi * 137 * 5.29e-9) * np.sqrt(au2amu)
+eig2wavenumber = 33.3564095 / (2*np.pi) 
+eig2thz = eig2wavenumber * 2.99792458e10 / 1e12 
+# conversion = eig2wavenumber
+au2cm = 5.29177249e-9
+c = 137
+conversion = 1.0 /(2*np.pi * c * au2cm)
+
+def load_xyz(fname):
+
+    xyzdat = np.genfromtxt(fname, skip_header=2, usecols=range(4), dtype=('U2', 'f8', 'f8', 'f8' ))
+    syms = [x[0] for x in xyzdat]
+    xyz  = np.array([list(x)[1:] for x in xyzdat])
+    return syms, xyz
+
+def converteig( eigs):
+    return np.sign( eigs) * np.sqrt( np.abs( eigs))*conversion
+
+# connect means whether to remove TR
+#    True -> remove TR
+#    False -> keep TR
+
+# CoordSysDict = {'cart':(CartesianCoordinates, False, False),
+#             'prim':(PrimitiveInternalCoordinates, False, False),
+#             'dlc':(DelocalizedInternalCoordinates, True, False),
+#             'hdlc':(DelocalizedInternalCoordinates, False, True),
+#             'tric':(DelocalizedInternalCoordinates, True, False)}
+# coordsys = 'prim' # kwargs.get('coordsys', 'tric')
+# CVals = None
+# CoordClass, connect, addcart = CoordSysDict[coordsys.lower()]
+# Cons = None
+
+# swap_internals = not connect
+debug = False
+
+def printortho( mat):
+    if not debug:
+       return 
+    for i in mat:
+        for x in i:
+            x = abs(x)
+            if x > 1e-8:
+                print("{:3.1f}".format(x), end="")
+            else:
+                print("Y", end="")
+        print()
+
+def printmat( mat, dec=4):
+    if not debug:
+       return 
+    for row in mat:
+        for val in row:
+            print(("{:"+str(dec+3)+"."+str(dec)+"f} ").format( val), end="")
+        print()
+
+
+mass_table = {'C': 12.011, 'H': 1.008, 'N': 14.007, 'O': 15.999, 'P': 30.973, \
+        'S': 32.06, 'F': 18.998, "Cl": 35.45, "CL": 35.45, "B": 10.81, "I": 126.9, "Br": 79.04, "Na": 22.90 }
+# mass_table = {'C': 12.000, 'H': 1.000, 'N': 14.000, 'O': 16.0, 'P': 31.00, \
+#         'S': 32.00, 'F': 19.0, "Cl": 35, "B": 11.00, "Br": 70.0, "I": 106., "Se": 68.0  }
+# mass_table = {'C': 12.000, 'H': 1.007825032230, 'N': 14.003074004430, 'O': 15.994914619570, 'P': 31.00, \
+#     'S': 32.00, 'F': 19.0, "CL": 35, "B": 11.00 }
[…]
