### 2024-07-20 c073886
change visiting order of primary nodes
