### 2024-07-20 b4df251
major update; initial basin hopping impl
