+++ besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py
@@ -5,6 +5,7 @@ besmarts.mechanics.optimizers_scipy
 
 import copy
 import scipy.optimize
+import numpy as np
 import datetime
 
 from besmarts.core import geometry
@@ -12,12 +13,13 @@ from besmarts.core import assignments
 from besmarts.core import compute
 from besmarts.core import arrays
 from besmarts.core import logs
+from besmarts.core import returns
 from besmarts.mechanics import molecular_models as mm
 from besmarts.mechanics import objectives
 from besmarts.mechanics import fits
 
 
-def optimize_positions_scipy(csys, psys: mm.physical_system):
+def optimize_positions_scipy(csys, psys: mm.physical_system, step_limit=1000, tol=1e-10):
 
     pos = copy.deepcopy(psys.models[0].positions[0])
     
@@ -26,18 +28,36 @@ def optimize_positions_scipy(csys, psys: mm.physical_system):
     # jac = objectives.array_geom_gradient
     # jac = None
     jac = True
+    hess = objectives.array_geom_hessian
 
+    opts = {'disp': False, 'maxls': 10, 'gtol': tol, 'ftol': 1e-9, 'maxiter': step_limit}
+    # opts = {'iprint': 11000}
+    # opts = {'disp': True}
+    method = 'L-BFGS-B'
+    opts = {'disp': False, 'ftol': tol, 'gtol': tol, 'maxiter': step_limit, 'maxls': 1000, 'maxcor': len(args)**2}
+    hess = None
+    # method = 'Newton-CG'
+    # opts = {'disp': False, 'maxiter': step_limit}
+    # method = 'Nelder-Mead'
+    # print("SciPy Minimizer Starting")
+    # method = 'CG'
+    # method = 'BFGS'
+    # opts = {'disp': False, 'gtol': tol, 'maxiter': step_limit}
+    # method = 'trust-exact'
+    # opts = {'disp': False, 'gtol': tol, 'maxiter': step_limit, 'max_trust_radius': 1e-1, 'initial_trust_radius': 1e-3 }
     result = scipy.optimize.minimize(
         objectives.array_geom_energy_gradient,
         args,
         jac=jac,
+        hess=hess,
         args=(keys, csys, psys),
-        options={'disp': False, 'gtol': 1e-3, 'ftol': 1e-4, 'maxiter': 1000},
-        method='L-BFGS-B',
+        options=opts,
+        tol=tol,
+        method=method,
     )
 
     n_confs = len(list(pos.selections.values())[0])
-    for (c, n, i), v in zip(keys, result.x):
+    for (c, n, i), v in zip(keys, arrays.array_round(result.x, 12)):
         pos.selections[n][c][i] = v
 
     return pos
@@ -114,16 +134,46 @@ def objective_gdb(csys, gdb, obj, psysref=None, reuse=None, ws=None):
         X += xi*x.scale
     return X
 
-def objective_gradient_gdb(args, keys, csys, gdb, obj, history=None, psysref=None, reuse=None, ws=None, verbose=False):
+def objective_hessp(x, p, *args):
+    history = args[6]
+    # return arrays.array_inner_product(list(history[-1][3]), list(p))
+    return np.array(arrays.array_multiply(list(history[-1][3]), list(p)))
+
+def objective_hess(x, *args):
+
+    history = args[6]
+
+    diag = history[-1][3]
+
+    hess = []
+    for i in range(len(diag)):
+        hess.append([0 for j in range(len(diag))])
+    for i in range(len(diag)):
+        hess[i][i] = diag[i]
+
+    return hess
+
+def objective_gradient_gdb(args, keys, csys, gdb, obj, priors, penalties=None, history=None, psysref=None, reuse=None, ws=None, verbose=False):
 
     if history is None:
         history = []
+    if penalties is None:
[…]
