### 2024-07-20 7362d96
change default separation to 0.1
