### 2024-07-20 dd631eb
new functionality to get bond/angle geom
