+++ besmarts-core/python/besmarts/mechanics/molecular_models.py
@@ -14,7 +14,6 @@ from besmarts.core import tree_iterators
 from besmarts.core import hierarchies
 from besmarts.core import assignments
 from besmarts.core import perception
-from besmarts.core import db
 
 
 class topology_term:
@@ -146,11 +145,16 @@ class chemical_model_procedure_smarts_assignment(chemical_model_procedure):
         self.topology_terms = topology_terms
         self.smarts_hierarchies: Dict[int, hierarchies.structure_hierarchy] = {}
 
-    def assign(self, cm, pm: physical_model) -> physical_model:
+        # if we don't find a match, use this instead
+        self.default_parameter = None
+
+    def assign(self, cm, pm: physical_model, overrides=None) -> physical_model:
         """
         this will return, for each selection, the reference
         """
 
+        if overrides is None:
+            overrides = {}
         # print(self.topology_terms)
         smiles = [x.smiles for x in pm.positions]
         topo = cm.topology
@@ -162,7 +166,7 @@ class chemical_model_procedure_smarts_assignment(chemical_model_procedure):
             smiles,
             self.smarts_hierarchies[unit_i].topology
         )
-        
+
         assn = []
         vals = []
         for x in lbls.assignments:
@@ -170,9 +174,16 @@ class chemical_model_procedure_smarts_assignment(chemical_model_procedure):
             v = {}
             for ic, lbl in x.selections.items():
                 if lbl is None:
-                    continue
+                    if self.default_parameter:
+                        lbl = self.default_parameter
+                    else:
+                        continue
                 names = self.get_term_labels((unit_i, lbl))
                 values = self.get_term_values((unit_i, lbl))
+                for l, lv in overrides.items():
+                    if l[1] == names.get(l[0]):
+                        # print(f"Override: {l}: {values[l[0]][l[2]]} -> {lv}")
+                        values[l[0]][l[2]] = lv
                 p[ic] = {term: l for (term, l), x  in zip(names.items(), values.values())}
                 v[ic] = {term: x for (term, l), x  in zip(names.items(), values.values())}
             assn.append(p)
@@ -422,10 +433,12 @@ def chemical_system_set_value_list(csys, key, values):
 
     if len(key) == 3:
         m, t, l = key
-        csys.models[m].topology_terms[t].values[l] = values
+        csys.models[m].topology_terms[t].values[l].clear()
+        csys.models[m].topology_terms[t].values[l].extend(values)
     elif len(key) == 2:
         m, t = key
-        csys.models[m].system_terms[t].values = values
+        csys.models[m].system_terms[t].values.clear()
+        csys.models[m].system_terms[t].values.extend(values)
 
 def chemical_system_set_value(csys, key, value):
 
@@ -443,6 +456,8 @@ def physical_system_set_value(psys: physical_system, key, value):
         m, t, l, i = key
         for lbls, values in zip(psys.models[m].labels, psys.models[m].values):
             for ic, ic_lbls in lbls.items():
+                if t not in ic_lbls:
+                    continue
                 term_lbl = ic_lbls[t]
                 if term_lbl == l:
                     # print(f"PSYS SETTING {ic}:{t}:{i} = {value}")
@@ -452,44 +467,6 @@ def physical_system_set_value(psys: physical_system, key, value):
         # TODO lol
         assert False
 
-    # unit_i = 0
-    # lbl = key[2]
-    # lbls = pm.get_term_labels((unit_i, lbl))
-
-    # pm = psys.models[key[0]]
-
[…]
