+++ besmarts-core/python/besmarts/mechanics/objectives.py
@@ -6,377 +6,18 @@ besmarts.mechanics.objectives
 import copy
 import pprint
 from besmarts.core import arrays
+from besmarts.core import primitives
 from besmarts.core import assignments
 from besmarts.mechanics import molecular_models as mm
 
-# def residual_squared_error(x1, x0):
 
-#     result = {}
-#     for k, r1 in x1.selections.items():
-#         r0 = x0.selections[k]
-#         result[k] = (r1 - r0)**2
-#     return result
-
-# def measure_rse(new_sys: system, ref_sys: system, measure):
-#     smi = ref_sys.positions.smiles
-#     graph = ref_sys.positions.graph
-
-#     x0 = measure(smi, graph, ref_sys.positions.selections)
-#     x1 = measure(smi, graph, new_sys.positions.selections)
-
-#     result = residual_squared_error(x1, x0)
-
-#     return assignments.graph_assignment(smi, result, g)
-
-# def bond_rse(new_sys: system, ref_sys: system):
-#     result = measure_rse(new_sys, ref_sys, assignments.graph_assignment_geometry_bonds)
-#     return result
-
-# class chemical_objective:
-#     def __init__(self, cm: chemical_model):
-#         self.total = 0
-#         self.results = Dict[str, Dict[str, float]]
-
-#     def create_compute_tasks(cm):
-#         return default_chem_objective
-
-
-
-# we need an objective from a pm that produces a sag
-# lets say the sag is 3d, then we need:
-# objective(psys, pos) -> sag
-# the sig would then be sag which are pos
-# we then also have an initial csys
-# then we would go csys, sag0 -> psys then to obj(psys) -> value
-"""
-sag0 could be a state with pos and grads and etc
-call it ref0
-then we would need a calculator for each sag, or in other words a sag provider
-csys2pys(ref, csys) -> pm -> obj(pm) -> sag
-
-ref is positions(3d), energy (mol) and grads (3d) for example
-
-csys is the initial ff
-
-so then we would get the psys and then calc pos, grad, and energy
-
-then the objective would take the ref and the psys and return a sag
-
-have a psys, which has a form for the analytic energies and gradients
-
-def modsem(pos, hes) -> sag:
-    pass
-    (then label, then mean)
-
-    # may want some initializers/computers such as bond length and fc
-
-# refs should be a state?
-# smiles_state = Dict[str, topology_assignment]
-
-sst.smiles
-sst.graph
-sst.assignments[POSITIONS].selections[(1,)] = [[0,0,0]]
-sst.assignments[GRADIENTS].selections[(1,)] = [[0,0,0]]
-sst.assignments[HESSIANS].selections[(1,1)] = [[[0,0,0],[0,0,0],[0,0,0]]]
-
-[smiles, assignments, selections]
-(0, POSITIONS, (1,), 0): [x, y, z]
-
-gid = db_graph_assignment_add_graph(db, g)
-aid = db_graph_assignment_add_(g)
-
-gid, aid, sid, vid (eid)
-
-topology_assignment_{str, float, unit, comment}
-
-xyz = smiles_dataset_get_values((0, POSITIONS, (1,), 0))
-
-db_graph_assignment:
-    graphs: list[graph_assignment]
-    assignments = {str: topology_assignment_terms}
-
-db_graph_assignment.graphs[0]
-db_graph_assignment.assignments[POSITIONS][0][(1,)] = [[x,y,z]]
-
-db_graph_topology_assignment
-    topology
-    selections[gid][sid]
-
-CLASSES:
[…]
