### 2024-07-20 0f07803
major update; freq fitting; fixes
