+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -11,13 +11,16 @@ import multiprocessing
 import collections
 import sys
 import math
-from typing import List, Dict, Tuple
+import itertools
+import numpy as np
+from typing import List, Dict, Tuple, Callable
 from besmarts.core import configs
 from besmarts.core import arrays
 from besmarts.core import assignments
 from besmarts.core import topology
 from besmarts.core import graphs
 from besmarts.core import codecs
+from besmarts.core import geometry
 from besmarts.core import compute
 from besmarts.core import clusters
 from besmarts.core import mapper
@@ -27,30 +30,17 @@ from besmarts.core import tree_iterators
 from besmarts.core import optimization
 from besmarts.core import primitives
 from besmarts.core import logs
-from besmarts.mechanics import optimizers_scipy 
+from besmarts.core import returns
+from besmarts.cluster import cluster_objective
+from besmarts.cluster import cluster_optimization
+from besmarts.core import array_numpy
+from besmarts.mechanics import optimizers_scipy
+from besmarts.mechanics import optimizers_openmm
 from besmarts.mechanics import objectives
+from besmarts.mechanics import hessians
+from besmarts.mechanics import vibration
 from besmarts.mechanics import molecular_models as mm
 
-
-class calculator_config_minimization:
-    def __init__(self):
-        self.minimize = True
-        self.minimize_force_convergence = 0.10
-        self.minimize_energy_convergence = 0.10
-        self.minimize_position_convergence = 0.10
-        self.step_limit = 100
-        self.bias_force = {}
-        self.constraints = {}
-
-
-class graph_db_address:
-    def __init__(self, *, eid=None, aid=None, gid=None, rid=None, sid=None, xid=None):
-        self.eid: list = [] if not eid else eid
-        self.aid: list = [] if not aid else aid
-        self.gid: list = [] if not gid else gid
-        self.rid: list = [] if not rid else rid
-        self.sid: list = [] if not sid else sid
-        self.xid: list = [] if not xid else xid
 
 class compute_config:
 
@@ -60,7 +50,7 @@ class compute_config:
 
     def __init__(self, addr):
         self.addr = addr 
-        self.keys = {}
+        self.keys = []
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_table]]:
         # print("Starting calculation", self)
@@ -70,38 +60,52 @@ class compute_config:
         tid = assignments.POSITIONS
         
 
-        results: List[Dict[assignments.tid_t, assignments.graph_db_table]] = []
+        all_results = []
+        for keyset in self.keys:
+            results: List[Dict[assignments.tid_t, assignments.graph_db_table]] = []
 
-        # for k, v in self.keys.items():
-            # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.6g} to {v:.6g} d={v-mm.chemical_system_get_value(csys, k):.6g}")
-            # mm.chemical_system_set_value(csys, k, v)
-            # mm.physical_system_set_value(self.psys, k, v)
+            # for k, v in self.keys.items():
+                # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.6g} to {v:.6g} d={v-mm.chemical_system_get_value(csys, k):.6g}")
+                # mm.chemical_system_set_value(csys, k, v)
+                # mm.physical_system_set_value(self.psys, k, v)
 
-        for eid, gdb in GDB.entries.items():
-            tbl = assignments.graph_db_table(topology.null)
-            for gid in gdb.graphs:
-                # pos0 = assignments.graph_db_entry_to_graph_assignment(gdb, tbl_idx, gid)
-                rid = 0
[…]
