+++ besmarts-core/python/besmarts/mechanics/smirnoff_models.py
@@ -8,6 +8,8 @@ from besmarts.core import topology
 from besmarts.core import assignments
 from besmarts.core import hierarchies
 from besmarts.core import trees
+from besmarts.core import arrays
+from besmarts.core import tree_iterators
 from besmarts.core import perception
 import pprint
 from besmarts.mechanics import molecular_models as mm
@@ -61,6 +63,7 @@ def chemical_model_bond_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
 
     uid = u.index
     for param in d["parameters"]:
+        param = param["Bond"]
         node = proc.smarts_hierarchies[u.index].index.node_add_below(
             root.index
         )
@@ -131,7 +134,7 @@ def chemical_model_angle_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
 
     uid = u.index
     for param in d["parameters"]:
-
+        param = param["Angle"]
         node = proc.smarts_hierarchies[u.index].index.node_add_below(
             root.index
         )
@@ -247,6 +250,10 @@ def chemical_model_torsion_periodic_smirnoff(
     cm.symbol = "T"
     cm.internal_function = assignments.graph_assignment_geometry_torsions
     cm.derivative_function = assignments.graph_assignment_jacobian_torsions
+
+    for i in range(len(d["parameters"])):
+        d["parameters"][i] = d["parameters"][i]["Proper"]
+
     smirnoff_dihedral_load(cm, pcp, d)
 
     return cm
@@ -262,6 +269,10 @@ def chemical_model_outofplane_periodic_smirnoff(
     cm.topology = topology.outofplane
     cm.internal_function = assignments.graph_assignment_geometry_outofplanes
     cm.derivative_function = assignments.graph_assignment_jacobian_outofplanes
+
+    for i in range(len(d["parameters"])):
+        d["parameters"][i] = d["parameters"][i]["Improper"]
+
     smirnoff_dihedral_load(cm, pcp, d)
 
     return cm
@@ -347,6 +358,227 @@ def chemical_model_electrostatics_smirnoff(d: Dict, pcp) -> mm.chemical_model:
 
     return cm
 
+def chemical_model_to_xml_dict(csys):
+
+    name_to_model = {
+        "Bonds": "Bonds",
+        "Angles": "Angles",
+        "Torsions": "ProperTorsion",
+        "OutOfPlanes": "ImproperTorsion",
+        "vdW": "vdW",
+        "LibraryCharges": "LibraryCharge",
+        "Constraints": "Constraint",
+        "ChargeIncrementModel": "ChargeIncrement"
+    }
+    model_to_parameter = {
+        "Bonds": "Bond",
+        "Angles": "Angle",
+        "ProperTorsions": "ProperTorsion",
+        "ImproperTorsions": "ImproperTorsion",
+        "vdW": "Atom",
+        "LibraryCharges": "LibraryCharge",
+        "Constraints": "Constraint",
+        "ChargeIncrementModel": "ChargeIncrement"
+    }
+    model_keys = {
+        "Bonds": ["smirks", "id", "k", "length"],
+        "Angles": ["smirks", "id", "k", "angle"],
+        "ProperTorsions": ["smirks", "id", "periodicity", "k", "phase"],
+        "ImproperTorsions": ["smirks", "id", "periodicity", "k", "phase"],
+        "vdW": ["smirks", "id", "epsilon", "sigma"],
+        "LibraryCharges": ["smirks", "id", "charge"],
+        "Constraints": ["smirks", "id", "c"],
+        "ChargeIncrementModel": ["smirks", "id", "charge"]
+    }
+    model_multi_parameter = {
+        "Bonds": False,
[…]
