+++ besmarts-core/python/besmarts/mechanics/smirnoff_xml.py
@@ -4,33 +4,36 @@ besmarts.mechanics.smirnoff_xml
 """
 
 from typing import List, Dict
-from xml.etree import ElementTree, Element
+from besmarts.core import trees
+from besmarts.core import tree_iterators
+from xml.etree.ElementTree import ElementTree, Element, iterparse, indent
 
 def smirnoff_xml_read(fname: str) -> Dict:
 
     xml = {}
 
-    has_parameters = [
-        "Bonds",
-        "Angles",
-        "ProperTorsions",
-        "ImproperTorsions",
-        "vdW",
-        "LibraryCharges",
-        "Constraints"
-        "ChargeIncrementModel"
-    ]
+    has_parameters = {
+        "Bonds": "Bond",
+        "Angles": "Angle",
+        "ProperTorsions": "Proper",
+        "ImproperTorsions": "Improper",
+        "vdW": "Atom",
+        "LibraryCharges": "LibraryCharge",
+        "Constraints": "Constraint",
+        "ChargeIncrementModel": "ChargeIncrement"
+    }
 
     active = ""
 
-    for event, element in ElementTree.iterparse(fname, ["start", "end"]):
+    for event, element in iterparse(fname, ["start", "end"]):
         # print(element.tag, event, active, element.attrib)
 
         if active and event == "start":
             if element.tag in has_parameters:
                 xml[active]["options"] = dict(element.attrib)
             else:
-                xml[active]["parameters"].append(dict(element.attrib))
+                p = {has_parameters[active]: dict(element.attrib)}
+                xml[active]["parameters"].append(p)
         elif element.tag in has_parameters:
             if event == "start":
                 active = element.tag
@@ -46,69 +49,23 @@ def smirnoff_xml_read(fname: str) -> Dict:
 
     return xml
 
-def smirnoff_xml_write_version_0p3(xml: Dict, fname: str):
+def smirnoff_xml_write(xml: Dict, fname: str):
     """
-    xml = {
-        Bonds: tag: {
-            options: {k:v}, # store in attribs
-            parameters: [ # these are subelements
-                Bond: tag: {k:v} # terms are attribs
-            ]
-    }
-
     """
-    model_to_parameter = {
-        "Bonds": "Bond",
-        "Angles": "Angle",
-        "ProperTorsions": "ProperTorsion",
-        "ImproperTorsions": "ImproperTorsion",
-        "vdW": "Atom",
-        "LibraryCharges": "LibraryCharge",
-        "Constraints": "Constraint",
-        "ChargeIncrementModel": "ChargeIncrement"
-    }
-    model_keys = {
-        "Bonds": ["smirks", "id", "k", "length"],
-        "Angles": ["smirks", "id", "k", "angle"],
-        "ProperTorsions": ["smirks", "id", "periodicity", "k", "phase"],
-        "ImproperTorsions": ["smirks", "id", "periodicity", "k", "phase"],
-        "vdW": ["smirks", "id", "epsilon", "sigma"],
-        "LibraryCharges": ["smirks", "id", "charge"],
-        "Constraints": ["smirks", "id", "c"],
-        "ChargeIncrementModel": ["smirks", "id", "charge"]
-    }
-    model_multi_parameter = {
-        "Bonds": False,
-        "Angles": False,
-        "ProperTorsions": True,
-        "ImproperTorsions": True,
-        "vdW": False,
-        "LibraryCharges": True,
-        "Constraints": True,
-        "ChargeIncrementModel": True
-    }
 
-    root_attrib = {
-        "version"="0.3" "aromaticity_model":"MDL"
-    }
-
-    root = Element("SMIRNOFF", attrib=root_attrib)
[…]
