+++ besmarts-core/python/besmarts/cluster/cluster_objective.py
@@ -6,7 +6,21 @@ Objective functions to use for optimizing cluster hierarchies.
 
 from typing import Sequence
 import itertools
+from besmarts.core import arrays
 from besmarts.core.clusters import clustering_objective
+
+def sum_obj(A):
+
+    obj = []
+    for a in A:
+        if not obj:
+            obj.extend(a)
+        else:
+            obj = arrays.array_add(obj, a)
+    return obj
+
+def format_array(a, s: str):
+    return "".join(map(s.format, a))
 
 
 class clustering_objective_mean_separation(clustering_objective):
@@ -21,14 +35,19 @@ class clustering_objective_mean_separation(clustering_objective):
         self, A: Sequence[Sequence[float]], B: Sequence[Sequence[float]], overlap=0.0
     ) -> float:
 
-        A: Sequence[float] = list(itertools.chain.from_iterable(A))
-        B: Sequence[float] = list(itertools.chain.from_iterable(B))
+        # A: Sequence[float] = list(itertools.chain.from_iterable(A))
+        # B: Sequence[float] = list(itertools.chain.from_iterable(B))
+        A: Sequence[float] = list(A)
+        B: Sequence[float] = list(B)
         if len(A) == 0 or len(B) == 0:
+            print(f"LEN A: {len(A):.8f} LEN B: {len(B):.8f}")
             return overlap
-        abar = sum(A) / len(A)
-        bbar = sum(B) / len(B)
-        d = abs(abar - bbar)
+
+        abar = arrays.array_scale(sum_obj(A), 1/len(A))
+        bbar = arrays.array_scale(sum_obj(B), 1/len(B))
+        d = max(map(abs, arrays.array_difference(abar, bbar)))
         # print(f"MEAN A: {abar:.8f} MEAN B: {bbar:.8f} ABS DIFF: {d:.8f}")
+        # print(f"MEAN A: {abar} MEAN B: {bbar} ABS DIFF: {d:.8f}")
 
         if d < self.split_separation:
             return d
@@ -39,19 +58,24 @@ class clustering_objective_mean_separation(clustering_objective):
         self, A: Sequence[Sequence[float]], B: Sequence[Sequence[float]], overlap=0.0
     ) -> float:
 
-        A: Sequence[float] = list(itertools.chain.from_iterable(A))
-        B: Sequence[float] = list(itertools.chain.from_iterable(B))
+        # A: Sequence[float] = list(itertools.chain.from_iterable(A))
+        # B: Sequence[float] = list(itertools.chain.from_iterable(B))
+        A: Sequence[float] = list(A)
+        B: Sequence[float] = list(B)
         if len(A) == 0:
             abar = 0.0
         else:
-            abar = sum(A) / len(A)
+            abar = arrays.array_scale(sum_obj(A), 1/len(A))
+            # abar = sum(A) / len(A)
 
         if len(B) == 0:
             bbar = 0.0
         else:
-            bbar = sum(B) / len(B)
+            bbar = arrays.array_scale(sum_obj(B), 1/len(B))
+            # bbar = sum(B) / len(B)
 
-        d = abs(abar - bbar)
+        d = max(map(abs, arrays.array_difference(abar, bbar)))
+        # d = abs(abar - bbar)
 
         if d < self.merge_separation:
             return -d
@@ -60,23 +84,34 @@ class clustering_objective_mean_separation(clustering_objective):
 
     def report(self, A: Sequence[Sequence[float]]) -> str:
 
-        A: Sequence[float] = list(itertools.chain.from_iterable(A))
+        # A: Sequence[float] = list(itertools.chain.from_iterable(A))
+        A: Sequence[float] = list(A)
         if len(A) == 0:
-            abar = 0
-            avar = 0
[…]
