+++ besmarts-core/python/besmarts/core/clusters.py
@@ -91,6 +91,8 @@ def get_objective(cst, assn, objfn, edits, splitting=True):
         m = hidx.index.above.get(n.index, None)
         if m is not None:
             m = hidx.index.nodes[m]
+            if m.type != "parameter":
+                continue
             n_match = new_match[n.name]
             m_match = new_match[m.name]
             # if not (n_match and m_match):
@@ -99,7 +101,7 @@ def get_objective(cst, assn, objfn, edits, splitting=True):
             #     continue
             n_group = tuple(((assn[i] for i in n_match)))
             m_group = tuple(((assn[i] for i in m_match)))
-            # print(f"Objective for Sj: {sma}")
+            # print(f"Objective for Sj: {n.name} {m.name} ->")
             obj = objfn(n_group, m_group, overlap=edits)
             if False and splitting:
                 if obj >= 0.0:
@@ -107,6 +109,7 @@ def get_objective(cst, assn, objfn, edits, splitting=True):
                 # print(f"Object increased to {obj} for {n.name} parent {m.name}")
                 # continue
             X += obj
+    # print("Total objective:", X)
     return keep, X
 
 
@@ -246,6 +249,7 @@ def perform_operations(
         group_number,
         Sj_sma,
         strategy,
+        prefix="p"
     ):
 
     topo = hidx.topology
@@ -259,14 +263,15 @@ def perform_operations(
         added = False
 
         if step.operation == strategy.SPLIT:
-            param_name = "p" + str(group_number)
+            param_name = prefix + str(group_number)
 
             # print(datetime.datetime.now(), '*** 2')
             hent = hidx.index.node_add(
                 S.index,
-                trees.tree_node(None, "parameter", "", param_name),
+                trees.tree_node(None, tuple((0,0,0)), "parameter", param_name),
                 index=0,
             )
+
             # print(datetime.datetime.now(), '*** 3')
             Sj = graphs.subgraph_relabel_nodes(
                 Sj, {n: i for i, n in enumerate(Sj.select, 1)}
@@ -471,11 +476,9 @@ def smarts_clustering_optimize(
     """
 
     group_number = max(
-        [
-            int(x.name[1:])
-            for x in hidx.index.nodes.values()
-            if x.name[0] == group_prefix_str
-        ]
+        [0] + [*map(int, [x for x in 
+            ["".join(a for a in x.name[1:] if a.isdigit())
+            for x in hidx.index.nodes.values()] if x])]
     )
 
     group_number += 1
@@ -541,12 +544,12 @@ def smarts_clustering_optimize(
         assignments,
         clustering_build_assignment_mappings(hidx, assignments),
     )
+    cst.group_prefix_str = group_prefix_str
     print(f"{datetime.datetime.now()} Checking consistency...")
     check_lbls_data_selections_equal(cst.group, sag)
     _, X0 = get_objective(
         cst, assn, objective.split, strategy.overlaps[0], splitting=True
     )
-
 
     if not strategy.steps:
         print("Optimization strategy is building steps...")
@@ -587,8 +590,7 @@ def smarts_clustering_optimize(
         nodes = [
             x
             for x in strategy.tree_iterator(cst.hierarchy.index, roots)
-            if strategy.cursor == -1
-            or strategy.cursor >= step_tracker.get(x.name, 0)
+            if x.type == "parameter" and cst.hierarchy.subgraphs.get(x.index) is not None and (strategy.cursor >= step_tracker.get(x.name, -1))
         ]
 
[…]
