### 2024-07-28 549ac87
functions to collection FCs
