+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -1464,7 +1464,9 @@ class forcefield_optimization_strategy(optimization.optimization_strategy):
         # After splitting and modifying periodicities, reset to a more 
         self.enable_dihedral_periodicity_reset = True
         self.dihedral_periodicity_reset_max_n = 6
-        self.dihedral_periodicity_reset_alpha = -0.25
+        self.dihedral_periodicity_reset_alpha = -.5
+        self.dihedral_periodicity_reset_min_k = 1e-3
+        self.dihedral_periodicity_reset_max_k = 5.0
 
         self.reference_list = []
 
@@ -1793,13 +1795,25 @@ def ff_optimize(
         "angle_l": strategy.enable_reset_angle_lengths,
         "angle_k": strategy.enable_reset_angle_stiffness,
         "torsion_k": strategy.enable_reset_torsion_stiffness,
-        "outofplane_k": strategy.enable_reset_outofplane_stiffness
+        "outofplane_k": strategy.enable_reset_outofplane_stiffness,
+        "dihedral_p": strategy.enable_dihedral_periodicity_reset,
+        "dihedral_max_n": strategy.dihedral_periodicity_reset_max_n,
+        "dihedral_alpha": strategy.dihedral_periodicity_reset_alpha,
+        "dihedral_min_k": strategy.dihedral_periodicity_reset_min_k,
+        "dihedral_max_k": strategy.dihedral_periodicity_reset_max_k
     }
     wq = compute.workqueue_local('0.0.0.0', configs.workqueue_port)
 
     # print(datetime.datetime.now(), "Resetting bonds and angles to reference values")
     ws = compute.workqueue_new_workspace(wq)
-    psystems = reset(reset_config, csys0, gdb, psystems, verbose=True, ws=ws)
+    psystems = reset(
+        reset_config,
+        csys0,
+        gdb,
+        psystems,
+        verbose=True,
+        ws=ws
+    )
     ws.close()
 
     csys = copy.deepcopy(csys0)
@@ -1829,8 +1843,6 @@ def ff_optimize(
     print("Initial parameter assignments of dataset:")
     print_chemical_system(csys, show_parameters=[x[1] for x in assigned_nodes])
 
-
-    # we can't split scales (yet) or cutoffs
     assigned_nodes = [x for x in assigned_nodes if x[0] in strategy.bounds and x[1] not in "sc"] 
     assigned_nodes = set([x[1] for x in assigned_nodes])
 
@@ -1895,7 +1907,13 @@ def ff_optimize(
                     psys.models[m].values.pop()
                     psys.models[m].labels.pop()
                 for proc in procs[1:]:
-                    psys.models[m] = proc.assign(csys.models[m], psys.models[m], overrides={k[1:]: v for k, v in kv.items() if k[0] == m})
+                    psys.models[m] = proc.assign(
+                        csys.models[m],
+                        psys.models[m],
+                        overrides={
+                            k[1:]: v for k, v in kv.items() if k[0] == m
+                        }
+                    )
 
     step_tracker = strategy.step_tracker
 
@@ -1919,7 +1937,7 @@ def ff_optimize(
             # and x.type == "parameter"
         ]
         # remove any that are not in the models
-        
+
         print(f"Targets for this macro step {strategy.cursor+1}:")
         for nidx, n in enumerate(nodes, 1):
             print(nidx, n.category, n.name)
@@ -1968,14 +1986,17 @@ def ff_optimize(
             gid = list(gde.tables[tid].graphs)[0]
             rid = 0
             pos = assignments.graph_db_graph_to_graph_assignment(gdb, eid, tid, gid, rid)
-            psys[eid] = mm.chemical_system_to_physical_system(csys, [pos], ref=psystems[eid], reuse=reuse)
[…]
