### 2024-07-28 56e1dec
support to drop linear torsions in B
