### 2024-07-28 d0fd960
turn off am1bcc minimization
