+++ besmarts-core/python/besmarts/codecs/codec_native.py
@@ -20,8 +20,6 @@ topology_tab = {
     "ATOM": topology.atom_topology(),
     "BOND": topology.bond_topology(),
     "ANGLE": topology.angle_topology(),
-    # "PROPER": topology.torsion_topology(),
-    # "IMPROPER": topology.outofplane_topology(),
     "TORSION": topology.torsion_topology(),
     "OUTOFPLANE": topology.outofplane_topology(),
 }
@@ -172,10 +170,12 @@ def graph_codec_native_read(f) -> Sequence:
 
     return graphs
 
+
 def graph_codec_native_load(fname) -> Sequence:
 
     with open(fname) as f:
         return graph_codec_native_read(f)
+
 
 def graph_codec_native_write(f, graphs):
 
@@ -195,6 +195,17 @@ def graph_codec_native_encode(graphs):
 
 
 class graph_codec_native(codecs.graph_codec):
+
+    """
+    The native graph codec implements the SMARTS and SMILES encoders which
+    can transform SMARTS primitives in binary form to string form.
+
+    To use this interface, supply a dictionary of primitive codecs that know how
+    to encode/decode the primitives, and then supply the initial list of
+    primitives that will be used when encoding SMARTS. Manipulating the lists
+    controls which primitives are active.
+    """
+
     def __init__(
         self,
         primitive_codecs: Dict[
@@ -203,6 +214,25 @@ class graph_codec_native(codecs.graph_codec):
         atom_primitives: Sequence[primitives.primitive_key],
         bond_primitives: Sequence[primitives.primitive_key],
     ):
+
+        """
+        Constructor initializer.
+
+        Parameters
+        ----------
+        primitive_codecs
+        The primitives that the codec will have access to
+
+        atom_primitives
+        The atom primitives that will be active when encoding is performed
+
+        bond_primitives
+        The bond primitives that will be active when encoding is performed
+
+        Returns
+        -------
+        graph_codec_native
+        """
 
         self.primitive_codecs = primitive_codecs
 
@@ -278,6 +308,18 @@ class graph_codec_native(codecs.graph_codec):
 
 
 def primitive_codecs_get() -> Dict[codecs.primitive_key, codecs.primitive_codec]:
+    """
+    Return the primitives that the BESMARTS native codec is aware of.
+
+    Parameters
+    ----------
+
+    Returns
+    -------
+    Dict[codecs.primitive_key, codecs.primitive_codec]
+    The map of primitive keys (e.g. "element") and the respective codec that
+    can encode/decode into binary form
+    """
 
     codecs_ = {}
     codecs_.update(primitive_codecs_get_atom())
@@ -285,7 +327,13 @@ def primitive_codecs_get() -> Dict[codecs.primitive_key, codecs.primitive_codec]
 
     return codecs_
 
-def primitive_codecs_get_atom() -> Dict[codecs.primitive_key, codecs.primitive_codec]:
+
+def primitive_codecs_get_atom(
+) -> Dict[codecs.primitive_key, codecs.primitive_codec]:
+    """
+    Return the node (atom) primitives that the BESMARTS native codec is aware
+    of.
+    """
     codecs_ = {
         "element": codecs.primitive_codec_element(),
         "hydrogen": codecs.primitive_codec_hydrogen(),
@@ -299,7 +347,13 @@ def primitive_codecs_get_atom() -> Dict[codecs.primitive_key, codecs.primitive_c
     }
     return codecs_
 
-def primitive_codecs_get_bond() -> Dict[codecs.primitive_key, codecs.primitive_codec]:
+
+def primitive_codecs_get_bond(
+) -> Dict[codecs.primitive_key, codecs.primitive_codec]:
+    """
[…]
