### 2024-07-29 dfefabe
add subgraph to ic functions
