+++ besmarts-core/python/examples/smarts_cluster_bond_lengths.py
@@ -4,6 +4,10 @@ from besmarts.cluster.cluster_objective import clustering_objective_mean_separat
 from besmarts.cluster.cluster_optimization import cluster_means
 from besmarts.codecs.codec_rdkit import graph_codec_rdkit
 from besmarts.assign import hierarchy_assign_rdkit
+from besmarts.core import configs
+
+configs.remote_compute_enable = False
+configs.workqueue_port = 59321
 
 gcd = graph_codec_rdkit()
 labeler = hierarchy_assign_rdkit.smarts_hierarchy_assignment_rdkit()
@@ -11,12 +15,12 @@ objective = clustering_objective_mean_separation(split_separation=0.1)
 
 smi = "[C:1]([H:3])#[C:2][H:4]"
 assns = {
-    (1,2): [1.1],
-    (1,3): [1.3],
-    (2,4): [1.3]
+    (1, 2): [1.1],
+    (1, 3): [1.3],
+    (2, 4): [1.3]
 }
 
-sa  = smiles_assignment_float(smi, assns)
+sa = smiles_assignment_float(smi, assns)
 
 sag = smiles_assignment_group_bonds([sa])
 cst = cluster_means(gcd, labeler, sag, objective=objective)
