+++ besmarts-core/python/besmarts/core/trees.py
@@ -6,6 +6,7 @@ Graphs with no cycles, a rough concept of directed edges, and have a root nodes
 
 from typing import Dict, List
 
+
 class tree_node:
     __slots__ = ("index", "category", "type", "name")
 
