+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -13,7 +13,7 @@ import sys
 import math
 import itertools
 import numpy as np
-from typing import List, Dict, Tuple, Callable
+from typing import List, Dict, Callable
 from besmarts.core import configs
 from besmarts.core import arrays
 from besmarts.core import assignments
@@ -41,15 +41,19 @@ from besmarts.mechanics import hessians
 from besmarts.mechanics import vibration
 from besmarts.mechanics import molecular_models as mm
 
+eid_t = int
 
 class compute_config:
 
     """
-    Default is single point energy
+    Configuration for computing some property of an entry in a graph_db. The
+    addr is the address of the entry in the graph_db. Keys are from a chemical
+    system, and are used to temporarility update the parameters. This is
+    generally used to computer numerical gradients and hessians.
     """
 
     def __init__(self, addr):
-        self.addr = addr 
+        self.addr = addr
         self.keys = []
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_table]]:
@@ -58,33 +62,21 @@ class compute_config:
         GDB = self.GDB
         tbl_idx = assignments.ENERGY
         tid = assignments.POSITIONS
-        
 
         all_results = []
         for keyset in self.keys:
             results: List[Dict[assignments.tid_t, assignments.graph_db_table]] = []
 
-            # for k, v in self.keys.items():
-                # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.6g} to {v:.6g} d={v-mm.chemical_system_get_value(csys, k):.6g}")
-                # mm.chemical_system_set_value(csys, k, v)
-                # mm.physical_system_set_value(self.psys, k, v)
-
             for eid, gdb in GDB.entries.items():
                 tbl = assignments.graph_db_table(topology.null)
                 for gid in gdb.graphs:
-                    # pos0 = assignments.graph_db_entry_to_graph_assignment(gdb, tbl_idx, gid)
                     rid = 0
                     pos0 = assignments.graph_db_graph_to_graph_assignment(gdb, eid, tid, gid, rid)
-
-                    # this means fold all graphs into a single graph
-                    # pos0 = assignments.graph_db_entry_to_graph_assignment(gdb, eid, tid)
 
                     psys = mm.chemical_system_to_physical_system(csys, [pos0], ref=self.psys, reuse=self.reuse)
                     reapply = set()
                     for k, v in keyset.items():
                         # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.10g} to {v:.10g} d={v-mm.chemical_system_get_value(csys, k):.10g}")
-                        # mm.chemical_system_set_value(csys, k, v)
-                        # mm.chemical_system_set_value(csys, k, v)
                         mm.physical_system_set_value(psys, k, v)
                         reapply.add(k[0])
                     for m in reapply:
@@ -97,9 +89,7 @@ class compute_config:
                                 psys.models[m] = proc.assign(csys.models[m], psys.models[m], overrides={k[1:]: v for k, v in keyset.items() if k[0] == m})
                     ene = objectives.physical_system_energy(psys, csys)
                     
-                    # ene_ga = assignments.graph_assignment("", {(0,): [[energy]]}, gdb.graphs[gid])
-                    # gdg = assignments.graph_assignment_to_graph_db_graph(ene_ga, topology.null)
-                    tbl[gid] = gdg
+                    tbl[gid] = ene
                 r = {tbl_idx: tbl}
                 # print("Done calculating", self)
                 # print("Result is\n", r)
@@ -114,15 +104,15 @@ class compute_config_energy(compute_config):
     """
 
[…]
