+++ besmarts-core/python/examples/init_bonds_angles.py
@@ -1,0 +1,180 @@
+"""
+examples/init_bonds_angles.py
+"""
+
+import os
+import tempfile
+from besmarts.mechanics import molecular_models as mm
+from besmarts.mechanics import smirnoff_models
+from besmarts.perception import perception_rdkit
+
+pcp = perception_rdkit.perception_model_rdkit()
+
+FF = """
+<SMIRNOFF version="0.3" aromaticity_model="OEAroModel_MDL">
+    <Bonds
+        version="0.4"
+        potential="harmonic"
+        fractional_bondorder_method="AM1-Wiberg"
+        fractional_bondorder_interpolation="linear"
+    >
+        <Bond
+            smirks="[*:1]~[*:2]"
+            id="b1"
+            length="1.5 * angstrom"
+            k="500.0 * angstrom**-2 * mole**-1 * kilocalorie"
+        ></Bond>
+        <Bond
+            smirks="[*:1]~[#1:2]"
+            id="b2"
+            length="1.0 * angstrom"
+            k="700.0 * angstrom**-2 * mole**-1 * kilocalorie"
+        ></Bond>
+    </Bonds>
+    <Angles version="0.3" potential="harmonic">
+        <Angle
+            smirks="[*:1]~[X4:2]~[*:3]"
+            angle="109.5 * degree"
+            k="60.0 * mole**-1 * radian**-2 * kilocalorie"
+            id="a1"
+        ></Angle>
+        <Angle
+            smirks="[*:1]~[X3:2]~[*:3]"
+            angle="120.0 * degree"
+            k="80.0 * mole**-1 * radian**-2 * kilocalorie"
+            id="a2"
+        ></Angle>
+        <Angle
+            smirks="[*:1]~[X2:2]~[*:3]"
+            angle="180.0 * degree"
+            k="10.0 * mole**-1 * radian**-2 * kilocalorie"
+            id="a3"
+        ></Angle>
+    </Angles>
+    <ProperTorsions
+        version="0.4"
+        potential="k*(1+cos(periodicity*theta-phase))"
+        default_idivf="auto"
+        fractional_bondorder_method="AM1-Wiberg"
+        fractional_bondorder_interpolation="linear"
+    >
+        <Proper
+            smirks="[*:1]~[*:2]~[*:3]~[*:4]"
+            id="t1"
+            periodicity1="3"
+            phase1="0.0 * degree"
+            k1="0.15 * mole**-1 * kilocalorie"
+            idivf1="1.0"
+        ></Proper>
+    </ProperTorsions>
+    <ImproperTorsions 
+        version="0.3"
+        potential="k*(1+cos(periodicity*theta-phase))"
+        default_idivf="auto"
+    > </ImproperTorsions>
+    <vdW
+        version="0.3"
+        potential="Lennard-Jones-12-6"
+        combining_rules="Lorentz-Berthelot"
+        scale12="0.0"
+        scale13="0.0"
+        scale14="0.5"
+        scale15="1.0"
+        cutoff="9.0 * angstrom"
+        switch_width="1.0 * angstrom"
+        method="cutoff"
+    >
+        <Atom
+            smirks="[*:1]"
+            id="n1"
[…]
