+++ besmarts-core/python/examples/besmarts_fit.py
@@ -1,0 +1,637 @@
+"""
+examples/besmarts_fit.py
+
+This performs automated chemical perception starting from OpenFF Sage 2.1. For
+expedience, only b4 is targeted and only the lengths are fit. The resulting
+force field was fit on positions and gradients of a single molecule.
+"""
+
+import tempfile
+import os
+from typing import Dict, List
+from besmarts.mechanics import fits
+from besmarts.mechanics import smirnoff_models
+from besmarts.mechanics import molecular_models as mm
+from besmarts.core import graphs
+from besmarts.core import topology
+from besmarts.core import perception
+from besmarts.core import arrays
+from besmarts.core import assignments
+from besmarts.assign import hierarchy_assign_rdkit
+from besmarts.codecs import codec_rdkit
+from besmarts.core import configs
+
+configs.processors = 4
+configs.remote_compute_enable = False
+configs.workqueue_port = 54321
+
+xyz_positions = """11
+
+  C -1.44819400 -0.84940800  0.16848900
+  C -1.59401300  0.50318700 -0.01678100
+  C -0.27397600  1.02622600 -0.13503500
+  C  0.58064400 -0.04716400 -0.01303100
+  C  2.03461200 -0.06860900 -0.05925200
+  O  2.72809700  0.90108700 -0.21909900
+ Cl  2.76214600 -1.70734100  0.14655600
+  O -0.13897300 -1.20044600  0.17351800
+  H -2.15226800 -1.65836100  0.30609000
+  H -2.52743000  1.04809900 -0.06180000
+  H  0.02935200  2.05273200 -0.28965800
+"""
+xyz_grad = """11
+
+  C      0.49755    0.17370   -0.04115
+  C     -0.00884    0.07632   -0.01031
+  C      0.20074   -0.69547    0.09073
+  C     -0.02955    1.24848   -0.17483
+  C      0.55229   -1.91119    0.25039
+  O     -0.15948    0.65794   -0.08724
+ Cl     -0.33030    0.82983   -0.10559
+  O     -0.73720   -0.66864    0.11909
+  H     -0.11502    0.11021   -0.01168
+  H     -0.00691    0.04737   -0.00649
+  H      0.02566   -0.05163    0.00657
+"""
+
+
+def load_xyz(flist, indices=None) -> assignments.graph_db_row:
+    s = 0
+    N = None
+    gdr = assignments.graph_db_row()
+    for f in flist:
+        lines = f.split('\n')
+        if N is None:
+            N = int(lines[0].split()[0])
+        if indices is None:
+            indices = [*range(1,N+1)]
+        assert N == int(lines[0].split()[0])
+        for chunk in arrays.batched(lines, N+2):
+            if chunk and chunk[0]:
+                sel = {}
+                for i, line in enumerate(chunk, -2):
+                    if i >= 0:
+                        sym, x, y, z = line.split()[:4]
+                        sel[indices[i],] = list(map(float, (x, y, z)))
+
+                gdc = assignments.graph_db_column()
+                gdc.selections.update(sel)
+                gdr.columns[s] = gdc
+                s += 1
+    return gdr
+
+def make():
[…]
