### 2024-07-29 09f1551
add gcd.rdmol_decode
