### 2024-07-29 7a09265
add gcd.rdmol_decode
