### 2024-07-29 05b45c1
set graph codec aromaticity from offxml
