+++ besmarts-core/python/besmarts/core/assignments.py
@@ -2,11 +2,34 @@
 besmarts.core.assignments
 
 Associates BESMARTS structures to data
+
+There are two intepretations of the data. The first simpler model is the
+the graph_assignments, which contain (ic, data) pairs. The data is assumed to
+be a list of lists of independent samples, i.e. multiple conformers. This
+distinction means a "system" where there are two molecules interacting requires
+a list of graph assignments. These assignments are quite flexible and agnostic
+of the data. An additional layer is required to interpret data, which is
+provided, at least partially, to graph_db objects.
+
+In constrast to the simple assignment objectives, the more complex graph_db
+system is explicitly designed to handle large systems with redundant graph
+information, e.g. a box of water where point triples use the same graph, and
+each triple exists in the same system and should be overlapping. See the
+documentation for graph_db for more information.
+
+In general, most of the work is performed on assignments, and graph_db objects
+are designed to be a management system for the various types of data assigned to
+graphs, such as positions, gradients, and hessians.
 """
 
-from typing import List, Dict, Sequence, Tuple, Any
-from besmarts.core import graphs, topology, codecs, hierarchies, geometry
 import math
+from typing import List, Dict, Sequence, Tuple, Any
+from besmarts.core import graphs
+from besmarts.core import topology
+from besmarts.core import codecs
+from besmarts.core import hierarchies
+from besmarts.core import geometry
+from besmarts.core import primitives
 
 assignment_mapping = Dict[str, List[List[Sequence[int]]]]
 cartesian_coordinates = List[List[Tuple[float, float, float]]]
@@ -26,7 +49,6 @@ RADII = 11
 PHYSICAL_MODEL_BOND_LABELS = 12
 
 ASSN_NAMES = {
-    # floats
     POSITIONS : "positions", # atom, 1x3 
     GRADIENTS : "gradients", # atom, 1x3
     HESSIANS : "hessians", # bond, 1x9
@@ -52,7 +74,20 @@ pid_t = int # topology id
 xid_t = int
 
 class graph_db_address:
-    def __init__(self, *, eid=None, tid=None, gid=None, rid=None, cid=None, sid=None, xid=None):
+    """
+    A query for information in a graph_db.
+    """
+    def __init__(
+        self,
+        *,
+        eid=None,
+        tid=None,
+        gid=None,
+        rid=None,
+        cid=None,
+        sid=None,
+        xid=None
+    ):
         self.eid: list = [] if not eid else eid
         self.tid: list = [] if not tid else tid
         self.gid: list = [] if not gid else gid
@@ -60,6 +95,7 @@ class graph_db_address:
         self.cid: list = [] if not cid else cid
         self.sid: list = [] if not sid else sid
         self.xid: list = [] if not xid else xid
+
 
 class graph_db_column:
 
@@ -89,6 +125,7 @@ class graph_db_row:
     def __setitem(self, cid: cid_t, columns):
         self.columns[cid] = columns
 
+
 class graph_db_graph:
 
     __slots__ = "rows",
@@ -101,6 +138,7 @@ class graph_db_graph:
 
     def __setitem(self, rid: rid_t, rows):
         self.columns[rid] = rows
+
 
 class graph_db_table:
     """
@@ -119,6 +157,7 @@ class graph_db_table:
     def __setitem(self, gid: rid_t, graph: graph_db_graph):
         self.graphs[gid] = graph
 
+
[…]
