+++ besmarts-core/python/besmarts/core/graphs.py
@@ -2667,3 +2667,11 @@ def structure_to_intvec(g: structure, atom_primitives, bond_primitives) -> array
     intvec = subgraph_to_intvec(g, atom_primitives, bond_primitives)
     intvec.v[4] = topology.index_of(g.topology)
     return intvec
+
+
+def graph_complexity(g: graph, scale=1.0, offset=0.0):
+    # C = len(g.nodes) + graphs.graph_bits_max(g) / graphs.graph_bits(g)  / len(g.nodes) / 100
+    # return scale*C + offset
+
+    C = graph_bits(g)  / len(g.nodes)
+    return C
