+++ besmarts-core/python/besmarts/mechanics/molecular_models.py
@@ -2,11 +2,8 @@
 besmarts.mechanics.molecular_models
 """
 
-from typing import Dict, List, Any
-import datetime
 import copy
-import math
-
+from typing import Dict, List, Any
 
 from besmarts.core import graphs
 from besmarts.core import trees
@@ -17,7 +14,16 @@ from besmarts.core import perception
 
 
 class topology_term:
-    def __init__(self, symbol, name, unit, cast, values, comment, value_comments):
+    def __init__(
+        self,
+        symbol,
+        name,
+        unit,
+        cast,
+        values,
+        comment,
+        value_comments
+    ):
         self.symbol: str = symbol
         self.name: str = name
         self.unit: str = unit
@@ -26,8 +32,9 @@ class topology_term:
         self.comment: str = comment
         self.value_comments: Dict[int, str] = value_comments
 
-    def copy():
+    def copy(self):
         return topology_term_copy(self)
+
 
 def topology_term_copy(t: topology_term):
     return topology_term(
@@ -36,9 +43,10 @@ def topology_term_copy(t: topology_term):
         t.unit,
         t.cast,
         t.values.copy(),
-        t.comment, 
+        t.comment,
         t.value_coments.copy()
     )
+
 
 class system_term:
     def __init__(self, name, symbol, unit, cast, values, comment=""):
@@ -49,14 +57,16 @@ class system_term:
         self.values: List = values
         self.comment: str = comment
 
+
 class physical_model:
-    """The functional form that can be evaluated as a function of positions.
+    """
+    The functional form that can be evaluated as a function of positions.
     Also tracks the positions as a cache mechanism to avoid recomputing
     Sort of assumes energies/forces are a function of internal coordinates
-    and masses and positions are cartesian"""
+    and masses and positions are cartesian
+    """
 
-    __slots__ =  "labels", "values", "positions", "topology"
-
+    __slots__ = "labels", "values", "positions", "topology"
 
     def __init__(self, positions, labels, values):
 
@@ -88,13 +98,13 @@ class physical_model_procedure:
     """
     calculates one or more physical properties of a system
     returns a bunch tables
-
     """
+
     def __init__(self, name, topo):
-        self.name = name # HESSIANS
+        self.name = name
         self.topology = topo
         self.procedure_parameters: Dict[str, int] = {}
-    
+
     def assign(self, pm: physical_model):
         """
         this will return the compute function and config (i.e. the task)
@@ -107,6 +117,7 @@ class physical_model_procedure:
 
     def get_term_values(self, key) -> Dict:
         assert False
+
 
 class chemical_model:
     def __init__(self, symbol, name, topo):
@@ -129,17 +140,23 @@ class physical_system:
     def __init__(self, models: List[physical_model]):
         self.models = models
 
+
 class chemical_system:
-    def __init__(self, pcp_model: perception.perception_model, models: List[chemical_model]):
+    def __init__(
+        self,
+        pcp_model: perception.perception_model,
+        models: List[chemical_model]
+    ):
         self.perception = pcp_model
         self.models = models
+
 
 class chemical_model_procedure_smarts_assignment(chemical_model_procedure):
     def __init__(self, pcp_model: perception.perception_model, topology_terms):
         self.name = ""
         self.perception = pcp_model
 
[…]
