### 2024-07-30 0e9987d
add minimization example
