### 2024-07-30 980055a
add minimization example
