+++ besmarts-core/python/besmarts/core/assignments.py
@@ -1313,3 +1313,50 @@ def graph_assignment_to_format_xyz(
             x, y, z = xyz[:3]
         lines.append(f"{sym:8s} {x:.6f} {y:.6f} {z:.6f}")
     return lines
+
+
+def parse_xyz(xyzdata: List[str]) -> Tuple[List[str], cartesian_coordinates]:
+
+    N = None
+    lines = xyzdata.split('\n')
+
+    if N is None:
+        N = int(lines[0].split()[0])
+
+    assert N == int(lines[0].split()[0])
+
+    syms = []
+    xyzs = []
+    for chunk in arrays.batched(lines, N+2):
+        sym = [None]*N
+        xyz = [None]*N
+        if chunk and chunk[0]:
+            for i, line in enumerate(chunk, -2):
+                if i >= 0:
+                    s, x, y, z = line.split()[:4]
+                    sym[i] = s
+                    xyz[i] = [[*map(float, (x, y, z))]]
+        if all(sym) and all(xyz):
+            syms.append(sym)
+            xyzs.append(xyz)
+    return syms, xyzs
+
+
+def xyz_to_graph_assignment(
+    gcd,
+    smi,
+    xyzdata: List,
+    indices=None
+) -> graph_assignment:
+    """
+    """
+    g = graphs.subgraph_as_graph(gcd.smiles_decode(smi))
+    sel = {}
+
+    for xyzs in xyzdata:
+        for ic, xyz in enumerate(xyzs, 1):
+            if (ic,) not in sel:
+                sel[ic,] = []
+            sel[ic,].extend(xyz)
+
+    return graph_assignment(smi, sel, g)
