+++ besmarts-openmm/python/besmarts/mechanics/optimizers_openmm.py
@@ -443,44 +443,3 @@ def molecular_dynamics(psys, ts, steps):
     sim.step(steps)
 
     return
-
-
-def print_xyz(pos, comment="") -> List[str]:
-    lines = []
-    lines.append(str(len(pos.selections)))
-    lines.append(comment)
-    for ic, xyz in pos.selections.items():
-        n = pos.graph.nodes[ic[0]]
-        sym = primitives.element_tr[str(n.primitives['element'].on()[0])]
-        try:
-            x, y, z = xyz[0][:3]
-        except TypeError:
-            x, y, z = xyz[:3]
-        lines.append(f"{sym:8s} {x:.6f} {y:.6f} {z:.6f}")
-    return lines
-
-def parse_xyz(xyzdata):
-
-    N = None
-    lines = xyzdata.split('\n')
-
-    if N is None:
-        N = int(lines[0].split()[0])
-
-    assert N == int(lines[0].split()[0])
-
-    syms = []
-    xyzs = []
-    for chunk in arrays.batched(lines, N+2):
-        sym = [None]*N
-        xyz = [None]*N
-        if chunk and chunk[0]:
-            for i, line in enumerate(chunk, -2):
-                if i >= 0:
-                    s, x, y, z = line.split()[:4]
-                    sym[i] = s
-                    xyz[i] = [[*map(float, (x, y, z))]]
-        if all(sym) and all(xyz):
-            syms.append(sym)
-            xyzs.append(xyz)
-    return syms, xyzs
