### 2024-07-30 c6e5901
move xyz processing
