### 2024-07-30 eab6a35
add 10_vibfreq.py
