### 2024-07-30 102acf3
add 10_vibfreq.py
