+++ besmarts-core/python/besmarts/mechanics/vibration.py
@@ -1,12 +1,7 @@
 #!/usr/bin/env python3
-import pickle
-import numpy as np
-# import matplotlib.pyplot as plt
-# from matplotlib.ticker import MultipleLocator, AutoMinorLocator
-# from geometric.internal import *
-# from geometric.molecule import Molecule as geometric_mol
-from scipy.linalg import svd
+
 import sys
+import numpy as np
 
 gpermol2au = 1836
 au2amu = 1/1836.
@@ -23,29 +18,8 @@ au2cm = 5.29177249e-9
 c = 137
 conversion = 1.0 /(2*np.pi * c * au2cm)
 
-def load_xyz(fname):
-
-    xyzdat = np.genfromtxt(fname, skip_header=2, usecols=range(4), dtype=('U2', 'f8', 'f8', 'f8' ))
-    syms = [x[0] for x in xyzdat]
-    xyz  = np.array([list(x)[1:] for x in xyzdat])
-    return syms, xyz
-
 def converteig( eigs):
     return np.sign( eigs) * np.sqrt( np.abs( eigs))*conversion
-
-# connect means whether to remove TR
-#    True -> remove TR
-#    False -> keep TR
-
-# CoordSysDict = {'cart':(CartesianCoordinates, False, False),
-#             'prim':(PrimitiveInternalCoordinates, False, False),
-#             'dlc':(DelocalizedInternalCoordinates, True, False),
-#             'hdlc':(DelocalizedInternalCoordinates, False, True),
-#             'tric':(DelocalizedInternalCoordinates, True, False)}
-# coordsys = 'prim' # kwargs.get('coordsys', 'tric')
-# CVals = None
-# CoordClass, connect, addcart = CoordSysDict[coordsys.lower()]
-# Cons = None
 
 # swap_internals = not connect
 debug = False
@@ -71,18 +45,25 @@ def printmat( mat, dec=4):
         print()
 
 
-mass_table = {'C': 12.011, 'H': 1.008, 'N': 14.007, 'O': 15.999, 'P': 30.973, \
-        'S': 32.06, 'F': 18.998, "Cl": 35.45, "CL": 35.45, "B": 10.81, "I": 126.9, "Br": 79.04, "Na": 22.90 }
-# mass_table = {'C': 12.000, 'H': 1.000, 'N': 14.000, 'O': 16.0, 'P': 31.00, \
-#         'S': 32.00, 'F': 19.0, "Cl": 35, "B": 11.00, "Br": 70.0, "I": 106., "Se": 68.0  }
-# mass_table = {'C': 12.000, 'H': 1.007825032230, 'N': 14.003074004430, 'O': 15.994914619570, 'P': 31.00, \
-#     'S': 32.00, 'F': 19.0, "CL": 35, "B": 11.00 }
-# mass_table_unit = {'C': 1.000, 'H': 1.000, 'N': 1.000, 'O': 1.0, 'P': 1.00, \
-#     'S': 1.00, 'F': 1.0, "Cl": 1.0, "B": 1.00 }
+mass_table = {
+    'C': 12.011,
+    'H': 1.008,
+    'N': 14.007,
+    'O': 15.999,
+    'P': 30.973,
+    'S': 32.06,
+    'F': 18.998,
+    "Cl": 35.45,
+    "B": 10.81,
+    "I": 126.9,
+    "Br": 79.04,
+    "Na": 22.90
+}
+for k, v in list(mass_table.items()):
+    mass_table[k.upper()] = v
 
-# mass_table = mass_table_unit
-
-
+# TODO clean and move these GS to a better home
+# The useful code uses QR so these are no longer needed here
 def gram_schmidt(vectors):
     basis = []
     for v in vectors:
[…]
