+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -28,7 +28,6 @@ from besmarts.core import splits
 from besmarts.core import trees
 from besmarts.core import tree_iterators
 from besmarts.core import optimization
-from besmarts.core import primitives
 from besmarts.core import logs
 from besmarts.core import returns
 from besmarts.cluster import cluster_objective
@@ -43,8 +42,12 @@ from besmarts.mechanics import molecular_models as mm
 
 eid_t = int
 
-class compute_config:
+POSITIONS = assignments.POSITIONS
+HESSIANS = assignments.HESSIANS
+GRADIENTS = assignments.GRADIENTS
 
+
+class compute_config:
     """
     Configuration for computing some property of an entry in a graph_db. The
     addr is the address of the entry in the graph_db. Keys are from a chemical
@@ -57,7 +60,7 @@ class compute_config:
         self.keys = []
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_table]]:
-        # print("Starting calculation", self)
+
         csys = self.csys
         GDB = self.GDB
         tbl_idx = assignments.ENERGY
@@ -65,18 +68,29 @@ class compute_config:
 
         all_results = []
         for keyset in self.keys:
-            results: List[Dict[assignments.tid_t, assignments.graph_db_table]] = []
+            results = []
+            results: List[Dict[assignments.tid_t, assignments.graph_db_table]]
 
             for eid, gdb in GDB.entries.items():
                 tbl = assignments.graph_db_table(topology.null)
                 for gid in gdb.graphs:
                     rid = 0
-                    pos0 = assignments.graph_db_graph_to_graph_assignment(gdb, eid, tid, gid, rid)
+                    pos0 = assignments.graph_db_graph_to_graph_assignment(
+                        gdb,
+                        eid,
+                        tid,
+                        gid,
+                        rid
+                    )
 
-                    psys = mm.chemical_system_to_physical_system(csys, [pos0], ref=self.psys, reuse=self.reuse)
+                    psys = mm.chemical_system_to_physical_system(
+                        csys,
+                        [pos0],
+                        ref=self.psys,
+                        reuse=self.reuse
+                    )
                     reapply = set()
                     for k, v in keyset.items():
-                        # print(f"-> Obj Setting {k} from {mm.chemical_system_get_value(csys, k):.10g} to {v:.10g} d={v-mm.chemical_system_get_value(csys, k):.10g}")
                         mm.physical_system_set_value(psys, k, v)
                         reapply.add(k[0])
                     for m in reapply:
@@ -86,9 +100,16 @@ class compute_config:
                                 psys.models[m].values.pop()
                                 psys.models[m].labels.pop()
                             for proc in procs[1:]:
-                                psys.models[m] = proc.assign(csys.models[m], psys.models[m], overrides={k[1:]: v for k, v in keyset.items() if k[0] == m})
+                                psys.models[m] = proc.assign(
+                                    csys.models[m],
+                                    psys.models[m],
+                                    overrides={
+                                        k[1:]: v for k, v in keyset.items()
+                                        if k[0] == m
+                                    }
+                                )
                     ene = objectives.physical_system_energy(psys, csys)
-                    
+
                     tbl[gid] = ene
                 r = {tbl_idx: tbl}
                 # print("Done calculating", self)
@@ -99,7 +120,6 @@ class compute_config:
 
 
 class compute_config_energy(compute_config):
-
     """
     """
 
@@ -107,7 +127,6 @@ class compute_config_energy(compute_config):
         """
         Compute the energy of the entries and store them in graph_db tables.
         """
-
         # print("Starting calculation", self)
         csys = self.csys
         gdb = self.GDB
[…]
