+++ besmarts-core/python/besmarts/mechanics/hessians.py
@@ -2,160 +2,34 @@
 besmarts.mechanics.hessians
 """
 
-import sys
 import copy
 import numpy as np
-import itertools
+import pprint
+
 from besmarts.core import assignments
-from besmarts.core import arrays
 from besmarts.core import graphs
 from besmarts.core import geometry
-from besmarts.mechanics import objectives
-from besmarts.mechanics import optimizers_scipy
-# from besmarts.mechanics import optimizers_openmm
-from besmarts.codecs import codec_rdkit
+from besmarts.core import configs
 from besmarts.mechanics import molecular_models as mm
-import numpy as np
-import sys
-import pprint
-from pprint import pprint
 
 from besmarts.mechanics import vibration
+
 au2kcal = 627.51/(0.529177**2)
 
-def get_xyz(fn):
-    frames = []
-    syms = []
-    frame = []
-    sym = []
-    with open(fn) as f:
-        for i, line in enumerate(f):
-            # print("loading line", i)
-            line = line.replace('\n','')
-            if i == 0:
-                natom = int(line.strip())
-                continue
-            if i % (natom+2) == 0 and frame:
-                frames.append(frame)
-                # print("Added frame", len(frames))
-                frame = []
-                syms.append(sym)
-                sym = []
-                continue
-            elif i % (natom+2) > 1:
-                # print("loading xyz", line)
-                s,x,y,z = line.split()
+PRECISION = configs.precision
 
-                xyz = [float(x), float(y), float(z)]
-                sym.append(s)
-                frame.append(xyz)
-        if frame:
-            frames.append(frame)
-            syms.append(sym)
-    return syms[0], np.array(frames[0])
-
-def load():
-    syms, frame = get_xyz(xyzfile)
-    g = gcd.smiles_decode(smiles)
-    N = len(g.nodes)
-
-def b2matrix(pos, bonds=True, angles=True, torsions=True, outofplanes=True, pairs=True, remove1_3=False):
-
-    N = len(pos.selections)
-    i = 0
-    weights = []
-    if bonds:
-        bonds = assignments.graph_assignment_jacobian2_bonds(pos)
-
-    if angles:
-        angles = assignments.graph_assignment_jacobian2_angles(pos)
-
-    if torsions:
-        torsions = assignments.graph_assignment_jacobian2_torsions(pos)
-
-    if outofplanes:
-        outofplanes = assignments.graph_assignment_jacobian2_outofplanes(pos)
-
-    if pairs:
-        pairs = assignments.graph_assignment_jacobian2_pairs(pos)
-        if angles and remove1_3:
-            for a in angles.selections:
-                if (a[0], a[2]) in pairs.selections:
-                    del pairs.selections[(a[0], a[2])]
-
-    B2 = []
-    ics = []
-
-    for valence in filter(None, (bonds, angles, torsions, outofplanes, pairs)):
-        ics.extend(valence.selections)
-        for ic, confs in valence.selections.items():
-
-            b2 = []
-            for conf in confs:
[…]
