+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -42,6 +42,8 @@ from besmarts.mechanics import molecular_models as mm
 
 eid_t = int
 
+PRECISION = configs.precision
+
 POSITIONS = assignments.POSITIONS
 HESSIANS = assignments.HESSIANS
 GRADIENTS = assignments.GRADIENTS
@@ -206,7 +208,7 @@ class compute_config_gradient(compute_config):
         """
         csys = self.csys
         gdb = self.GDB
-        tbl_idx = assignments.GRADIENTS
+        tbl_idx = GRADIENTS
         tid = assignments.POSITIONS
 
         all_results = []
@@ -256,10 +258,10 @@ class compute_config_gradient(compute_config):
                                 )
 
                     gx = optimizers_openmm.physical_system_gradient_openmm(
-                        csys,
-                        psys
+                        psys,
+                        csys
                     )
-                    gx = arrays.array_round(gx, 12)
+                    gx = arrays.array_round(gx, PRECISION)
                     tbl.values.extend(gx)
 
                 r = {tbl_idx: tbl}
@@ -282,7 +284,7 @@ class compute_config_hessian(compute_config):
         # print("Starting calculation", self)
         csys = self.csys
         gdb = self.GDB
-        tbl_idx = assignments.HESSIANS
+        tbl_idx = HESSIANS
         tid = assignments.POSITIONS
 
         all_results = []
@@ -357,8 +359,8 @@ class compute_config_hessian(compute_config):
                     tbl_grad.values.extend(gx)
 
                 r = {
-                    assignments.HESSIANS: tbl_hess,
-                    assignments.GRADIENTS: tbl_grad
+                    HESSIANS: tbl_hess,
+                    GRADIENTS: tbl_grad
                 }
 
                 results.append(r)
@@ -464,7 +466,7 @@ class compute_config_penalty(compute_config):
         for k, ref in self.targets.items():
             val = keys.get(k)
             if val is not None:
-                results[k] = round(val - ref, 12)
+                results[k] = round(val - ref, PRECISION)
         return results
 
 
@@ -542,14 +544,14 @@ class objective_config:
                                 dxb.extend(dv)
 
         dxdp = arrays.array_difference(dxb, dxa)
-        dxdp = arrays.array_round(dxdp, 12)
+        dxdp = arrays.array_round(dxdp, PRECISION)
         dxdp = arrays.array_scale(dxdp, 1/(2*h))
 
         d2xdp2 = arrays.array_add(
             arrays.array_scale(X0, -2),
             arrays.array_add(dxb, dxa)
         )
-        d2xdp2 = arrays.array_round(d2xdp2, 12)
+        d2xdp2 = arrays.array_round(d2xdp2, PRECISION)
         d2xdp2 = arrays.array_scale(d2xdp2, 1.0/(h*h))
 
         return dxdp, d2xdp2
@@ -584,7 +586,7 @@ class objective_config:
 
                             for ic, x1, x0 in zip(gdc0.selections, x1l, x0l):
                                 x = arrays.array_difference(x1, x0)
-                                x = arrays.array_round(x, 12)
+                                x = arrays.array_round(x, PRECISION)
                                 obj.extend(x)
                                 xyz1 = " ".join([f"{xi:8.4f}" for xi in x1])
                                 xyz0 = " ".join([f"{xi:8.4f}" for xi in x0])
@@ -603,7 +605,7 @@ class objective_config:
                 f"Total Position SSE: {X2:10.5f} A^2",
                 f"RMSE: {rmse:10.5f} A"
             ]))
[…]
