### 2024-07-31 a67e21b
add example 11_aicp.py
