+++ besmarts-core/python/examples/11_aicp.py
@@ -1,0 +1,1751 @@
+"""
+examples/11_aicp.py
+
+Ab-Initio (valence) chemical perception (AICP)
+
+This is a full valence fit of parameters starting from a base, empty set of
+parameters. The initial parameters are first rapidly expanded by clustering
+bonds, angles, and torsions based on Hessian projection, followed by a BESMARTS
+parameter search which will refine the parameters to fit the objective. Here
+we fit on the geometry, gradient, and frequencies of a single molecule. The
+molecule is chemically diverse and has 25 atoms, so it presents a typical
+challenge in force field design.
+
+Warning: This is not a quick calculation. With 100-200 cores this completed in
+approximately one day.
+"""
+
+import os
+import pickle
+import numpy as np
+import base64
+import tempfile
+from besmarts.mechanics import fits
+from besmarts.mechanics import smirnoff_models
+from besmarts.core import perception
+from besmarts.core import assignments
+from besmarts.cluster import cluster_objective
+from besmarts.cluster import cluster_optimization
+from besmarts.assign import hierarchy_assign_rdkit
+from besmarts.codecs import codec_rdkit
+from besmarts.core import configs
+
+configs.processors = 16
+configs.remote_compute_enable = True
+# To use this, from a terminal run
+# python -m besmarts.worker <IP> 55555 np
+# where ip is your publically visible IP and np is the number of cores
+# to assign to the worker. Recommended value for this example is 4-10 with a
+# maximum number of 30 workers (due to tier acceptance).
+configs.workqueue_port = 55555
+
+# Known issue: if the command is abnormally terminated, some processes will
+# not be killed and will not free the port. This will prevent the script from
+# running again as it will try to bind the occupied port. To get around this
+# (on Linux), run
+# pkill -f 11_besmarts_fit_aicp.py
+
+au2ang = 0.529177249
+hess2kcal = 627.51 / (0.529177249**2)
+au2kj = 627.51*4.184
+
+
+def mystrategy():
+    splitter = configs.smarts_splitter_config(
+        1, 2, 0, 0, 0, 0, True, True, 0, False, True, True, True
+    )
+    splitter.split_specific = True
+    splitter.split_general = False
+    extender = configs.smarts_extender_config(
+        0, 0, True
+    )
+    cfg = configs.smarts_perception_config(splitter, extender)
+    strategy = cluster_optimization.optimization_strategy_default(cfg)
+    strategy.overlaps = [0]
+    strategy.macro_accept_max_per_cluster = 1
+    strategy.micro_accept_max_per_cluster = 1
+
+    strategy.macro_accept_max_total = 0
+    strategy.micro_accept_max_total = 0
+
+    return strategy
+
+
+def pkl(self, fnm):
+    with open(fnm, 'wb') as f:
+        pickle.dump(self, f)
+
+
+def load_xml(xml):
+
+    gcd = codec_rdkit.graph_codec_rdkit()
+    labeler = hierarchy_assign_rdkit.smarts_hierarchy_assignment_rdkit()
+    pcp = perception.perception_model(gcd, labeler)
+
+    fd, ff_fname = tempfile.mkstemp(suffix=".offxml")
+    with open(fd, 'w') as f:
+        f.write(xml)
+    csys = smirnoff_models.smirnoff_load(ff_fname, pcp)
+    os.remove(ff_fname)
+    return csys
+
+
+def configure_tiers(objs, penalties, fit_models, fit_symbols):
+
+    # The initial fit to determine the reference
+    initial = fits.objective_tier()
+    final = fits.objective_tier()
+
+    initial.objectives = objs
+
+    # Don't perform any stochastic optimization
+    initial.anneal = 0
+
+    # line search in LBFGS
+    initial.maxls = 30
[…]
