### 2024-07-31 6ff191f
add example 11_aicp.py
