+++ besmarts-core/python/examples/11_aicp.py
@@ -20,40 +20,61 @@ import pickle
 import numpy as np
 import base64
 import tempfile
-from besmarts.mechanics import fits
-from besmarts.mechanics import smirnoff_models
+
+from besmarts.core import configs
 from besmarts.core import perception
 from besmarts.core import assignments
 from besmarts.cluster import cluster_objective
 from besmarts.cluster import cluster_optimization
-from besmarts.assign import hierarchy_assign_rdkit
+from besmarts.mechanics import fits
+from besmarts.mechanics import smirnoff_models
+
 from besmarts.codecs import codec_rdkit
-from besmarts.core import configs
+from besmarts.assign import hierarchy_assign_rdkit
 
 configs.processors = 16
-configs.remote_compute_enable = True
-# To use this, from a terminal run
+configs.remote_compute_enable = False
+# To use additional workers, from a terminal run (on the separate work machine)
 # python -m besmarts.worker <IP> 55555 np
-# where ip is your publically visible IP and np is the number of cores
-# to assign to the worker. Recommended value for this example is 4-10 with a
-# maximum number of 30 workers (due to tier acceptance).
+# where ip is your publically visible IP where this is being run and np is the
+# number of cores to assign to the worker. Recommended value for this example
+# is 4-10 with a maximum number of 30 workers (due to tier acceptance).
 configs.workqueue_port = 55555
+
+# Show a little more output during the longer distributed compute
+# to know it's doing something :)
+# Currently will show an update every 10% progress or 60 seconds, whichever
+# comes first.
+configs.compute_verbosity = 1
 
 # Known issue: if the command is abnormally terminated, some processes will
 # not be killed and will not free the port. This will prevent the script from
 # running again as it will try to bind the occupied port. To get around this
-# (on Linux), run
-# pkill -f 11_besmarts_fit_aicp.py
+# (on Linux), run the following in a terminal
+# pkill -f 11_aicp.py
 
+# Boring constants yet to find a home. Need because gradient and hessian
+# are in atomic units. Currently, we expect gradients to be in kJ/A, but
+# Hessians in kcal/A/A. This will change in the future.
 au2ang = 0.529177249
 hess2kcal = 627.51 / (0.529177249**2)
 au2kj = 627.51*4.184
 
 
 def mystrategy():
+
+    # The chemical perception strategy. Use a SMARTS search depth of up to 2
+    # bits, and do not look at neighboring atoms.
+
     splitter = configs.smarts_splitter_config(
         1, 2, 0, 0, 0, 0, True, True, 0, False, True, True, True
     )
+
+    # Disable general patterns like [!#6][!#8]. In general these will lead
+    # to a hierarchy with fewer parameters as it greatly widens the search
+    # space with almost no extra cost, but are harder to read and reason
+    # with by SMARTS non-experts. We prefer specific splits here, like
+    # [#6X4]-[#7]
     splitter.split_specific = True
     splitter.split_general = False
     extender = configs.smarts_extender_config(
@@ -62,11 +83,14 @@ def mystrategy():
     cfg = configs.smarts_perception_config(splitter, extender)
     strategy = cluster_optimization.optimization_strategy_default(cfg)
     strategy.overlaps = [0]
-    strategy.macro_accept_max_per_cluster = 1
-    strategy.micro_accept_max_per_cluster = 1
 
+    # Accept an unlimited number of parameters per iteration
     strategy.macro_accept_max_total = 0
     strategy.micro_accept_max_total = 0
+
+    # But only accept at most one split per iteration
[…]
