+++ besmarts-core/python/examples/07_besmarts_fit.py
@@ -8,7 +8,7 @@ force field was fit on positions and gradients of a single molecule.
 
 import tempfile
 import os
-from typing import Dict, List
+from typing import Dict, List, Optional
 from besmarts.mechanics import fits
 from besmarts.mechanics import smirnoff_models
 from besmarts.mechanics import molecular_models as mm
@@ -58,7 +58,24 @@ xyz_grad = """11
 """
 
 
-def load_xyz(flist, indices=None) -> assignments.graph_db_row:
+def load_xyz(flist: list[str], indices=Optional[list[int]]) -> assignments.graph_db_row:
+    """
+    Convert a list of xyz strings into a graph_db_row.
+    A graph_db_row is a collection of graph_db_columns, each of which
+    contains a selection of atoms and their coordinates.
+
+    Parameters
+    ----------
+    flist : list[str]
+        A list of xyz strings
+    indices: list[int]
+        A list of indices to use for the atoms. If None, the indices are
+        taken from the xyz files.
+
+    Returns
+    -------
+    assignments.graph_db_row
+    """
     s = 0
     N = None
     gdr = assignments.graph_db_row()
@@ -83,7 +100,17 @@ def load_xyz(flist, indices=None) -> assignments.graph_db_row:
                 s += 1
     return gdr
 
-def make():
+def make() -> dict[str, list[dict[int, str]]]:
+    """
+    Make a dataset for the optimization.
+    This returns a dictionary with a single molecule.
+    The key is its SMILES string. The value is a list with one dictionary.
+    The dictionary keys are essentially
+    integer enum values for property types,
+    where the integer value represents the index of the property in a
+    :class:`besmarts.core.assignments.graph_db_entry`.
+    """
+
     global smi
     global xyz_positions
     global xyz_grad
@@ -100,7 +127,11 @@ def make():
     return d
 
 
-def new_gdb(f: Dict[str, List[str]]) -> assignments.graph_db:
+def new_gdb(f: Dict[str, List[Dict[int, str]]]) -> assignments.graph_db:
+    """
+    This creates a ``graph_db`` from a dataset
+    as generated by ``make()``.
+    """
     gcd = codec_rdkit.graph_codec_rdkit()
     gdb = assignments.graph_db()
 
