### 2024-08-08 47dad24
remove redundant enumerate
