+++ besmarts-core/python/besmarts/core/clusters.py
@@ -91,7 +91,7 @@ def get_objective(cst, assn, objfn, edits, splitting=True):
         m = hidx.index.above.get(n.index, None)
         if m is not None:
             m = hidx.index.nodes[m]
-            if m.type != "parameter":
+            if m.type != "parameter" or n.type != "parameter":
                 continue
             n_match = new_match[n.name]
             m_match = new_match[m.name]
@@ -113,24 +113,9 @@ def get_objective(cst, assn, objfn, edits, splitting=True):
     return keep, X
 
 
-class find_successful_candidates_ctx:
-    candidates = None
-    pq = None
-    group_number = None
-    step = None
-    hidx = None
-    topology = None
-    labeler = None
-    Sj_sma = None
-    groups = None
-    assn = None
-    strategy = None
-    gcd = None
-    objective = None
-
-
 
 def find_successful_candidates_distributed(S, Sj, operation, edits, shm=None):
+
     sag = shm.sag
     cst = shm.cst
     hidx = cst.hierarchy.copy()
@@ -159,7 +144,7 @@ def find_successful_candidates_distributed(S, Sj, operation, edits, shm=None):
         # print(datetime.datetime.now(), '*** 2')
         hent = hidx.index.node_add(
             S.index,
-            trees.tree_node(None, "parameter", "", param_name),
+            trees.tree_node(None, S.category, S.type, param_name),
             index=0,
         )
         # print(datetime.datetime.now(), '*** 3')
@@ -203,10 +188,11 @@ def find_successful_candidates_distributed(S, Sj, operation, edits, shm=None):
 
             if not (cst.mappings[S.name] and cst.mappings[hent.name]):
                 keep = False
+            # print(datetime.datetime.now(), '*** 8', keep, X, obj)
 
             # keep the splits that match the most (general)
             # match_len = len(cst.mappings[S.name])
-            
+
             # keep the splits that match the least (specific)
             # this is better since it leaves more in S, so more can be split
             # at a time
@@ -268,7 +254,7 @@ def perform_operations(
             # print(datetime.datetime.now(), '*** 2')
             hent = hidx.index.node_add(
                 S.index,
-                trees.tree_node(None, tuple((0,0,0)), "parameter", param_name),
+                trees.tree_node(None, S.category, S.type, param_name),
                 index=0,
             )
 
@@ -302,111 +288,6 @@ def perform_operations(
             hidx.smarts.pop(hent.index)
 
     return hidx, nodes
-
-
-def find_successful_candidates(cnd_i, key, return_hier=False):
-    candidates = find_successful_candidates_ctx.candidates
-    # pq = find_successful_candidates_ctx.pq
-    # x, key = pq[cnd_i]
-    group_number = find_successful_candidates_ctx.group_number
-    # step = find_successful_candidates_ctx.step
-    Sj_sma = find_successful_candidates_ctx.Sj_sma
-    labeler = find_successful_candidates_ctx.labeler
-    groups = find_successful_candidates_ctx.groups
-    assn = find_successful_candidates_ctx.assn
-    hidx = find_successful_candidates_ctx.hidx.copy()
-    strategy = find_successful_candidates_ctx.strategy
-    gcd = find_successful_candidates_ctx.gcd
-    smiles = find_successful_candidates_ctx.smiles
-    objective = find_successful_candidates_ctx.objective
-    # topo = find_successful_candidates_ctx.topology
-
-    topo = hidx.topology
-
-    (S, Sj, step, _, _, _, _) = candidates[key]
-    (edits, _, p_j) = key
-    param_name = "p."
-    sma = ""
-    added = False
-    dX = 0.0
-
-    if step.operation == strategy.SPLIT:
[…]
