+++ besmarts-core/python/besmarts/core/mapper.py
@@ -12,18 +12,17 @@ import multiprocessing
 
 from typing import Dict, Sequence, List, Tuple
 
-from besmarts.codecs import codec_native
 from besmarts.core import configs, chem, graphs, db, codecs
 from besmarts.core.graphs import structure
-from besmarts.core.primitives import primitive_key
-from besmarts.core import compute
 from besmarts.core import arrays
 from besmarts.core.logs import dprint, timestamp
+from besmarts.core import compute
 
 # TODO: remove this shim
 from besmarts.core.graphs import structure_extend as mapper_smarts_extend
 
 mapping = Dict[int, int]
+
 
 class mapped_type:
     """
@@ -1853,6 +1852,7 @@ def union_list(
 
     return graphs.structure_copy(ref)
 
+
 def intersection_list_dispatch(
     indices,
 ) -> graphs.structure:
@@ -1864,6 +1864,35 @@ def intersection_list_dispatch(
     return intersection_list(
         A, config, max_depth, reference, sort=True, executor=None, verbose=False
     )
+
+
+def union_list_dispatch_distributed(
+    indices: List[int],
+    shm=None
+) -> graphs.structure:
+
+    topo = shm.topology
+
+    reference = shm.reference
+    config = shm.config
+    max_depth = shm.max_depth
+
+    icd: codecs.intvec_codec = shm.icd
+    G, sel = shm.A
+
+    sel = [sel[i] for i in indices]
+
+    work = union_list_parallel(
+        G,
+        sel,
+        topo,
+        config=config,
+        max_depth=max_depth,
+        reference=reference,
+        icd=icd,
+        verbose=False
+    )
+    return icd.structure_encode(work)
 
 
 def union_list_dispatch(
@@ -1947,10 +1976,12 @@ def intersection_list_parallel(
 
     return work[0]
 
-def union_list_parallel(
+
+def union_list_distributed(
     G: Sequence[graphs.graph],
     selections,
     topo,
+    wq,
     config: configs.mapper_config = None,
     max_depth=None,
     reference=None,
@@ -1995,9 +2026,137 @@ def union_list_parallel(
     procs = min(os.cpu_count(), len(indices))
     procs = min(procs, configs.processors)
 
+    work = None
+
+
+    while len(indices) > 1:
+        # print(timestamp(), f"Initializing pool")
+
+        if len(indices) > 1000:
+            print(timestamp(), f"Distributed Union merging={len(indices)}")
+            shm = compute.shm_local(0, data={
+                "reference": reference,
+                "result": union_ctx.result,
+                "topology": topo,
+                "config": config,
+                "max_depth": max_depth,
+                "A": (G, selections),
+                "icd": icd
+            })
+            ws = compute.workqueue_new_workspace(
+                wq,
+                shm=shm
+            )
+            
+            chunk_n = max(2, len(indices) // procs)
+            chunk_n = min(chunk_n, 10000)
+
+            iterable = {
+                i: ((x,), {}) for i, x in enumerate(arrays.batched(indices, chunk_n))
+            }
+
+            chunksize = 1
+
+            work = compute.workspace_submit_and_flush(
+                ws,
+                union_list_dispatch_distributed,
+                iterable,
+                chunksize,
+                1,
+                len(iterable),
+                verbose=True
+            )
+            ws.close()
+            work = list(work.values())
+        else:
+            print(timestamp(), f"Parallel Union merging={len(indices)}")
+
+            with multiprocessing.pool.Pool(processes=procs) as pool:
+                # print(timestamp(), f"Generating batches")
+                chunk_n = max(2, len(indices) // procs)
[…]
