### 2024-08-08 e40a9e5
distributed union; enable_minimization flag in objectives
