+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -40,6 +40,7 @@ from besmarts.mechanics import hessians
 from besmarts.mechanics import vibration
 from besmarts.mechanics import molecular_models as mm
 
+import scipy.optimize
 eid_t = int
 
 PRECISION = configs.precision
@@ -60,6 +61,7 @@ class compute_config:
     def __init__(self, addr):
         self.addr = addr
         self.keys = []
+        self.enable_minimization = True
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_table]]:
 
@@ -294,6 +296,7 @@ class compute_config_hessian(compute_config):
             for eid, gde in gdb.entries.items():
                 tbl_hess = assignments.graph_db_table(topology.null)
                 tbl_grad = assignments.graph_db_table(topology.null)
+                tbl_pos = assignments.graph_db_table(topology.atom)
                 rids = assignments.graph_db_table_get_row_ids(gde[tid])
                 for rid in rids:
                     system = []
@@ -336,9 +339,11 @@ class compute_config_hessian(compute_config):
 
                     pos = psys.models[0].positions[0]
 
-                    args, keys = objectives.array_flatten_assignment(
-                        pos.selections
-                    )
+                    if self.enable_minimization:
+                        minpos = optimizers_openmm.optimize_positions_openmm(csys, psys, tol=1e-7)
+                        pos = minpos
+                        tbl_pos.values = minpos
+                        psys.models[0].positions[0] = minpos
 
                     hess_mm = optimizers_openmm.physical_system_hessian_openmm(
                         psys,
@@ -359,6 +364,7 @@ class compute_config_hessian(compute_config):
                     tbl_grad.values.extend(gx)
 
                 r = {
+                    POSITIONS: tbl_pos,
                     HESSIANS: tbl_hess,
                     GRADIENTS: tbl_grad
                 }
@@ -487,6 +493,7 @@ class objective_config:
         self.include = include
         self.cache = {}
         self.batch_size = None
+        self.enable_minimization = True
 
     def get_task(
         self,
@@ -610,6 +617,10 @@ class objective_config:
 
 class objective_config_gradient(objective_config):
 
+    def __init__(self, *args, **kwds):
+        super().__init__(*args, **kwds)
+        self.enable_minimization = False
+
     def get_task(
         self,
         GDB: assignments.graph_db,
@@ -718,7 +729,7 @@ class objective_config_gradient(objective_config):
             else:
                 rmse = 0
 
-            X2 = arrays.array_inner_product(X,X)
+            X2 = arrays.array_inner_product(X, X)
             out.append(" ".join([
                 f" Total Gradient SSE: {X2:10.5f} (kJ/mol/A)^2",
                 f"RMSE: {rmse:10.5f} kJ/mol/A"
@@ -727,6 +738,10 @@ class objective_config_gradient(objective_config):
 
 
 class objective_config_hessian(objective_config):
+
+    def __init__(self, *args, **kwds):
+        super().__init__(*args, **kwds)
+        self.enable_minimization = False
 
     def get_task(
         self,
@@ -746,6 +761,8 @@ class objective_config_hessian(objective_config):
         else:
             cc.keys.append({})
         cc.reuse = reuse
+        cc.enable_minimization = self.enable_minimization
+
         return cc
 
     def compute_gradient_2pt(self, ref, X0, E0, E1, h):
@@ -769,9 +786,36 @@ class objective_config_hessian(objective_config):
             grad0 = list(dtbls[GRADIENTS].values)
             x0 = hessians.hessian_frequencies(g, hess0, grad0, DL, ics, B, B2)
 
[…]
