### 2024-08-08 8df816a
check for linear torsions
