### 2024-08-09 1f9ebe4
add types and comment gdb creation
