### 2024-08-12 9c11575
add some more comments and signatures
