+++ besmarts-core/python/examples/07_besmarts_fit.py
@@ -58,7 +58,7 @@ xyz_grad = """11
 """
 
 
-def load_xyz(flist: list[str], indices=Optional[list[int]]) -> assignments.graph_db_row:
+def load_xyz(flist: list[str], indices: Optional[list[int]] = None) -> assignments.graph_db_row:
     """
     Convert a list of xyz strings into a graph_db_row.
     A graph_db_row is a collection of graph_db_columns, each of which
@@ -204,21 +204,53 @@ def run():
     """
 
     # 1 Build the dataset and FF
-    d = make()
-    csys = load_sage_csys()
-    gdb = new_gdb(d)
-    psys = fits.gdb_to_physical_systems(gdb, csys)
+    example_training_data: dict[str, list[dict]] = make()
+
+    # load the Sage SMIRNOFF force field into a
+    # ``molecular_mechanics.chemical_system``.
+    #  The ``chemical_system`` has the following models in order:
+    # bonds, angles, torsions, outofplanes, electrostatics, vdw.
+    # Each model is an instance of ``molecular_mechanics.chemical_model``.
+    csys: mm.chemical_system = load_sage_csys()
+    gdb: assignments.graph_db = new_gdb(example_training_data)
+
+    # set up the physical systems
+    # a molecular_mechanics.physical_system is a collection of
+    # ``molecular_mechanics.physical_model`` objects.
+    # A physical model is the functional form that can be evaluated
+    # as a function of positions
+    psys_dict: dict[int, mm.physical_system] = fits.gdb_to_physical_systems(
+        gdb, csys
+    )
 
     # 1 Configure the fitting strategy
+    # models should be a dictionary where the keys are the indices
+    # of each ``molecular_mechanics.chemical_model`` object
+    # in the ``csys``. Therefore,
+    # the 0 index here refers to the bond model.
+    # if we wanted to train an angle, we would need
+    # to set the key to 1.
     models = {0: ["b4"]}
-    strat = fits.forcefield_optimization_strategy_default(csys, models=models)
-    co = fits.chemical_objective
+
+    # this sets up the default optimization strategy for the fit.
+    # It uses the ``models`` dictionary to determine which models
+    # to fit in the ``csys.models`` list.
+    # default settings are also applied to splitting, extending, perception
+    # configs etc.
+    strategy = fits.forcefield_optimization_strategy_default(
+        csys, models=models
+    )
+    chemical_objective_function = fits.chemical_objective
 
     # 2. Configure the objective tiers
     final = fits.objective_tier()
+    # set the objectives for the final tier
+    # each value is an ``objective_config``
     final.objectives = {
         0: fits.objective_config_position(
                 assignments.graph_db_address(
+                    # eid stands for entry ID
+                    # here it's set to 0 for the first molecule
                     eid=[0],
                 ),
                 scale=1
@@ -231,48 +263,52 @@ def run():
         ),
     }
 
-    fit_models = [0]
+    fit_models = [0]  # 0 here indicates bonds
     final.fit_models = fit_models
-    final.fit_symbols = ["l"]
+    final.fit_symbols = ["l"]  # l indicates equilibrium length
 
     onestep = fits.objective_tier()
+    # set same objectives as final tier
     onestep.objectives = final.objectives
+    # only do two fitting steps
     onestep.step_limit = 2
+    # the top 3 candidates will be passed to the final tier
     onestep.accept = 3
 
+    # fit the same models, symbols as the final tier
     onestep.fit_models = fit_models
     onestep.fit_symbols = ["l"]
[…]
