+++ besmarts-rdkit/python/setup.py
@@ -14,7 +14,12 @@ setuptools.setup(
     url='https://github.com/trevorgokey/besmarts',
     long_description=long_description,
     long_description_content_type="text/markdown",
-    packages=['besmarts'],
+    packages=[
+        'besmarts',
+        'besmarts.assign',
+        'besmarts.codecs',
+        'besmarts.perception',
+    ],
     install_requires=requirements,
     classifiers=[
         'Programming Language :: Python :: 3.6',
