+++ besmarts-core/python/setup.py
@@ -14,7 +14,14 @@ setuptools.setup(
     url='https://github.com/trevorgokey/besmarts',
     long_description=long_description,
     long_description_content_type="text/markdown",
-    packages=['besmarts'],
+    packages=[
+        'besmarts',
+        'besmarts.assign',
+        'besmarts.cluster',
+        'besmarts.codecs',
+        'besmarts.core',
+        'besmarts.mechanics'
+    ],
     install_requires=requirements,
     classifiers=[
         'Programming Language :: Python :: 3.6',
