### 2024-08-12 77aec2b
clean; memory opts
