+++ besmarts-core/python/besmarts/core/compute.py
@@ -435,139 +435,30 @@ class Server(managers.Server):
         idx = 0
         while True:
             # threads = [x for x in threads if x.is_alive()]
-            if True:
-                try:
-                    # print("LISTENING")
-                    # t0 = time.perf_counter()
-                    c = self.listener.accept()
-                    # t = time.perf_counter()
-                    # print(f"CONNECTION RECV time {t-t0:.6f}")
-                except OSError as e:
-                    # except Exception as e:
-                    # print(f"acceptor exception {type(e)} {e}")
-                    # print(
-                    #     f"Warning, there are {len(threads)} active connections, "
-                    #     f"waiting 10 seconds before accepting new connections."
-                    # )
-                    # time.sleep(10.0)
-                    continue
-                # if len(threads) > 600:
-                #     print(f"there are {len(threads)} active connections, waiting 2")
-                #     time.sleep(2)
-                # if len(threads) > 700:
-                #     print(f"there are {len(threads)} active connections, waiting 3")
-                #     time.sleep(3)
-                # if len(threads) > 800:
-                #     print(f"there are {len(threads)} active connections, waiting 4")
-                #     time.sleep(4)
-                # if len(threads) > 900:
-                #     print(f"there are {len(threads)} active connections, waiting 5")
-                #     time.sleep(5)
-                # self.handle_request(c)
-                t = threading.Thread(target=self.handle_request, args=(c,))
-                # pool.apply_async(self.handle_request, (c,))
-                t.daemon = True
-                try:
-                    t.start()
-                except RuntimeError:
-                    # possible on large jobs with many threads?
-                    continue
+            try:
+                # print("LISTENING")
+                # t0 = time.perf_counter()
+                c = self.listener.accept()
+                # t = time.perf_counter()
+                # print(f"CONNECTION RECV time {t-t0:.6f}")
+            except OSError as e:
+                # except Exception as e:
+                # print(f"acceptor exception {type(e)} {e}")
+                # print(
+                #     f"Warning, there are {len(threads)} active connections, "
+                #     f"waiting 10 seconds before accepting new connections."
+                # )
+                # time.sleep(10.0)
+                continue
 
-            else:
-                if len(recvs) < 32:
-                    new_out = []
-                    new_err = []
-                    new_conn = []
-                    new_t = threading.Thread(
-                        target=self.recv_request,
-                        args=(new_conn, new_out, new_err),
-                        name=f"recv_{idx}",
-                    )
-                    new_t.daemon = True
-                    new_t.start()
-                    recvs[idx] = (new_conn, new_t, new_out, new_err)
-                    idx += 1
+            t = threading.Thread(target=self.handle_request, args=(c,))
+            t.daemon = True
 
-                for i, (c, tr, out, err) in list(recvs.items()):
-                    if tr.is_alive():
-                        continue
-
-                    if not c:
-                        tr.join()
-                        del recvs[i]
-                        continue
-
-                    if err and not out:
-                        t = threading.Thread(
-                            target=self.send_request,
-                            args=(c[0], err[0]),
-                            name=f"senderr_{i}",
-                        )
-                        t.daemon = True
-                        t.start()
-                        sends[i] = (c, t, None, None)
-                    elif out:
-                        msg = []
[…]
