### 2024-08-12 c9d9903
clean; allow singlepoint with 0 steps
