+++ besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py
@@ -7,6 +7,8 @@ import copy
 import scipy.optimize
 import numpy as np
 import datetime
+import time
+# import tracemalloc
 
 from besmarts.core import geometry
 from besmarts.core import assignments
@@ -88,49 +90,49 @@ def ff_obj(x, keys, csys, refpsys, ref, refcsys):
     # print("Total obj:", obj)
     return obj
 
-def objective_gdb(csys, gdb, obj, psysref=None, reuse=None, ws=None):
+# def objective_gdb(csys, gdb, obj, psysref=None, reuse=None, ws=None):
 
-    tasks = {}
-    for i, x in obj.items():
-        psys = None
-        if psysref:
-            psys = psysref[x.addr.eid[0]]
-        tasks[i] = x.get_task(gdb, csys, psys=psys, reuse=reuse)
+#     tasks = {}
+#     for i, x in obj.items():
+#         psys = None
+#         if psysref:
+#             psys = psysref[x.addr.eid[0]]
+#         tasks[i] = x.get_task(gdb, csys, psys=psys, reuse=reuse)
 
-    # print("Starting tasks")
-    if ws:
-        # wq = compute.workqueue_local('127.0.0.1', 55555)
-        # ws = compute.workqueue_new_workspace(wq, shm={})
+#     # print("Starting tasks")
+#     if ws:
+#         # wq = compute.workqueue_local('127.0.0.1', 55555)
+#         # ws = compute.workqueue_new_workspace(wq, shm={})
 
-        results = compute.workspace_submit_and_flush(
-            ws,
-            fits.objective_run_distributed,
-            {i: ([x], {}) for i, x in tasks.items()}
-        )
-        # print("############################################")
+#         results = compute.workspace_submit_and_flush(
+#             ws,
+#             fits.objective_run_distributed,
+#             {i: ([x], {}) for i, x in tasks.items()}
+#         )
+#         # print("############################################")
         
-        # results = {i:v for i, v in enumerate(
-        #     compute.workspace_submit_and_flush(
-        #         ws,
-        #         fits.objective_run,
-        #         tasks.values()
-        #     )
-        # )}
-        # ws.close()
+#         # results = {i:v for i, v in enumerate(
+#         #     compute.workspace_submit_and_flush(
+#         #         ws,
+#         #         fits.objective_run,
+#         #         tasks.values()
+#         #     )
+#         # )}
+#         # ws.close()
 
-    else:
-        results = {}
-        for i, task in tasks.items():
-            # print(f"RUNNING {i}")
-            results[i] = task.run()
-        # results = {i: task.run() for i, task in tasks.items()}
-    # print("Tasks done. Calculating objective")
-    X = 0
-    for i, x in obj.items():
-        # print(f"Objective {i}", x)
-        xi = x.compute_total(assignments.graph_db_get_entries(gdb, x.addr.eid), results[i])
-        X += xi*x.scale
-    return X
+#     else:
+#         results = {}
+#         for i, task in tasks.items():
+#             # print(f"RUNNING {i}")
+#             results[i] = task.run()
+#         # results = {i: task.run() for i, task in tasks.items()}
+#     # print("Tasks done. Calculating objective")
+#     X = 0
+#     for i, x in obj.items():
+#         # print(f"Objective {i}", x)
+#         xi = x.compute_total(assignments.graph_db_get_entries(gdb, x.addr.eid), results[i])
+#         X += xi*x.scale
+#     return X
 
 def objective_hessp(x, p, *args):
     history = args[6]
@@ -151,225 +153,33 @@ def objective_hess(x, *args):
 
     return hess
 
[…]
