+++ besmarts-openmm/python/setup.py
@@ -14,7 +14,10 @@ setuptools.setup(
     url='https://github.com/trevorgokey/besmarts',
     long_description=long_description,
     long_description_content_type="text/markdown",
-    packages=['besmarts'],
+    packages=[
+        'besmarts',
+        'besmarts.mechanics'
+    ],
     install_requires=requirements,
     classifiers=[
         'Programming Language :: Python :: 3.6',
