+++ besmarts-core/python/examples/09_minimize.py
@@ -6,6 +6,7 @@
 import sys
 import os
 import pickle
+import tempfile
 from typing import Dict, List, Tuple
 from besmarts.mechanics import smirnoff_models
 from besmarts.mechanics import optimizers_scipy
@@ -28,21 +29,21 @@ def graph_db_insert_sdf_files(gdb, gcd, fname_pos, fname_grad=None, fname_hessia
     pos, extras = gcd.sdf_decode(fname_pos)
     smi = pos.smiles
 
-    ene = extras.pop('qm_energy (kcal/mol)')
+    ene = extras.get('qm_energy (kcal/mol)')
     if ene is not None and energy is None:
         energy = ene * 4.184
 
     if fname_grad:
         with open(fname_grad) as f:
             xyzdata = f.read()
-        sym, xyz = parse_xyz(xyzdata)
-        gx = xyz_to_graph_assignment(gcd, smi, xyz)
+        sym, xyz = assignments.parse_xyz(xyzdata)
+        gx = assignments.xyz_to_graph_assignment(gcd, smi, xyz)
 
     if fname_hessian:
         with open(fname_grad) as f:
             xyzdata = f.read()
-        sym, xyz = parse_xyz(xyzdata)
-        gx = xyz_to_graph_assignment(gcd, smi, xyz)
+        sym, xyz = assignments.parse_xyz(xyzdata)
+        gx = assignments.xyz_to_graph_assignment(gcd, smi, xyz)
 
     eid, gid = assignments.graph_db_add_single_molecule_state(gdb, pos, gradients=gx, hessian=hx, energy=energy)
     return eid, gid
@@ -70,20 +71,467 @@ def run(ff_fname, sdf_fname):
     
     eid, gid = graph_db_insert_sdf_files(gdb, gcd, sdf_fname)
 
-    if os.path.exists("psys.p"):
-        print("Loading cached psys.p")
-        with open("psys.p", 'rb') as f:
-            psys = pickle.load(f)
-
     psys = fits.gdb_to_physical_systems(gdb, csys)[eid]
-    # psys = psys[6878]
+
     print("\n".join(print_xyz(psys.models[0].positions[0])))
 
     pos = optimizers_scipy.optimize_positions_scipy(csys, psys)
 
     print("\n".join(print_xyz(pos)))
 
+
+def load():
+    
+    fd, ff = tempfile.mkstemp(suffix=".offxml")
+    with os.fdopen(fd, 'w') as f:
+        f.write(xml)
+
+    fd, sdf = tempfile.mkstemp(suffix=".sdf")
+
+    with os.fdopen(fd, 'w') as f:
+        f.write(mol)
+
+    return ff, sdf
+
+
+mol = """
+  -OEChem-04202311503D
+
+ 23 24  0     0  0  0  0  0  0999 V2000
+    3.7552   -0.3135    5.4325 C   0  0  0  0  0  0  0  0  0  0  0  0
+    2.4571   -0.7173    5.5284 C   0  0  0  0  0  0  0  0  0  0  0  0
+    0.7045    0.4205    2.2696 C   0  0  0  0  0  0  0  0  0  0  0  0
+    0.2717   -0.2893    1.1874 C   0  0  0  0  0  0  0  0  0  0  0  0
+    2.6306    1.5065    5.0890 C   0  0  0  0  0  0  0  0  0  0  0  0
[…]
