+++ besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py
@@ -7,8 +7,6 @@ import copy
 import scipy.optimize
 import numpy as np
 import datetime
-import time
-# import tracemalloc
 
 from besmarts.core import geometry
 from besmarts.core import assignments
@@ -21,7 +19,12 @@ from besmarts.mechanics import objectives
 from besmarts.mechanics import fits
 
 
-def optimize_positions_scipy(csys, psys: mm.physical_system, step_limit=1000, tol=1e-10):
+def optimize_positions_scipy(
+    csys,
+    psys: mm.physical_system,
+    step_limit=1000,
+    tol=1e-10
+):
 
     pos = copy.deepcopy(psys.models[0].positions[0])
     args, keys = objectives.array_flatten_assignment(pos.selections)
@@ -29,23 +32,20 @@ def optimize_positions_scipy(csys, psys: mm.physical_system, step_limit=1000, to
     # jac = objectives.array_geom_gradient
     # jac = None
     jac = True
+
     hess = objectives.array_geom_hessian
 
-    opts = {'disp': False, 'maxls': 10, 'gtol': tol, 'ftol': 1e-9, 'maxiter': step_limit}
-    # opts = {'iprint': 11000}
-    # opts = {'disp': True}
     method = 'L-BFGS-B'
-    opts = {'disp': False, 'ftol': tol, 'gtol': tol, 'maxiter': step_limit, 'maxls': 1000, 'maxcor': len(args)**2}
+    opts = {
+        'disp': False,
+        'ftol': tol,
+        'gtol': tol,
+        'maxiter': step_limit,
+        'maxls': 1000,
+        'maxcor': len(args)**2
+    }
     hess = None
-    # method = 'Newton-CG'
-    # opts = {'disp': False, 'maxiter': step_limit}
-    # method = 'Nelder-Mead'
-    # print("SciPy Minimizer Starting")
-    # method = 'CG'
-    # method = 'BFGS'
-    # opts = {'disp': False, 'gtol': tol, 'maxiter': step_limit}
-    # method = 'trust-exact'
-    # opts = {'disp': False, 'gtol': tol, 'maxiter': step_limit, 'max_trust_radius': 1e-1, 'initial_trust_radius': 1e-3 }
+
     result = scipy.optimize.minimize(
         objectives.array_geom_energy_gradient,
         args,
@@ -62,9 +62,9 @@ def optimize_positions_scipy(csys, psys: mm.physical_system, step_limit=1000, to
 
     return pos
 
+
 def ff_obj(x, keys, csys, refpsys, ref, refcsys):
-    # print("input array:")
-    # print(x)
+
     pos = copy.deepcopy(ref)
     recalc = set([x[0] for x in keys])
 
@@ -72,7 +72,12 @@ def ff_obj(x, keys, csys, refpsys, ref, refcsys):
         mm.chemical_system_set_value(csys, k, v)
 
     reuse = set(range(len(csys.models))).intersection(recalc)
-    psys = mm.chemical_system_to_physical_system(csys, pos, ref=refpsys, reuse=reuse)
+    psys = mm.chemical_system_to_physical_system(
+        csys,
+        pos,
+        ref=refpsys,
+        reuse=reuse
+    )
     optpos = optimize_positions_scipy(csys, psys)
 
     obj = 0
@@ -82,62 +87,13 @@ def ff_obj(x, keys, csys, refpsys, ref, refcsys):
             dx = geometry.array_distance(a, b)
             dx2 = dx*dx
             obj += dx2
-            # print(obj, dx2)
-    # for k in keys:
-    #     newv = mm.chemical_system_get_value(csys, k)
-    #     refv = mm.chemical_system_get_value(refcsys, k)
[…]
