+++ besmarts-core/python/examples/07_besmarts_fit.py
@@ -21,7 +21,7 @@ from besmarts.assign import hierarchy_assign_rdkit
 from besmarts.codecs import codec_rdkit
 from besmarts.core import configs
 
-configs.processors = 4
+configs.processors = 1
 configs.remote_compute_enable = False
 configs.workqueue_port = 54321
 
@@ -58,96 +58,22 @@ xyz_grad = """11
 """
 
 
-def load_xyz(flist, indices=None) -> assignments.graph_db_row:
-    s = 0
-    N = None
-    gdr = assignments.graph_db_row()
-    for f in flist:
-        lines = f.split('\n')
-        if N is None:
-            N = int(lines[0].split()[0])
-        if indices is None:
-            indices = [*range(1,N+1)]
-        assert N == int(lines[0].split()[0])
-        for chunk in arrays.batched(lines, N+2):
-            if chunk and chunk[0]:
-                sel = {}
-                for i, line in enumerate(chunk, -2):
-                    if i >= 0:
-                        sym, x, y, z = line.split()[:4]
-                        sel[indices[i],] = list(map(float, (x, y, z)))
+def new_gdb() -> assignments.graph_db:
 
-                gdc = assignments.graph_db_column()
-                gdc.selections.update(sel)
-                gdr.columns[s] = gdc
-                s += 1
-    return gdr
-
-def make():
-    global smi
-    global xyz_positions
-    global xyz_grad
-    s = xyz_positions
-    g = xyz_grad
-    d  = {
-        smi: [
-            {
-                assignments.POSITIONS: s,
-                assignments.GRADIENTS: g,
-            },
-        ],
-    }
-    return d
-
-
-def new_gdb(f: Dict[str, List[str]]) -> assignments.graph_db:
     gcd = codec_rdkit.graph_codec_rdkit()
     gdb = assignments.graph_db()
 
-    for smi, fn_dict in f.items():
+    xyzdata = assignments.parse_xyz(xyz_positions)
+    pos = assignments.xyz_to_graph_assignment(gcd, smi, xyz_positions)
+    xyzdata = assignments.parse_xyz(xyz_grad)
+    gx = assignments.xyz_to_graph_assignment(gcd, smi, xyz_grad)
 
-        g = gcd.smiles_decode(smi)
-        gid = assignments.graph_db_add_graph(gdb, smi, g)
+    eid, gid = assignments.graph_db_add_single_molecule_state(
+        gdb,
+        pos,
+        gradients=gx
+    )
 
-        gdb.graphs[gid] = g
-        gdb.smiles[gid] = smi
-        gdb.selections[topology.index_of(topology.atom)] = {
-            gid: {k: v for k, v in enumerate(graphs.graph_atoms(g))}
-        }
-        gde = assignments.graph_db_entry()
-        gdb.entries[len(gdb.entries)] = gde
-        for rid, rdata in enumerate(fn_dict):
-            tid = assignments.POSITIONS
-            gdt = assignments.graph_db_table(topology.atom)
-            gdg = assignments.graph_db_graph()
-            gdt.graphs[gid] = gdg
-            fn = rdata[tid]
-            # indices=dict(sorted([(j, x) for j, x in enumerate(g.nodes, 1)], key=lambda x: x[1]))
-            indices = None
-            gdr = load_xyz([fn], indices=indices)
-            gdg.rows[0] = gdr
-            gde.tables[tid] = gdt
-            tid = assignments.GRADIENTS
-            if tid in rdata:
-                gdt = assignments.graph_db_table(topology.atom)
-                gdg = assignments.graph_db_graph()
-                gdt.graphs[gid] = gdg
-                fn = rdata[tid]
-                # indices=dict(sorted([(j, x) for j, x in enumerate(g.nodes)], key=lambda x: x[1]))
[…]
