### 2024-08-14 #6653 @trevorgokey
I've updated and added more examples in main/besmarts-core/python/examples, including comments that guide the user through the code.
