### 2024-08-21 #6653 @lucydot
A quick note that I am on leave for the following week and will be back at desk on 2nd Sept.
