+++ besmarts-core/python/besmarts/core/compute.py
@@ -56,7 +56,6 @@ distributed_function = Tuple[Callable, Sequence, Mapping]
 
 TIMEOUT = 3
 
-
 SHM_GLOBAL = {}
 
 ## from https://stackoverflow.com/questions/34361035/python-thread-name-doesnt-show-up-on-ps-or-htop
@@ -254,8 +253,8 @@ class Listener(connection.Listener):
             raise OSError("listener is closed")
         c = self._listener.accept()
         if self._authkey:
-            deliver_challenge(c, self._authkey)
-            answer_challenge(c, self._authkey)
+            connection.deliver_challenge(c, self._authkey)
+            connection.answer_challenge(c, self._authkey)
         # t = time.perf_counter()
         # dprint(f"SERVER ACCEPT TIME: {t - t0:.6f}")
         return c
@@ -265,8 +264,7 @@ class Server(managers.Server):
     def __init__(self, *args, **kwargs):
         dprint("HELLO FROM SERVER")
         super().__init__(*args, **kwargs)
-        # self.shm_pool = multiprocessing.Pool(32, server_shm_init, )
-        self.shm_pool = None
+
         self.shm_msg_cache = None
         self.shm_msg_cache_lock = threading.Lock()
 
@@ -344,7 +342,7 @@ class Server(managers.Server):
 
             except AttributeError:
                 if methodname is None:
-                    msg = ("#TRACEBACK", format_exc())
+                    msg = ("#TRACEBACK", traceback.format_exc())
                 else:
                     try:
                         fallback_func = self.fallback_mapping[methodname]
@@ -579,11 +577,6 @@ class Server(managers.Server):
             f"REQUEST TOTAL {t2-t0:.6f} RECV {tf-t0:.6f} PROC {t1 - tf:.6f} SEND {t2-t1:.6f}"
         )
 
-    def shutdown(self, c):
-        if self.shm_pool:
-            self.shm_pool.close()
-        super().shutdown(c)
-
     def incref(self, c, ident):
         with self.mutex:
             try:
@@ -721,61 +714,166 @@ managers.AutoProxy = AutoProxy
 # managers.BaseManager._Server = Server
 
 
-def _repopulate_pool(self):
-    """Bring the number of pool processes up to the specified number,
-    for use after reaping workers which have exited.
-    """
-    for i in range(self._processes - len(self._pool)):
-        w = self.Process(
-            target=multiprocessing.pool.worker,
-            args=(
-                self._inqueue,
-                self._outqueue,
-                self._initializer,
-                self._initargs,
-                self._maxtasksperchild,
-                self._wrap_exception,
-            ),
-        )
-        self._pool.append(w)
-        w.name = w.name.replace("Process", "PoolWorker")
-        w.daemon = False
-        w.start()
-        # util.debug('added worker')
+# def _repopulate_pool(self):
+#     """Bring the number of pool processes up to the specified number,
+#     for use after reaping workers which have exited.
+#     """
+#     for i in range(self._processes - len(self._pool)):
+#         w = self.Process(
+#             target=multiprocessing.pool.worker,
+#             args=(
+#                 self._inqueue,
+#                 self._outqueue,
+#                 self._initializer,
+#                 self._initargs,
+#                 self._maxtasksperchild,
+#                 self._wrap_exception,
+#             ),
+#         )
+#         self._pool.append(w)
+#         w.name = w.name.replace("Process", "PoolWorker")
+#         w.daemon = False
+#         w.start()
+#         # util.debug('added worker')
 
 
-multiprocessing.pool.Pool._repopulate_pool = _repopulate_pool
+workspace_global = []
+pool_global = []
 
-process_global = []
+SIGINT_MAIN_TRACEBACK = False
+SIGINT_STATE = 0
 
-def kill_processes(sig, frame):
[…]
