+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -39,6 +39,7 @@ from besmarts.mechanics import objectives
 from besmarts.mechanics import hessians
 from besmarts.mechanics import vibration
 from besmarts.mechanics import molecular_models as mm
+from besmarts.mechanics import smirnoff_models
 
 import scipy.optimize
 eid_t = int
@@ -48,6 +49,9 @@ PRECISION = configs.precision
 POSITIONS = assignments.POSITIONS
 HESSIANS = assignments.HESSIANS
 GRADIENTS = assignments.GRADIENTS
+
+au2kcal = 627.51
+kj2kcal = 1/4.184
 
 
 class compute_config:
@@ -62,6 +66,8 @@ class compute_config:
         self.addr = addr
         self.keys = []
         self.enable_minimization = True
+        self.fc_reset_config = None
+        self.method_vib_modes = "min"
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_table]]:
 
@@ -184,6 +190,21 @@ class compute_config_energy(compute_config):
                                     }
                                 )
 
+                    if self.fc_reset_config:
+                        if self.fc_reset_config.get("unique", False):
+                            psys = reset_unique(
+                                self.fc_reset_config,
+                                gde,
+                                psys
+                            )
+                        else:
+                            psys = reset(
+                                self.fc_reset_config,
+                                csys,
+                                assignments.graph_db_get_entries(gdb, eid),
+                                psystems={eid: psys}
+                            )[eid].value
+
                     ene = objectives.physical_system_energy(psys, csys)
 
                     tbl.values.append(ene)
@@ -258,6 +279,20 @@ class compute_config_gradient(compute_config):
                                         if k[0] == m
                                     }
                                 )
+                    if self.fc_reset_config:
+                        if self.fc_reset_config.get("unique", False):
+                            psys = reset_unique(
+                                self.fc_reset_config,
+                                gde,
+                                psys
+                            )
+                        else:
+                            psys = reset(
+                                self.fc_reset_config,
+                                csys,
+                                assignments.graph_db_get_entries(gdb, eid),
+                                psystems={eid: psys}
+                            )[eid].value
 
                     gx = optimizers_openmm.physical_system_gradient_openmm(
                         psys,
@@ -339,11 +374,31 @@ class compute_config_hessian(compute_config):
 
                     pos = psys.models[0].positions[0]
 
-                    if self.enable_minimization:
+                    if self.fc_reset_config:
+                        if self.fc_reset_config.get("unique", False):
+                            psys = reset_unique(
+                                self.fc_reset_config,
+                                gde,
+                                psys
+                            )
+                        else:
+                            psys = reset(
+                                self.fc_reset_config,
+                                csys,
+                                assignments.graph_db_get_entries(gdb, eid),
+                                psystems={eid: psys}
+                            )[eid].value
+
+                    if self.enable_minimization or self.method_vib_modes == "min":
                         minpos = optimizers_openmm.optimize_positions_openmm(csys, psys, tol=1e-7)
                         pos = minpos
                         tbl_pos.values = minpos
                         psys.models[0].positions[0] = minpos
+
[…]
