### 2024-08-27 93c9ac9
modified chunk logic
