+++ besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py
@@ -7,6 +7,7 @@ import copy
 import scipy.optimize
 import numpy as np
 import datetime
+import math
 
 from besmarts.core import geometry
 from besmarts.core import assignments
@@ -164,7 +165,8 @@ def objective_gradient_gdb(
     psysref=None,
     reuse=None,
     ws=None,
-    verbose=False
+    verbose=False,
+    return_gradient=True
 ):
 
     # if ws:
@@ -190,9 +192,11 @@ def objective_gradient_gdb(
 
     # big job, try to start computing while tasks are being generated
     # also, reap objective as it comes due to memory consumption
-    async_compute = True and (ws is not None) and len(args)*len(objlst) > 10000
+    async_compute = True and (ws is not None) and len(args)*len(objlst) > 1000
     # async_compute = True and (ws is not None)
     async_compute = False
+
+    verbose = verbose and not async_compute
 
     line = (
         f"{logs.timestamp()} Generating {len(objlst)} objectives. "
@@ -201,6 +205,9 @@ def objective_gradient_gdb(
     out.append(line)
     if verbose:
         print(line)
+
+    verbose = verbose and not async_compute
+
     dcsys = csys
     if ws:
         dcsys = None
@@ -209,29 +216,63 @@ def objective_gradient_gdb(
     hi = 1e-6
     h = tuple(([hi] * len(z)))
 
-    hp = tuple(1e-6/j[1] for j in priors)
+    hp = tuple()
+    if return_gradient:
+        hp = tuple(1e-6/j[1] for j in priors)
+
 
     n_finished = 0
 
-    chunksize = 20  # aim for about the expected size of workers
+    chunksize = 300
 
-    objbatchsize = 100 if len(objlst) > 100 else len(objlst)
-    if objbatchsize < len(objlst):
+    targetbatch = 300*len(keys)
+    if not return_gradient:
+        targetbatch *= 3
+    objbatches = []
+    cur_batch_score = 0
+    cur_batch = []
+
+    for i, obj in objlst.items():
+        bsz = obj.batch_size
+        if bsz is None:
+            cost = 1
+        else:
+            cost = len(keys) // bsz + bool(len(keys) % bsz)
+
+        if cur_batch_score + cost < targetbatch or len(cur_batch) == 0:
+            cur_batch.append((i, obj))
+            cur_batch_score += cost
+        else:
+            objbatches.append(cur_batch)
+            cur_batch = []
+
+    if cur_batch:
+        objbatches.append(cur_batch)
+        cur_batch = []
+        cur_batch_score = 0
+
+    objbatchsize = 500 if len(objlst) > 500 else len(objlst)
+
+    n_objbatches = len(objbatches)
+    if len(objbatches) > 1 and verbose:
 
         line = (
-            f"{logs.timestamp()} Processing objectives in batches of "
-            f"{objbatchsize}"
+            f"{logs.timestamp()} Processing objectives in "
+            f"{len(objbatches)} batches "
         )
-        if verbose:
-            print(line)
+        print(line)
 
-    n_objbatches = len(objlst)//objbatchsize + bool(len(objlst)%objbatchsize)
-    for obsz, obj in enumerate(arrays.batched(objlst.items(), objbatchsize), 1):
+    # n_objbatches = len(objlst)//objbatchsize + bool(len(objlst)%objbatchsize)
+    # for obsz, obj in enumerate(arrays.batched(objlst.items(), objbatchsize), 1):
+    batches_cum_sum = 0
+    for obsz, obj in enumerate(objbatches, 1):
         tasks = {}
         full_results = {}
         batch_map = {}
         chunk = {}
[…]
