### 2024-09-02 #6653 @lucydot
@AntObi @Ruibin-Liu  - are you able to take another look at the documentation/installation/tests after improvements made by authors?
