### 2024-09-05 trevorgokey/besmarts#1 @AntObi
Hi @trevorgokey ,

I've taken another look at the repository and have some further issues I've spotted

# Installation
Currently, the installation in a fresh environment fails for me, which I have identified the causes.

For the setup.py for `besmarts-rdkit`, please remove `besmarts.perceptron` as `besmarts-rdkit/python/besmarts/perceptron` does not exist. It prevents the user from installing `besmarts-rdkit`:
https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-rdkit/python/setup.py#L17-L22

As well for `besmarts-scipy`, please remove `besmarts.core` in the setup.py file.
https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-scipy/python/setup.py#L18-L22

# Tests

Currently 22/23 tests pass. The test `test_import_core` in  `besmarts-core/python/tests/test_import.py` fails  with an `ImportError` ` ImportError: cannot import name 'tree_merge' from 'besmarts.core' (unknown location)`. I've searched the codebase and `tree_merge` does not exist. Please remove `tree_merge` from the lines below.

https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-core/python/tests/test_import.py#L10-L34

# Examples

## [besmarts-core/python/examples/](https://github.com/trevorgokey/besmarts/tree/joss-manuscript/besmarts-core/python/examples)

My configuration for running these examples was as follows:
* OS: macOS
* Python: 3.10
* Packages installed: besmarts-core, besmarts-rdkit, besmarts-scipy, besmarts-openmm

In [besmarts-core/python/besmarts/mechanics/fits.py](https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-core/python/besmarts/mechanics/fits.py#L35) the following import statement exists https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-core/python/besmarts/mechanics/fits.py#L35

`array_numpy` does not appear to exist in `besmarts.core` which is causing all the examples that use `from besmarts.mechanics import fits` to fail for me. These examples would be examples  07, 09, and 11. As this import statement is also in [besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py](https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py#L20) https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-scipy/python/besmarts/mechanics/optimizers_scipy.py#L20  Example 10 also fails because of this.

Example 4 is failing for me because of this line: https://github.com/trevorgokey/besmarts/blob/93c9ac99344ada317050f566b7b2689a455750c5/besmarts-core/python/examples/04_init_bonds_angles.py#L11

`perception_rdkit` does not appear to exist in this codebase.

I will follow up by testing the examples in the documentation as well as elsewhere in the repository where I find them
