### 2024-09-05 #6653 @AntObi
I've had another look and I've found new issues with the installation and the examples, see my new comment in https://github.com/trevorgokey/besmarts/issues/1#issuecomment-2331334662
