### 2024-09-07 0091175
add perception_rdkit.py
