+++ besmarts-scipy/python/setup.py
@@ -17,7 +17,6 @@ setuptools.setup(
     long_description_content_type="text/markdown",
     packages=[
         'besmarts',
-        'besmarts.core',
         'besmarts.mechanics',
     ],
     install_requires=requirements,
