+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -41,7 +41,8 @@ from besmarts.mechanics import vibration
 from besmarts.mechanics import molecular_models as mm
 from besmarts.mechanics import smirnoff_models
 
-import scipy.optimize
+# import scipy.optimize
+# import scipy.stats
 eid_t = int
 
 PRECISION = configs.precision
@@ -240,6 +241,7 @@ class compute_config_gradient(compute_config):
             results: List[Dict[assignments.tid_t, assignments.graph_db_table]]
             for eid, gde in gdb.entries.items():
                 tbl = assignments.graph_db_table(topology.atom)
+                tbl.values = []
                 rids = assignments.graph_db_table_get_row_ids(gde[tid])
                 for rid in rids:
                     system = []
@@ -298,8 +300,20 @@ class compute_config_gradient(compute_config):
                         psys,
                         csys
                     )
+                    # gx = objectives.physical_system_gradient(psys, csys)
+
                     gx = arrays.array_round(gx, PRECISION)
+                    # print(f"GRADIENT EID {eid}", gx)
                     tbl.values.extend(gx)
+
+                    # gq = objectives.physical_system_gradient_internal(psys, csys)
+                    # tbl.values['q'] = gq
+
+                    # qic = objectives.physical_system_internals(psys, csys)
+                    # tbl.values['ic'] = qic
+
+                    # jac = objectives.physical_system_bmatrix(psys, csys)
+                    # tbl.values['b'] = jac
 
                 r = {tbl_idx: tbl}
                 # print("Result is\n", r)
@@ -812,8 +826,20 @@ class objective_config_gradient(objective_config):
                     dxyz = " ".join([f"{xi:12.4e}" for xi in totalxx])
                     out.append(f"   Sum: MM: {xyz1} QM: {xyz0} dG: {dxyz}")
 
+                    # show the IC grads
+                    # i = len(out)
+                    # out.append(f" EID {eid} IC gradients:")
+                    # gq = tbl.values['q']
+                    # gqic = tbl.values['ic']
+                    # jac = tbl.values['b']
+                    # for m, vals in gq.items():
+                    #     for ic, v in vals.items():
+                    #         q0 = " ".join([f"{xi:12.4e}" for xi in v])
+                    #         icq = " ".join([f"{xi:12.4e}" for xi in gqic[m][ic]])
+                    #         b = " ".join([f"{xi:12.4e}" for x in jac[m].selections[ic] for xi in x[0]])
+                    #         out.append(f"{str(ic)} {q0} kJ/mol/q q: {icq} b: {b}")
                     if verbose:
-                        for i in range(len(out) - nlines):
+                        for i in range(len(out)-nlines, len(out)):
                             print(out[i])
                 X.extend(x)
             if rmse:
@@ -890,7 +916,7 @@ class objective_config_hessian(objective_config):
                 pos1 = etbls[POSITIONS].values
                 xyz = list([x[0] for x in pos1.selections.values()])
 
-            if self.method_vib_modes in ["min", "mm"]:
+            if False and self.method_vib_modes in ["min", "mm"]:
                 sym = graphs.graph_symbols(g)
                 mass = [[vibration.mass_table[sym[n]]]*3 for n in sym]
                 sym = list(sym.values())
@@ -1018,7 +1044,7 @@ class objective_config_hessian(objective_config):
[…]
