### 2024-09-07 25ac3ed
initial array_numpy.py
