### 2024-09-07 1d1e165
initial array_numpy.py
