+++ besmarts-core/python/besmarts/core/compute.py
@@ -996,7 +996,10 @@ def remote_connect(mgr, timeout=TIMEOUT):
             success,
         ),
     )
-    t.start()
+    try:
+        t.start()
+    except RuntimeError:
+        return False
 
     connected = True
 
@@ -1065,7 +1068,10 @@ def manager_remote_get_status(wq, timeout=5.0):
         args=(wq, out),
     )
     print("manager_remote_get_status: starting thread")
-    t.start()
+    try:
+        t.start()
+    except RuntimeError:
+        return workspace_status.INVALID
 
     t.join(timeout=timeout)
     if len(out):
@@ -1083,7 +1089,10 @@ def manager_remote_get_state(wq, timeout=TIMEOUT):
         args=(wq, out),
     )
     print("manager_remote_get_state: starting thread")
-    t.start()
+    try:
+        t.start()
+    except RuntimeError:
+        return None
 
     t.join(timeout=timeout)
     if len(out):
@@ -1140,7 +1149,10 @@ def workqueue_remote_get_workspaces(wq, timeout=TIMEOUT):
         target=workqueue_remote_get_workspaces_thread,
         args=(wq.remote_workspaces, out, success),
     )
-    t.start()
+    try:
+        t.start()
+    except RuntimeError:
+        return {}
 
     t.join(timeout=timeout)
     if (not t.is_alive()) or success.is_set():
@@ -1157,7 +1169,10 @@ def workqueue_remote_put_workspaces(wq, wss, timeout=TIMEOUT):
         target=workqueue_remote_put_workspaces_thread,
         args=(wq.mgr, wss, success),
     )
-    t.start()
+    try:
+        t.start()
+    except RuntimeError:
+        return False
     t.join(timeout=timeout)
 
     return success.is_set()
@@ -1320,6 +1335,7 @@ def queue_get_nowait(q, block=False, timeout=TIMEOUT, n=1):
     # print(f"queue_get_nowait")
     # success = threading.Event()
     out = []
+    result = None
     thread_timeout = timeout
     if timeout is not None and timeout > 1.0:
         # give some space since the timeout is really for the thread timeout
@@ -1327,7 +1343,10 @@ def queue_get_nowait(q, block=False, timeout=TIMEOUT, n=1):
     t = threading.Thread(
         target=queue_get_nowait_thread, args=(q, out, n, block, timeout)
     )
-    t.start()
+    try:
+        t.start()
+    except RuntimeError:
+        return result
     # print(f"queue_get_nowait starting thread")
     t.join(timeout=thread_timeout)
     # print(f"queue_get_nowait joined")
@@ -1336,7 +1355,6 @@ def queue_get_nowait(q, block=False, timeout=TIMEOUT, n=1):
     # print(f"queue_get_nowait alive {alive}")
     # is_success = success.is_set()
     # print(f"queue_get_nowait success {is_success}")
-    result = None
     # print(f"out is {out}")
     if out:
         result = out[0]
@@ -1353,7 +1371,10 @@ def manager_remote_queue_qsize(q, timeout=TIMEOUT):
         target=manager_remote_queue_qsize_thread, args=(q, out)
     )
     # print(f"manager_remote_queue_qsize starting thread")
-    t.start()
+    try:
+        t.start()
+    except RuntimeError:
+        return sz
 
     t.join(timeout=timeout)
     # print(f"manager_remote_get_iqueue joined")
@@ -1373,7 +1394,10 @@ def manager_remote_get_iqueue(mgr, timeout=TIMEOUT):
         target=manager_remote_get_iqueue_thread, args=(mgr, out)
     )
     print(f"manager_remote_get_iqueue starting thread")
-    t.start()
+    try:
+        t.start()
[…]
