### 2024-09-07 9ed9ff7
handle linear molecules better
