### 2024-09-07 a6722c7
handle linear molecules better
