+++ besmarts-core/python/besmarts/core/splits.py
@@ -836,7 +836,7 @@ def process_split_distributed(pack, shm=None):
 
     try:
         bit, mapping = pack[0]
-    except:
+    except Exception:
         pack = (pack,)
 
     bj = None
@@ -1006,7 +1006,7 @@ def process_split_difference_distributed(pack, T, S0, splitter, A, selections, i
 
     try:
         bit, mapping = pack[0]
-    except:
+    except Exception:
         pack = (pack,)
 
     makes_split = False
@@ -1042,7 +1042,7 @@ def process_split_difference_distributed(pack, T, S0, splitter, A, selections, i
         # print(f"4WORKER DIFFERENCE FOR A {hash(Sj)} SPLITS {makes_split}")
         return Tsj, shard, h, matches, makes_split
 
-    if S0 == Sj:
+    if graphs.graph_same(S0, Sj):
         # print(f"4WORKER DIFFERENCE FOR A {hash(Sj)} SPLITS {makes_split}")
         return None, shard, h, matches, makes_split
 
@@ -1085,7 +1085,7 @@ def process_split_intersect_distributed(pack, T, S0, splitter, A, selections, ic
 
     try:
         bit, mapping = pack[0]
-    except:
+    except Exception:
         pack = (pack,)
 
     makes_split = False
@@ -1118,7 +1118,7 @@ def process_split_intersect_distributed(pack, T, S0, splitter, A, selections, ic
         # print(f"4WORKER INTERSECT FOR A {hash(Sj)} SPLITS {makes_split}")
         return Tsj, shard, h, matches, makes_split
 
-    if S0 == Sj:
+    if graphs.graph_same(S0, Sj):
         return None, shard, h, matches, makes_split
 
     matches, makes_split = process_split_matches(Sj, A, selections, icd, False)
@@ -1460,7 +1460,13 @@ def split_subgraphs_distributed(
     # this should be encoded and whatnot before start
     # shm = shm_split_subgraphs(splitter, S0, A)
 
-    shm = {"splitter": splitter, "S0": S0, "G": G, "selections": selections, "icd": icd}
+    shm = {
+        "splitter": splitter,
+        "S0": S0,
+        "G": G,
+        "selections": selections,
+        "icd": icd
+    }
 
     # we need 1 for this main process, and the other for the workspace server
     nproc = max(1, configs.processors - 1)
@@ -1552,7 +1558,7 @@ def split_subgraphs_distributed(
         for i in range(min_bits, uptobits):
             n = offsets[i]
             these_results = compute.workspace_flush(
-                ws, set([x[0] for x in batch if x[0] < n]), timeout=0.0,
+                ws, set([x[0] for x in batch if x[0] < n]), timeout=0.1,
             )
             for idx, splits in sorted(
                 these_results.items(), key=lambda x: x[0]
@@ -1607,10 +1613,12 @@ def split_subgraphs_distributed(
 
                     if matches is None or len(matches) == 0:
                         # print(f"MATCHES IS {matches}")
+                        # print(f"CND SPLITS=O {sma}")
                         continue
 
                     if hashshard in visited:
                         # print(f"HASH DUP: {hashshard}")
+                        # print(f"{idx+1:5d} CND DUPLICATE {sma}")
                         continue
                     else:
                         # print(f"HASH NEW: {hashshard}")
@@ -1619,6 +1627,7 @@ def split_subgraphs_distributed(
                     sma = gcd.smarts_encode(Sj)
 
                     if sma in sma_visited:
+                        # print(f"{idx+1:5d} CND DUPLICATE {sma}")
                         continue
                     else:
                         sma_visited.add(sma) 
@@ -1640,8 +1649,9 @@ def split_subgraphs_distributed(
[…]
