### 2024-09-07 635940d
expose internal grads/coords
