+++ besmarts-core/python/besmarts/mechanics/smirnoff_models.py
@@ -340,16 +340,9 @@ def chemical_model_electrostatics_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     # proc.topology_parameters[(0, i.name)] = {"s": i.name}
     # cm.topology_terms["s"].values[i.name] = [0.0]
 
-    # 13 scaling (on)
-    i = proc.smarts_hierarchies[0].index.node_add_below(None)
-    i.name = "s3"
-    i.type = "parameter"
-    i.category = [-1, pid, uid]
-    proc.smarts_hierarchies[0].smarts[i.index] = "[*:1]~[*]~[*:2]"
-    proc.topology_parameters[(0, i.name)] = {"s": i.name}
-    cm.topology_terms["s"].values[i.name] = [0.0]
-
     # 14 scaling (0.83)
+    # Put this before s3 because rings would match this first, when 1-3 is
+    # closer (think r5)
     i = proc.smarts_hierarchies[0].index.node_add_below(None)
     i.name = "s4"
     i.type = "parameter"
@@ -359,54 +352,19 @@ def chemical_model_electrostatics_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     cm.topology_terms["s"].values[i.name] = [1 / 1.2]
     cm.procedures.append(proc)
 
+    # 13 scaling (on)
+    i = proc.smarts_hierarchies[0].index.node_add_below(None)
+    i.name = "s3"
+    i.type = "parameter"
+    i.category = [-1, pid, uid]
+    proc.smarts_hierarchies[0].smarts[i.index] = "[*:1]~[*]~[*:2]"
+    proc.topology_parameters[(0, i.name)] = {"s": i.name}
+    cm.topology_terms["s"].values[i.name] = [0.0]
+
     return cm
 
 def chemical_model_to_xml_dict(csys):
 
-    name_to_model = {
-        "Bonds": "Bonds",
-        "Angles": "Angles",
-        "Torsions": "ProperTorsion",
-        "OutOfPlanes": "ImproperTorsion",
-        "vdW": "vdW",
-        "LibraryCharges": "LibraryCharge",
-        "Constraints": "Constraint",
-        "ChargeIncrementModel": "ChargeIncrement"
-    }
-    model_to_parameter = {
-        "Bonds": "Bond",
-        "Angles": "Angle",
-        "ProperTorsions": "ProperTorsion",
-        "ImproperTorsions": "ImproperTorsion",
-        "vdW": "Atom",
-        "LibraryCharges": "LibraryCharge",
-        "Constraints": "Constraint",
-        "ChargeIncrementModel": "ChargeIncrement"
-    }
-    model_keys = {
-        "Bonds": ["smirks", "id", "k", "length"],
-        "Angles": ["smirks", "id", "k", "angle"],
-        "ProperTorsions": ["smirks", "id", "periodicity", "k", "phase"],
-        "ImproperTorsions": ["smirks", "id", "periodicity", "k", "phase"],
-        "vdW": ["smirks", "id", "epsilon", "sigma"],
-        "LibraryCharges": ["smirks", "id", "charge"],
-        "Constraints": ["smirks", "id", "c"],
-        "ChargeIncrementModel": ["smirks", "id", "charge"]
-    }
-    model_multi_parameter = {
-        "Bonds": False,
-        "Angles": False,
-        "ProperTorsions": True,
-        "ImproperTorsions": True,
-        "vdW": False,
-        "LibraryCharges": True,
-        "Constraints": True,
-        "ChargeIncrementModel": True
-    }
-
-    root_attrib = {
-        "version": "0.3", "aromaticity_model": "MDL"
-    }
     xml = {
         "SMIRNOFF": {
[…]
