### 2024-09-07 4c3cc4a
reinstate smarts resolver
