+++ besmarts-core/python/besmarts/core/resolve.py
@@ -1,0 +1,407 @@
+"""
+besmarts.core.enumerate
+
+Enumerate a SMARTS pattern into one or more SMILES
+"""
+
+
+from typing import Sequence, List, Tuple
+import itertools
+
+from besmarts.core import graphs
+from besmarts.core import compute
+from besmarts.core.rulesets import visitor_ruleset
+
+
+def resolve_smiles_recurse_iter(
+    beg: graphs.graph,
+    beg_adj,
+    visitor_rulesets: List[visitor_ruleset],
+    frag_cache,
+    nodes,
+    edges,
+    adj,
+    seen,
+):
+    for idx, node in beg.nodes.items():
+        if idx in nodes:
+            continue
+
+        edge_frags = get_edge_frags(beg, edges, beg_adj, adj, idx)
+
+        frags = frag_cache.get(hash(node), None)
+        if frags is None:
+            frags = [
+                frag
+                for frag in node.to_fragments()
+                if all(
+                    (ruleset.on_node(idx, frag) for ruleset in visitor_rulesets)
+                )
+            ]
+
+        for frag in frags:
+            nodes[idx] = frag
+
+            # molecule-so-far based rules
+            if not all(ruleset.on_nodes(nodes) for ruleset in visitor_rulesets):
+                nodes.pop(idx)
+                continue
+
+            nodes.pop(idx)
+
+            for edge_set in itertools.product(*edge_frags):
+                # atom and bonds rules
+                if not all(
+                    ruleset.on_node_edges(idx, frag, edge_set)
+                    for ruleset in visitor_rulesets
+                ):
+                    continue
+
+                for eidx, e in edge_set:
+                    edges[eidx] = e
+
+                # print("ACCEPTED")
+                nodes[idx] = frag
+
+                frag_hash = hash_fragment(nodes, edges)
+
+                if frag_hash in seen:
+                    continue
+
+                seen.add(frag_hash)
+                if len(nodes) == len(beg.nodes) and len(edges) == len(
+                    beg.edges
+                ):
+                    if all(
+                        (
+                            ruleset.on_graph(nodes, edges)
+                            for ruleset in visitor_rulesets
+                        )
+                    ):
+                        P = graphs.graph_copy(graphs.graph(nodes, edges))
+                        # print("-----------")
+                        # print(P.nodes)
+                        # print(P.edges)
+                        # print("***********")
+                        yield P
+                P = graphs.graph_copy(graphs.graph(nodes, edges))
+                yield from resolve_smiles_recurse_iter(
+                    beg,
+                    beg_adj,
+                    visitor_rulesets,
+                    frag_cache,
+                    P.nodes,
+                    P.edges,
+                    {i: j.copy() for i, j in adj.items()},
+                    seen,
+                )
+
+
+def resolve_process_fragment(
+    beg,
+    beg_adj,
+    visitor_rulesets,
+    frag_cache,
+    edge_frags,
+    nodes,
+    edges,
+    adj,
+    seen,
+    idx,
+    gcd,
+    frag,
+    shm=None,
+):
+
+    # print("Enter RPF")
+    results = []
+
+    nodes[idx] = frag
+
+    # molecule-so-far based rules
+    if not all(ruleset.on_nodes(nodes) for ruleset in visitor_rulesets):
+        nodes.pop(idx)
+        return []
+
+    nodes.pop(idx)
+
+    prods = list(itertools.product(*edge_frags))
+    # nprod = list(itertools.product(range(0, len(edge_frags))))
+    for i, edge_set in enumerate(prods, 0):
+        # atom and bonds rules
[…]
