+++ besmarts-core/python/besmarts/core/rulesets.py
@@ -1,0 +1,411 @@
+"""
+besmarts.core.rulesets
+
+Defines the rules for determining whether a given graph is "valid" where the
+current state is whether the graph is a valid molecule/SMILES
+"""
+
+from besmarts.core.primitives import primitive_key
+from besmarts.core.chem import bechem
+from besmarts.core.graph_visitors import graph_visitor
+from besmarts.core.graphs import graph
+from besmarts.core import arrays
+
+
+class visitor_ruleset(graph_visitor):
+    def __init__(self, primitive_codecs):
+        self.primitive_codecs = primitive_codecs
+        self.beg: graph = None
+        self.small_molecule_cutoff = 5
+        self.hydrogen_rule = ("n_minus", {None: None})
+        self.noncarbon_rule = ("n_max", {None: None})
+
+
+    def on_start(self, g: graph):
+        self.beg = g
+        return None
+
+    def on_stop(self, g: graph, result):
+        self.beg: graph = None
+        return None
+
+    def on_graph(self, nodes, edges):
+
+        codec = self.primitive_codecs[primitive_key.FORMAL_CHARGE]
+
+        if abs(codec.count_charge_smiles(nodes)) > 2:
+            return False
+
+        N = len(self.beg.nodes)
+        if len(nodes) == N:
+            # codec = self.primitive_codecs[primitive_key.ELEMENT]
+            # C = codec.count_element_smiles(nodes, n=6)
+
+            name, rule = self.hydrogen_rule
+            if name == "n_minus":
+                codec = self.primitive_codecs[primitive_key.HYDROGEN]
+                H = codec.count_hydrogen_smiles(nodes)
+
+                x = rule.get(N, rule[None])
+                if type(x) is float:
+                    # print("H", H, "total", N, (N-H)/(N+H))
+                    if H/N < x:
+                        return False
+                elif type(x) is int:
+                    if H < N - x:
+                        return False
+
+            if self.noncarbon_rule[0] == "n_max":
+                codec = self.primitive_codecs[primitive_key.ELEMENT]
+
+                C = codec.count_carbon_smiles(nodes)
+                N = len(self.beg.nodes)
+                calc = len(nodes) - C
+                rule = self.noncarbon_rule[1]
+
+                x = rule.get(N, rule[None])
+                # print("Calc:", calc, len(nodes), C, len(self.beg.nodes), x)
+                if x is not None and calc > x:
+                    return False
+
+        return True
+
+    def on_nodes(self, nodes):
+
+
+        return True
+
+    def on_node(self, idx, besmarts: bechem, *args, **kwargs):
+
+        valid = True
+        name = primitive_key.FORMAL_CHARGE
+        codec = self.primitive_codecs[name]
+        arr = besmarts.primitives[name]
+        q = codec.decode_int(arr.on_first())
+
+        name = primitive_key.ELEMENT
+        codec = self.primitive_codecs[name]
+        arr = besmarts.primitives[name]
+        e = codec.decode_int(arr.on_first())
+
+        name = primitive_key.VALENCE
+        codec = self.primitive_codecs[name]
+        arr = besmarts.primitives[name]
+        valence = codec.decode_int(arr.on_first())
+
+        # name = primitive_key.RING_SMALLEST
+        # codec = self.primitive_codecs[name]
+        # arr = besmarts.primitives[name]
+        # r = codec.decode_int(arr.on_first())
+
+        # name = primitive_key.CONNECTIVITY_RING
+        # codec = self.primitive_codecs[name]
+        # arr = besmarts.primitives[name]
+        # x = codec.decode_int(arr.on_first())
+
+        # name = primitive_key.AROMATIC
[…]
