+++ doc/gen.sh
@@ -1,3 +1,5 @@
-sphinx-apidoc -M --ext-autodoc --implicit-namespaces -e -f  -o besmarts ~/projects/besmarts/besmarts-core/python/besmarts
-sphinx-apidoc -M --ext-autodoc --implicit-namespaces -e -f  -o besmarts-rdkit ~/projects/besmarts/besmarts-rdkit/python/besmarts
-sphinx-apidoc -M --ext-autodoc --implicit-namespaces -e -f  -o besmarts-scipy ~/projects/besmarts/besmarts-scipy/python/besmarts
+opts="-M --ext-autodoc --implicit-namespaces -d 3 --remove-old -e -f"
+sphinx-apidoc $opts -o besmarts        ../besmarts-core/python/besmarts
+sphinx-apidoc $opts -o besmarts-rdkit  ../besmarts-rdkit/python/besmarts
+sphinx-apidoc $opts -o besmarts-scipy  ../besmarts-scipy/python/besmarts
+sphinx-apidoc $opts -o besmarts-openmm ../besmarts-scipy/python/besmarts
