+++ doc/besmarts/besmarts.core.resolve.rst
@@ -1,0 +1,7 @@
+besmarts.core.resolve module
+============================
+
+.. automodule:: besmarts.core.resolve
+   :members:
+   :undoc-members:
+   :show-inheritance:
