+++ doc/besmarts/modules.rst
@@ -2,6 +2,6 @@ besmarts
 ========
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 3
 
    besmarts
