+++ doc/besmarts-scipy/besmarts.core.array_numpy.rst
@@ -1,0 +1,7 @@
+besmarts.core.array\_numpy module
+=================================
+
+.. automodule:: besmarts.core.array_numpy
+   :members:
+   :undoc-members:
+   :show-inheritance:
