+++ doc/besmarts/besmarts.mechanics.hessians.rst
@@ -1,0 +1,7 @@
+besmarts.mechanics.hessians module
+==================================
+
+.. automodule:: besmarts.mechanics.hessians
+   :members:
+   :undoc-members:
+   :show-inheritance:
