+++ doc/besmarts/besmarts.sample.rst
@@ -1,0 +1,7 @@
+besmarts.sample module
+======================
+
+.. automodule:: besmarts.sample
+   :members:
+   :undoc-members:
+   :show-inheritance:
