+++ doc/besmarts/besmarts.mechanics.fits.rst
@@ -1,0 +1,7 @@
+besmarts.mechanics.fits module
+==============================
+
+.. automodule:: besmarts.mechanics.fits
+   :members:
+   :undoc-members:
+   :show-inheritance:
