+++ doc/besmarts/besmarts.mechanics.rst
@@ -1,17 +1,22 @@
 besmarts.mechanics namespace
 ============================
 
+.. py:module:: besmarts.mechanics
+
 Submodules
 ----------
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 3
 
+   besmarts.mechanics.fits
    besmarts.mechanics.force_harmonic
    besmarts.mechanics.force_pairwise
    besmarts.mechanics.force_periodic
+   besmarts.mechanics.hessians
    besmarts.mechanics.masses
    besmarts.mechanics.molecular_models
    besmarts.mechanics.objectives
    besmarts.mechanics.smirnoff_models
    besmarts.mechanics.smirnoff_xml
+   besmarts.mechanics.vibration
