+++ doc/besmarts/besmarts.cluster.rst
@@ -1,11 +1,13 @@
 besmarts.cluster namespace
 ==========================
 
+.. py:module:: besmarts.cluster
+
 Submodules
 ----------
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 3
 
    besmarts.cluster.cluster_assignment
    besmarts.cluster.cluster_objective
