+++ doc/besmarts/besmarts.smiles.rst
@@ -1,0 +1,7 @@
+besmarts.smiles module
+======================
+
+.. automodule:: besmarts.smiles
+   :members:
+   :undoc-members:
+   :show-inheritance:
