### 2024-09-10 d416fe7
increase gradient obj scale
