### 2024-09-10 #6653 @trevorgokey
Sorry, I thought I had fixed all of the installation issues with a colleague. I did another pass and I can now install a fresh working environment again
