### 2024-09-10 61eb2fe
increase verbosity
