### 2024-09-10 trevorgokey/besmarts#1 @trevorgokey
It turned out I had a lot of local files that were not commited to the repo so you got a lot of missing file issues. This should all be fixed. Thanks for your patience, persistence, and diligence here.
