### 2024-10-02 #6653 @lucydot
@AntObi @Ruibin-Liu -  have you had a chance to look through after changes @trevorgokey has made?

@trevorgokey  - I encourage you to double check installation and examples will correctly run, so that our reviewers can progress with other aspects of the review when able.
