### 2024-11-24 #6653 @Ruibin-Liu
Hi @lucydot, I am reviewing the docs. @trevorgokey Here are what I found that might need to be improved.

- In this line [`wq = compute.workqueue_local("127.0.0.1", 63210)`](https://besmarts.readthedocs.io/en/latest/splitting.html), I got the error `_pickle.PicklingError: Can't pickle <function AutoProxy at 0x1050c8ae0>: it's not the same object as multiprocessing.managers.AutoProxy` when I tested in MacOS 12.5 and python3.12. The other code block in the `Numerical/Iterative search` section hasn't been tested yet because of the error.
- For the code that increases maxmoves to 2, I couldn't get the results as shown in the docs but just
```
0 [#6H3:1](-[#1H0])(-[#1H0])(-[#1H0])-[#6H2:2](-[#1H0])(-[#1H0])-[#8H1]-[#1H0]
1 [#6H2:2](-[#6H3](-[#1H0])(-[#1H0])-[#1H0])(-[#1H0])(-[#1H0])-[#8H1:3]-[#1H0]
3 [#6H3:1](-[#1H0])(-[#1H0:5])(-[#1H0])-[#6H2](-[#1H0])(-[#1H0])-[#8H1]-[#1H0]
4 [#6H3:1](-[#1H0])(-[#1H0])(-[#1H0:6])-[#6H2](-[#1H0])(-[#1H0])-[#8H1]-[#1H0]
5 [#6H2:2](-[#6H3](-[#1H0])(-[#1H0])-[#1H0])(-[#1H0:7])(-[#1H0])-[#8H1]-[#1H0]
6 [#6H2:2](-[#6H3](-[#1H0])(-[#1H0])-[#1H0])(-[#1H0])(-[#1H0:8])-[#8H1]-[#1H0]
7 [#8H1:3](-[#1H0:9])-[#6H2](-[#1H0])(-[#1H0])-[#6H3](-[#1H0])(-[#1H0])-[#1H0]
2 -> [#6H3:1](-[#1H0:4])(-[#1H0])(-[#1H0])-[#6H2](-[#1H0])(-[#1H0])-[#8H1]-[#1H0]
LUN:  [#6H3:1]-[#1H0:4]
LHS:  [#6H3:1]-[#1H0:4]
RUN:  [#6,#8;!H0!H4:1]-[#1,#6;H0,H2:2]
RHS:  [__:1]-[__:2]
LHS_DIFF:  [#6H3:1]_[#1H0:4]
LHS_INVE:  [#6H3:1]-[#1H0:4]
RHS_DIFF:  [__:1]_[__:2]
BESTLHS:  [#6H3:1]-[#1H0:2]
```
where I just copied all the code blocks in the `Analytic/Direct search` section and run as a python script.

- In the `Bitwise Operations` chapter, the line `A, B = graphs.graph_to_structure_atoms(g)[:2]:` should not have the last `:`.
- In the `Clustering` chapter, the lines `cst = cluster_classifications(gcd, labeler, sag)` and `cst = cluster_means(gcd, labeler, sag, objective=objective)` lead to the same `_pickle.PicklingError` as above.
- In the `Force Field Fitting` chapter, it seems the file `openff-2.1.0.offxml` cannot be found. Do we need to download it from somewhere?
- `FileNotFoundError: [Errno 2] No such file or directory: 'sqm'` occurs for the `4. Initializing FF parameters (04_init_bonds_angles.py)`, `7. Automated chemical perception for force field design (07_besmarts_fit.py)`, `9. Minimize a molecules energy by optimizing positions (09_minimize.py)`, and `10. Vibrational frequencies of ethane with OpenMM (10_vibfreq.py)`.
- The same `_pickle.PicklingError:` error occurs for the `6. Cluster bond lengths (06_smarts_cluster_bond_lengths.py)`, `8. Find all splits using numerical SMARTS search (08_smarts_split.py)`, and `11. Ab-Initio (valence) chemical perception (AICP) (11_aicp.py)` examples.

In addition, I think it would be better if the package is available in PyPI.
