### 2024-12-02 b700fb8
detect linear torsions
