+++ besmarts-core/python/besmarts/core/splits.py
@@ -449,7 +449,7 @@ def split_partition(
 
         winner = False
         this_lhs_nummoves = len(lhs_removeA) + len(lhs_removeB)
-        if (lhs_nummoves == 0) and lhs:
+        if (this_lhs_nummoves == 0) and lhs:
             relabel = {x: i for i, x in enumerate(lhs.select, 1)}
             for i, x in enumerate(lhs.nodes, len(lhs.select) + 1):
                 if x not in lhs.select:
@@ -463,7 +463,7 @@ def split_partition(
             winner = True
             lhs_nummoves = this_lhs_nummoves
         this_rhs_nummoves = len(rhs_removeA) + len(rhs_removeB)
-        if (rhs_nummoves == 0) and rhs:
+        if (this_rhs_nummoves == 0) and rhs:
             relabel = {x: i for i, x in enumerate(rhs.select, 1)}
             for i, x in enumerate(rhs.nodes, len(rhs.select) + 1):
                 if x not in rhs.select:
@@ -479,10 +479,15 @@ def split_partition(
         if winner:
             break
 
-        lhs_better = lhs and this_lhs_nummoves < lhs_nummoves
-        rhs_better = rhs and this_rhs_nummoves < rhs_nummoves
-        first = (bestlhs is None and bestrhs is None) and (lhs or rhs)
-        if first or lhs_better:
+        lhs_better = bool(lhs and this_lhs_nummoves <= lhs_nummoves)
+        rhs_better = bool(rhs and this_rhs_nummoves <= rhs_nummoves)
+        if lhs_better and rhs_better:
+            if lhs_nummoves <= rhs_nummoves:
+                rhs_better = False
+            else:
+                lhs_better = False
+        first = bool((bestlhs is None and bestrhs is None) and (lhs or rhs))
+        if lhs and (first or lhs_better):
             relabel = {x: i for i, x in enumerate(lhs.select, 1)}
             for i, x in enumerate(lhs.nodes, len(lhs.select) + 1):
                 if x not in lhs.select:
@@ -491,10 +496,16 @@ def split_partition(
             if gcd:
                 print("BESTLHS: ", gcd.smarts_encode(lhs))
             bestlhs = graphs.structure_copy(lhs)
+
             best_lhs_removeA = lhs_removeA
             best_lhs_removeB = lhs_removeB
+
+            best_rhs_removeA = rhs_removeA
+            best_rhs_removeB = rhs_removeB
             lhs_nummoves = this_lhs_nummoves
-        if first or rhs_better:
+            rhs_nummoves = this_rhs_nummoves
+
+        if rhs and (first or rhs_better):
             relabel = {x: i for i, x in enumerate(rhs.select, 1)}
             for i, x in enumerate(rhs.nodes, len(rhs.select) + 1):
                 if x not in rhs.select:
@@ -503,8 +514,12 @@ def split_partition(
             if gcd:
                 print("BESTRHS: ", gcd.smarts_encode(rhs))
             bestrhs = graphs.structure_copy(rhs)
+
+            best_lhs_removeA = lhs_removeA
+            best_lhs_removeB = lhs_removeB
             best_rhs_removeA = rhs_removeA
             best_rhs_removeB = rhs_removeB
+            lhs_nummoves = this_lhs_nummoves
             rhs_nummoves = this_rhs_nummoves
 
     return success(
@@ -1509,22 +1524,25 @@ def split_subgraphs_distributed(
     k = 0
     n = 0
     chunksize = nproc
-    while len(completed) < len(iterable):
+    while len(iterable):
         Bn = n_ops * (
             math.factorial(len(single_bits))
             // (math.factorial(len(single_bits) - i) * math.factorial(i))
         )
 
-
-        unfinished = {
-            idx: unit
-            for idx, unit in iterable.items()
[…]
