### 2024-12-02 f92ecfb
analytic MM hessians
