+++ besmarts-core/python/besmarts/mechanics/objectives.py
@@ -10,6 +10,7 @@ from besmarts.core import primitives
 from besmarts.core import assignments
 from besmarts.mechanics import molecular_models as mm
 
+import numpy as np
 
 def array_flatten_assignment(selections):
     lst = []
@@ -76,6 +77,81 @@ def physical_system_force(psys: mm.physical_system, csys):
     return arrays.array_scale(force, 4.184)
 
 
+def physical_system_gradient_system(psys: mm.physical_system, csys):
+    f = physical_system_force_system(psys)
+    for ic, x in f.items():
+        f[ic] = [(k, -q) for k, q in x]
+    return f
+
+def physical_system_force_system(psys: mm.physical_system, csys):
+    """
+    csys is for the reference functions and system params
+    psys is for the positions and param values
+    """
+    force = list([0.0]*3*len(psys.models[0].positions[0].graph.nodes))
+
+    for m, pm in enumerate(psys.models):
+
+        refpos = pm.positions[0]
+        pos = assignments.graph_assignment_float(
+                refpos.graph,
+                {k: v.copy() for k, v in refpos.selections.items()}
+        )
+
+        if csys.models[m].derivative_function:
+
+            icq = csys.models[m].internal_function(pos)
+
+            system_terms = {
+                k: v.values for k, v in csys.models[m].system_terms.items()
+            }
+
+            params = pm.values
+            f = mm.smiles_assignment_function(
+                csys.models[m].force_system,
+                system_terms,
+                params,
+                icq
+            )
+
+    return f
+
+
+def physical_system_force_gradient(psys: mm.physical_system, csys):
+    """
+    csys is for the reference functions and system params
+    psys is for the positions and param values
+    """
+    force = list([0.0]*3*len(psys.models[0].positions[0].graph.nodes))
+
+    hq = {}
+    for m, pm in enumerate(psys.models):
+
+        refpos = pm.positions[0]
+        pos = assignments.graph_assignment_float(
+                refpos.graph,
+                {k: v.copy() for k, v in refpos.selections.items()}
+        )
+
+        if csys.models[m].derivative_function:
+
+            icq = csys.models[m].internal_function(pos)
+
+            system_terms = {
+                k: v.values for k, v in csys.models[m].system_terms.items()
+            }
+
+            params = pm.values
+            f = mm.smiles_assignment_function(
+                csys.models[m].force_gradient_system,
+                system_terms,
+                params,
+                icq
+            )
+            hq[m] = f
+
+    return hq
+
 def physical_system_bmatrix(psys: mm.physical_system, csys):
     """
     csys is for the reference functions and system params
@@ -125,6 +201,58 @@ def physical_system_internals(psys: mm.physical_system, csys):
             }
 
     return ic
+
+
+def physical_system_force_gradient_internal(psys: mm.physical_system, csys):
+    """
+    csys is for the reference functions and system params
+    psys is for the positions and param values
+    """
+
+    fgrad = {}
+
+    for m, pm in enumerate(psys.models):
+
+        refpos = pm.positions[0]
+        pos = assignments.graph_assignment_float(
+                refpos.graph,
+                {k: v.copy() for k, v in refpos.selections.items()}
+        )
+
+        if csys.models[m].derivative_function:
+
+            fgrad[m] = {}
+            icq = csys.models[m].internal_function(pos)
+
+            system_terms = {
[…]
