+++ besmarts-core/python/besmarts/mechanics/force_harmonic.py
@@ -2,15 +2,10 @@
 besmarts.mechanics.force_harmonic
 """
 
-import math
 from typing import List
 
 from besmarts.core import topology
 from besmarts.core import assignments
-from besmarts.core import trees
-from besmarts.core import codecs
-from besmarts.core import hierarchies
-from besmarts.core import primitives
 from besmarts.core import perception
 
 from besmarts.mechanics import molecular_models as mm
@@ -41,6 +36,45 @@ def energy_function_spring(
 
     return result
 
+def force_gradient_function_spring(
+    *, k: List[float], l: List[float], x: List[float]
+) -> List[float]:
+    result = []
+    for xi in x[0]:
+        xres = []
+        for kj, lj in zip(k, l):
+            xres.append(-kj)
+        result.append(xres)
+
+    return result
+
+
+def force_gradient_system_spring(
+    *, k: List[float], l: List[float], x: List[float]
+) -> List[float]:
+    result = []
+    for xi in x[0]:
+        xres = []
+        for kj, lj in zip(k, l):
+            xres.append((-kj, -1.0))
+        result.append(xres)
+
+    return result
+
+
+def force_system_spring(
+    *, k: List[float], l: List[float], x: List[float]
+) -> List[float]:
+    result = []
+    for xi in x[0]:
+        xres = []
+        for kj, lj in zip(k, l):
+            r = xi - lj
+            xres.append((-kj, -r))
+        result.append(xres)
+
+    return result
+
 
 def smiles_assignment_energy_function_spring(pos, params):
     ene = {}
@@ -62,6 +96,11 @@ def chemical_model_bond_harmonic(
 
     cm.energy_function = energy_function_spring
     cm.force_function = force_function_spring
+    cm.force_gradient_function = force_gradient_function_spring
+
+    cm.force_system = force_system_spring
+    cm.force_gradient_system = force_gradient_system_spring
+
     cm.internal_function = assignments.graph_assignment_geometry_bonds
     cm.derivative_function = assignments.graph_assignment_jacobian_bonds
 
@@ -87,6 +126,11 @@ def chemical_model_angle_harmonic(
 
     cm.energy_function = energy_function_spring
     cm.force_function = force_function_spring
+    cm.force_gradient_function = force_gradient_function_spring
+
+    cm.force_system = force_system_spring
+    cm.force_gradient_system = force_gradient_system_spring
+
     cm.internal_function = assignments.graph_assignment_geometry_angles
     cm.derivative_function = assignments.graph_assignment_jacobian_angles
 

[…]
