### 2024-12-02 ceb8724
truncate B to 99%
