+++ besmarts-core/python/besmarts/mechanics/smirnoff_models.py
@@ -18,23 +18,12 @@ from besmarts.mechanics import smirnoff_xml
 from besmarts.mechanics import force_harmonic, force_periodic, force_pairwise
 
 PRECISION = configs.precision
-sigma2rmin_half =  1 / 2 ** (5 / 6)
+sigma2rmin_half = 1 / 2 ** (5 / 6)
 
 def chemical_model_bond_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
-    cm = mm.chemical_model("B", "Bonds", topology.bond)
 
-    cm.energy_function = force_harmonic.energy_function_spring
-    cm.force_function = force_harmonic.force_function_spring
-    cm.internal_function = assignments.graph_assignment_geometry_bonds
-    cm.derivative_function = assignments.graph_assignment_jacobian_bonds
-
-    # define the terms of this model
-    cm.topology_terms = {
-        "k": mm.topology_term(
-            "k", "stiffness", "float", "kcal/mol/A/A", {}, "", {}
-        ),
-        "l": mm.topology_term("l", "length", "float", "A", {}, "", {}),
-    }
+    cm = force_harmonic.chemical_model_bond_harmonic(pcp)
+    cm.name = "Bonds"
 
     ############################################################################
     # the terms are determined by a smarts matching
@@ -58,19 +47,28 @@ def chemical_model_bond_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
     )
     proc.smarts_hierarchies = {uid: h}
 
+
     root = proc.smarts_hierarchies[u.index].index.node_add_below(None)
     root.name = "Bonds"
     root.category = [-1, pid, uid]
     root.type = "hierarchy"
     proc.smarts_hierarchies[u.index].smarts[root.index] = None
 
+    label_to_id = {}
+    label_to_id[root.name] = root.index
+
+
     uid = u.index
     for param in d["parameters"]:
         param = param["Bond"]
+
+        above_name = param.get("parent_id", root.name)
+        above_idx = label_to_id[above_name]
         node = proc.smarts_hierarchies[u.index].index.node_add_below(
-            root.index
+            above_idx
         )
         node.name = param.get("id", "")
+        label_to_id[node.name] = node.index
         node.type = "parameter"
         node.category = [-1, pid, uid]
         h.smarts[node.index] = param.get("smirks", None)
@@ -93,20 +91,9 @@ def chemical_model_bond_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
 
 
 def chemical_model_angle_harmonic_smirnoff(d: Dict, pcp) -> mm.chemical_model:
-    cm = mm.chemical_model("A", "Angles", topology.angle)
 
-    cm.energy_function = force_harmonic.energy_function_spring
-    cm.force_function = force_harmonic.force_function_spring
-    cm.internal_function = assignments.graph_assignment_geometry_angles
-    cm.derivative_function = assignments.graph_assignment_jacobian_angles
-
-    cm.topology_terms = {
-        "k": mm.topology_term(
-            "k", "stiffness", "float", "kcal/mol/rad/read", {}, "", {}
-        ),
-        "l": mm.topology_term("l", "length", "float", "rad", {}, "", {}),
-    }
-
+    cm = force_harmonic.chemical_model_angle_harmonic(pcp)
+    cm.name = "Angles"
     ############################################################################
     # the terms are determined by a smarts matching
[…]
