+++ besmarts-core/python/examples/11_aicp.py
@@ -27,11 +27,13 @@ from besmarts.cluster import cluster_objective
 from besmarts.cluster import cluster_optimization
 from besmarts.mechanics import fits
 from besmarts.mechanics import smirnoff_models
+from besmarts.mechanics import molecular_models as mm
 
 from besmarts.codecs import codec_rdkit
 from besmarts.assign import hierarchy_assign_rdkit
 
-configs.processors = 1
+configs.compute_task_chunksize = 1000
+configs.processors = 16
 configs.remote_compute_enable = False
 # To use additional workers, from a terminal run (on the separate work machine)
 # python -m besmarts.worker <IP> 55555 np
@@ -44,7 +46,7 @@ configs.workqueue_port = 55555
 # to know it's doing something :)
 # Currently will show an update every 10% progress or 60 seconds, whichever
 # comes first.
-configs.compute_verbosity = 1
+configs.compute_verbosity = 10
 
 # Known issue: if the command is abnormally terminated, some processes will
 # not be killed and will not free the port. This will prevent the script from
@@ -66,7 +68,7 @@ def mystrategy():
     # bits, and do not look at neighboring atoms.
 
     splitter = configs.smarts_splitter_config(
-        1, 2, 0, 0, 0, 0, True, True, 0, False, True, True, True
+        1, 1, 0, 0, 0, 0, True, True, 0, False, True, True, True
     )
 
     # Disable general patterns like [!#6][!#8]. In general these will lead
@@ -128,6 +130,8 @@ def configure_tiers(objs, penalties, fit_models, fit_symbols):
 
     # Don't perform any stochastic optimization
     initial.anneal = 0
+    initial.step_limit = 200
+    initial.method = 'L-BFGS-B'
 
     # line search in LBFGS
     initial.maxls = 20
@@ -150,12 +154,12 @@ def configure_tiers(objs, penalties, fit_models, fit_symbols):
     # optimization.
     initial.priors = {
        "k": (None, 20),
-       ("k", "b"): (None, 50),
+       ("k", "b"): (None, 1),
        ("l", "b"): (None, .01),
-       ("k", "a"): (None, 50),
-       ("l", "a"): (None, .01),
-       ("k", "t"): (None, 5.0),
-       ("k", "i"): (None, 5.0),
+       ("k", "a"): (None, .5),
+       ("l", "a"): (None, .1),
+       ("k", "t"): (None, .1),
+       ("k", "i"): (None, .1),
        ("s", "s"): (None, .1),
        None: (None, None)
     }
@@ -172,6 +176,8 @@ def configure_tiers(objs, penalties, fit_models, fit_symbols):
 
     final.bounds = initial.bounds
     final.priors = initial.priors
+
+    final.method = "L-BFGS-B"
 
     tier = fits.objective_tier()
     tier.objectives = final.objectives
@@ -227,7 +233,7 @@ def configure_chem_perception_strat(csys, split_models):
 
     # The default is to merge then split, we want split then merge
     strat.build_steps()
-    strat.steps = list(reversed(strat.steps))
+    # strat.steps = list(reversed(strat.steps))
 
     return strat
 
@@ -249,10 +255,11 @@ def configure_objectives(opt_mols):
                 assignments.graph_db_address(
                     eid=[eid],
                 ),
-                scale=1e-4,
+                scale=1e-3,
                 include=True
         )
         o.batch_size = 1
+        # o.freq_range = (500, 600)
         o.method_vib_modes = "qm"
         o.verbose = 2
         objs[len(objs)] = o
@@ -262,12 +269,14 @@ def configure_objectives(opt_mols):
                 assignments.graph_db_address(
                     eid=[eid],
                 ),
[…]
