### 2024-12-02 #6653 @trevorgokey
Hi @Ruibin-Liu, thanks so much for going through everything! I don't think I would have noticed the docs mismatch without your help. It should work if you `git pull` now.

> I got the error _pickle.PicklingError

This is specifically a macos thing; in particular the difference to linux is the default context for `multiprocessing` where the linux default appears to be incompatible with macos, and the current besmarts implementation assumes the linux default at certain parts. I am going to poke around on the web and see if I can get a macos environment setup on linux to test things out. For now, since I can't test this, can you please add ```multiprocessing.set_start_method('fork')``` to the top of each tested file as a temporary fix? It would need to be added before anything in `besmarts.compute` is called. Additionally, another route is to set `besmarts.configs.processors = 1` at the top; ideally this will invoke a codepath that doesn't make use of `multiprocessing`, but I need to get a macos environment up to ensure that doing so will squash the issue and I don't have any uncovered `multiprocessing` calls. If setting the context to 'fork' works, I think I'll go ahead with updating `main` to do this explicitly until I can get it working with the macos default 'spawn'. The last remaining option is to test on a linux environment if you have one available.

> FileNotFoundError: [Errno 2] No such file or directory: 'sqm'

The ambertools `sqm` program calculates atomic partial charges which is optional but necessary for pipelines explored in some examples. It can be installed via source or conda (`conda install -c conda-forge ambertools`). This feeds into the pypi remark,

> I think it would be better if the package is available in PyPI

Right now since I assume that changes to the codebase are necessary and somewhat frequent (like the issue you found with overlap!), it doesn't make sense to keep reuploading to pypi. Additionally, not everything, e.g. ambertools and previously openmm, is on pypi and so uploading there won't exactly make things easier to install in some cases. Right now, I feel the easiest thing to do is to `git clone`, `pip install -e .` and then `git pull` as we go through the review process. After the dust settles, I do plan to upload a version to pypi. However if you feel strongly that an installable package exists on pypi right now, I can go ahead and get it uploaded.

> it seems the file openff-2.1.0.offxml cannot be found

The force field `openff-2.1.0.offxml` does indeed need to be downloaded. It can be found at [https://github.com/openforcefield/openff-forcefields/blob/main/openforcefields/offxml/openff-2.1.0.offxml](). I've updated the docs with the link as well.
