### 2024-12-05 #6653 @trevorgokey
Here is the follow up on the mac issue. I was able to get an OSX test system running with [https://github.com/kholia/OSX-KVM](OSX-KVM) (i.e. an emulated hackintosh), but only running Ventura because it is their latest recommend version. I get a different error than the `error _pickle.PicklingError` above, but I believe is still related to the same issue. Changing to 1 processor did not circumvent the issue, but adding `multiprocessing.set_start_method('fork')` did allow `11_aicp.py` to run successfully.
