### 2024-12-17 #6653 @trevorgokey
I've added a snippet to always set the start method to fork at [https://github.com/trevorgokey/besmarts/blob/main/besmarts-core/python/besmarts/core/compute.py#L34C1-L37C9]() so just an update should fixed the failed examples.
