### 2024-12-17 #6653 @lucydot
Hi @trevorgokey  - am I correct in thinking that a git pull of main branch will mean all of the tutorials can be ran on OSX?

> In addition, I think it would be better if the package is available in PyPI.

Is this something you have considered?
