### 2024-12-17 0f6ba5c
add analytic hessian comparisons
