+++ doc/fffit.rst
@@ -38,7 +38,7 @@ Furthermore, only the equilibrium lengths of the bonds will be modified.
 >>> from besmarts.core import compute
 >>> from besmarts.core import configs
 >>> 
->>> configs.processors = 4
+>>> configs.processors = 1
 >>> configs.remote_compute_enable = False
 >>> configs.workqueue_port = 54321
 >>> 
@@ -175,12 +175,12 @@ Furthermore, only the equilibrium lengths of the bonds will be modified.
 >>>     fit_models = [0]
 >>>     final = fits.objective_tier()
 >>>     final.objectives = {
->>>         # 0: fits.objective_config_position(
->>>         #         fits.graph_db_address(
->>>         #             eid=[0],
->>>         #         ),
->>>         #         scale=1
->>>         # ),
+>>>         0: fits.objective_config_position(
+>>>                 fits.graph_db_address(
+>>>                     eid=[0],
+>>>                 ),
+>>>                 scale=1
+>>>         ),
 >>>         1: fits.objective_config_gradient(
 >>>                 fits.graph_db_address(
 >>>                     eid=[0],
@@ -188,9 +188,13 @@ Furthermore, only the equilibrium lengths of the bonds will be modified.
 >>>                 scale=1
 >>>         ),
 >>>     }
+>>>     final.objectives[0].verbose = 2
+>>>     final.objectives[1].verbose = 2
 >>>     # final.key_filter = lambda x: x[0] in fit_models and x[1] == 'l'
 >>>     final.fit_models = fit_models
 >>>     final.fit_symbols = ["l"]
+>>> 
+>>>     final.method = "L-BFGS-B" 
 >>> 
 >>>     onestep = fits.objective_tier()
 >>>     onestep.objectives = final.objectives
@@ -199,6 +203,8 @@ Furthermore, only the equilibrium lengths of the bonds will be modified.
 >>>     # onestep.key_filter = lambda x: x[0] in fit_models and x[1] == 'l'
 >>>     onestep.fit_models = fit_models
 >>>     onestep.fit_symbols = ["l"]
+>>>     onestep.method = "L-BFGS-B" 
+>>> 
 >>>     tiers = [onestep] # have a default
 >>> 
 >>>     initial = final
@@ -248,11 +254,13 @@ the `forcefield_optimization_strategy` class. We also indicate that we only want
 to fit to equilibrium lengths as given by the parameter term symbol "l" (one
 could also include "k" to also fit spring force constants).
 
-Now for the output:
+Now for the a few snippets from the output:
 
+
+-- Snip 1 --
 .. code-block::
 
-    Initial assignments:
+    ### BESMARTS chemical perception on the following assignments ###
     Model:
     Tree:
      0   0 Bonds  
@@ -262,1151 +270,88 @@ Now for the output:
      1   0  b21  [#6:1]=[#8X1+0,#8X2+1:2] k: [1527.019744047] l: [1.221668642702]
      1   0  b70  [#6:1]-[#17:2] k: [368.4266150848] l: [1.722215272811]
      1   0  b85  [#6X3:1]-[#1:2] k: [775.3853383846] l: [1.081823673944]
-    Tree:
-     0   1 Angles  
-    Tree:
-     0   2 Torsions  
-    Tree:
-     0   3 OutOfPlanes  
-    Tree:
-    Tree:
-     0   5 vdW   
-    Tree:
[…]
