+++ doc/clustering.rst
@@ -43,20 +43,19 @@ same labels. Here is how to do it:
 
 
 .. code-block::
-    
+
     Assigning molecule     1/1 at depth 0
-    
     Labels per unique structure that need more depth
     There are 2/3 unique structures at depth 0
     There are 0 problems:
     Max depth is set to 0
-    2024-06-02 10:59:16.684934 Labeling subgraphs
-    2024-06-02 10:59:16.685488 Checking consistency...
+    2024-12-17 10:52:34.477112 Labeling subgraphs
+    2024-12-17 10:52:34.477964 Checking consistency...
     Optimization strategy is building steps...
-    2024-06-02 10:59:16.685556 The optimization strategy has the following iterations:
-    ->   1. op= 1 a=[0] b=1->1 d=0->0 n=0->3
+    2024-12-17 10:52:34.478116 The optimization strategy has the following iterations:
+    ->   1. op= 1 a=[0] b=1->1 d=0->0 n=0->1
          2. op=-1 a=[0] b=0->0 d=0->0 n=0->0
-         3. op= 1 a=[0] b=2->2 d=0->0 n=0->3
+         3. op= 1 a=[0] b=2->2 d=0->0 n=0->2
          4. op=-1 a=[0] b=0->0 d=0->0 n=0->0
          5. op= 1 a=[0] b=3->3 d=0->0 n=0->3
          6. op=-1 a=[0] b=0->0 d=0->0 n=0->0
@@ -64,19 +63,20 @@ same labels. Here is how to do it:
     1 p0
     N Targets: 1
     Step tracker for current macro step 1
+    p0 1
     
     
     *******************
-     2024-06-02 10:59:16.685609 iteration=   1 macro=  1/6 X=        0 params=(1|2) G=Y S=Y bits=1->3 depth=0->0 branch=0->3
+     2024-12-17 10:52:34.478233 iteration=   1 macro=  1/6 X=        0 params=(2|2) G=Y S=Y bits=1->3 depth=0->0 branch=0->3
     *******************
     
     Tree:
-    **  0   0 p0   {'a', 'b'} [*:1]~[*:2]
+    **  0   0 p0   {'b', 'a'} [*:1]~[*:2]
     =====
     
-    2024-06-02 10:59:16.685637 Saving checkpoint to chk.cst.p
-    2024-06-02 10:59:16.686071 Collecting SMARTS for p0 N=3/3 and setting to depth=0
-     == iteration=   2 macro=  1/6 micro=  1/1 operation=1 params=(1|2) cluster=p0   N= 3 overlap=[0] bits=1->1 depth=0->0 branch=0->3
+    2024-12-17 10:52:34.478282 Saving checkpoint to chk.cst.p
+    2024-12-17 10:52:34.479079 Collecting SMARTS for p0 N=3/3 and setting to depth=0
+     == iteration=   2 macro=  1/6 micro=  1/1 operation=1 params=(2|2) cluster=p0   N= 3 overlap=[0] bits=1->1 depth=0->0 branch=0->1
     
     Attempting to split p0:
     S0: [*:1]~[*:2]
@@ -85,13 +85,12 @@ same labels. Here is how to do it:
     000002 (0, (1, 2))              {'a'} [#6H1X2x0!rA+0:1]!@;#[#6H1X2x0!rA+0:2]
     000003 (0, (2, 4))              {'b'} [#6H1X2x0!rA+0:2]!@;-[#1H0X1x0!rA+0:4]
     
-    2024-06-02 10:59:16.689140 Union merging=3
-    2024-06-02 10:59:16.703539 Union merging=2
-    2024-06-02 10:59:16.739951 Union merging=1
[…]

[…]
