### 2024-12-19 #6653 @lucydot
> Yes, I made some comments in my earlier Dec. 2 post

Ah yes, apologies I forgot about this.

@Ruibin-Liu and @AntObi it sounds that OSX is now supported. I can see that functionality is the main outstanding issue, so if you are able to verify this it seems we are close to wrapping up. Thank-you very much for your reviewing efforts.

FYI @trevorgokey @Ruibin-Liu @AntObi  - I am going to take two weeks leave, and will be back on the 6th of Jan.
