### 2025-01-28 #6653 @lucydot
@Ruibin-Liu and @AntObi it sounds that OSX is now supported. I can see that functionality is the main outstanding issue, so if you are able to verify this it seems we are close to wrapping up. Do you have capacity for this over the next couple of weeks?
