### 2025-02-16 #6653 @Ruibin-Liu
I think it's good to go. A side comment before the authors push it to pypi. Maybe using optional dependencies for different use cases can unify the packaging and using experience, even for the one not listed in the subpackages like `sqm` in `ambertools`.
