### 2025-02-17 #6653 @trevorgokey
Thanks so much @Ruibin-Liu. I think the optional dependencies is a great idea and I should get that working when pypi packages are made. I think I wasn't able to get it working when installing directly from git and so manual installation of each module was necessary. I will definitely make ambertools a dependency so that manual installation is not required.
