### 2025-02-18 #6653 @AntObi
Hi @lucydot @trevorgokey , I'd say it good to go for JOSS.

Echoing @Ruibin-Liu 's comments, I would recommend looking into optional dependencies for the package in order to make the PyPI push easier. I'd also recommend setting CI tests using GitHub Actions to be able to verify that any updates to the package in pull requests don't break the existing package.
