### 2025-02-19 #6653 @lucydot
Thank you very much for your time and effort in reviewing this package @AntObi and @Ruibin-Liu. One final request - could you tick off the remaining checklist item if you think it has been achieved?

@trevorgokey I believe your plan is to upload this to PyPI once this part of the review is complete?
