### 2025-02-20 #6653 @trevorgokey
@lucydot Yes, if there are no other changes requested I would be happy to upload to pypi. I gather it might be a prerequisite to completing the review, so I'll be able to prioritize this beginning next week.
