### 2025-02-25 #6653 @lucydot
@trevorgokey  - yes, I agree a PyPI upload is important, as this will enable the most seamless installation for most people. Please let me know once it is uploaded to PyPI and I will test the install myself.
