### 2025-03-04 4f6fcc6
Document several API points as used by ff_optimize
