### 2025-03-04 c308ef6
Document several API points as used by ff_optimize
