### 2025-03-04 3d677df
Document several API points as used by ff_optimize
