+++ besmarts-mechanics/python/besmarts/mechanics/fits.py
@@ -10,6 +10,7 @@ import pickle
 import multiprocessing
 import collections
 import sys
+from typing import Protocol
 import math
 import itertools
 import numpy as np
@@ -754,10 +755,30 @@ class compute_config_penalty(compute_config):
 
 class objective_config:
     """
-    Configuration of an objective function for force field fitting. The
-    objective has two parts. The first part is how the properties are
-    calculated and is configured with a compute_config. The next is how the
+    Configuration of an objective function for force field fitting.
+
+    The objective has two parts. The first part is how the properties are
+    calculated and is configured with a :py:class:`compute_config` as returned
+    by the :py:meth:`get_task` method. The second is how the
     objective is computed from the properties and is configured here.
+
+    This is a base class and should generally not be instantiated directly;
+    instead, see the following classes for commonly used objectives:
+
+    :py:class:`objective_config_position`
+        Performs a geometry optimization and calculates the sum of squared
+        error (SSE). The root of the mean SSE is the RMSD.
+    :py:class:`objective_config_gradient`
+        The geometry is held fixed at the reference, and calculates the
+        objective as the SSE of the difference in QM/MM forces.
+
+    .. todo: More of the above, write a Parameters section
+
+    See Also
+    --------
+    objective_config_energy, objective_config_hessian, objective_config_penalty,
+    objective_config_position, objective_config_gradient
+
     """
 
     def __init__(self, addr, include=True, scale=1, coordinate_system="C"):
@@ -2215,12 +2236,36 @@ class objective_config_penalty(objective_config):
 
 class objective_tier:
     """
+    Description of a single tier in a multi-tier optimization.
+
     A configuration for how to compute a total objective over a list of
     objectives. A tier is meant to be calculated for 1 or more force fields,
     which is what happens during parameter searching. Tiers have a fitting
     aspect that is configured here. For example, an objective might only
     perform 1 or 2 fitting steps for each force field, and the top N force
-    fields would be accepted (and passed to the next tier.
+    fields would be accepted (and passed to the next tier).
+
+    Attributes
+    ----------
+    objectives: Dict[int, objective_config]
+        The objective functions that should be optimized in this tier. Note that
+        each molecule in the graph database requires its own entry in this dict.
+    fit_models: list[int] | None
+        Which models in the chemical system to fit with this tier. Fits all
+        models if ``None``. See the ``models`` attribute of
+        :py:class:`besmarts.mechanics.molecular_models.chemical_system`.
+    fit_symbols: list[str] | None
+        Which symbols of the fitted models to fit with this tier. Fits all
+        symbols if ``None``. See the ``symbol`` attribute of
+        :py:class:`besmarts.mechanics.molecular_models.chemical_model`.
+    step_limit: int | None
+        Stop after this many iterations of this tier.
+    accept: int
+        Pass the best this many candidates on to the next tier. Passes all
+        candidates on if 0.
+
+
+    .. TODO: Why is the objectives attribute a dict rather than a list?
     """
 
     def __init__(self):
@@ -2962,13 +3007,45 @@ def fit(csys, gdb, objective, psystems, nodes, wq=None, verbose=False):
     )
     return ret
 
+class ChemicalObjective(Protocol):
+    """A function to compute an objective function from a chemical system.
+
+    See :py:func:`chemical_objective` for an example implementation.
+
+    Parameters
+    ----------
[…]
