### 2025-03-05 3d2b8ef
better handling of dimers
