+++ besmarts-core/python/besmarts/core/assignments.py
@@ -48,6 +48,7 @@ GRID = 9
 ESP = 10
 RADII = 11
 PHYSICAL_MODEL_BOND_LABELS = 12
+SAPT = 13
 
 ASSN_NAMES = {
     POSITIONS: "positions",  # atom, 1x3
@@ -100,6 +101,7 @@ class graph_db_address:
 
 class graph_db_column:
 
+    # each column in a row 
     __slots__ = "selections",
 
     def __init__(self):
@@ -115,6 +117,7 @@ class graph_db_column:
 
 class graph_db_row:
 
+    # each row is an non-interacting system
     __slots__ = "columns",
 
     def __init__(self):
@@ -431,6 +434,12 @@ class graph_assignment_float(smiles_assignment):
         self.graph: str = graph
         self.selections: Dict[Sequence[int], List[float]] = selections
 
+class graph_assignment_float_matrix(smiles_assignment):
+    __slots__ = "graph", "selections"
+
+    def __init__(self, graphs, selections):
+        self.graph: List[graphs.graph] = graphs
+        self.selections: Dict[Sequence[int], List[float]] = selections
 
 class smarts_assignment:
     __slots__ = "smarts", "selections"
@@ -468,7 +477,25 @@ class pair_assignment_float(structure_assignment_float):
         self.topology = topology.pair
         self.selections = selections
 
+class pair_assignment_float_matrix(structure_assignment_float):
+    __slots__ = "topology", "selections"
+    def __init__(self, selections):
+        self.topology = topology.pair
+        self.selections = selections
+
+class bond_assignment_float_matrix(structure_assignment_float):
+    __slots__ = "topology", "selections"
+    def __init__(self, selections):
+        self.topology = topology.bond
+        self.selections = selections
+
 class angle_assignment_float(structure_assignment_float):
+    __slots__ = "topology", "selections"
+    def __init__(self, selections):
+        self.topology = topology.angle
+        self.selections = selections
+
+class angle_assignment_float_matrix(structure_assignment_float):
     __slots__ = "topology", "selections"
     def __init__(self, selections):
         self.topology = topology.angle
@@ -480,7 +507,19 @@ class torsion_assignment_float(structure_assignment_float):
         self.topology = topology.torsion
         self.selections = selections
 
+class torsion_assignment_float_matrix(structure_assignment_float):
+    __slots__ = "topology", "selections"
+    def __init__(self, selections):
+        self.topology = topology.torsion
+        self.selections = selections
+
 class outofplane_assignment_float(structure_assignment_float):
+    __slots__ = "topology", "selections"
+    def __init__(self, selections):
+        self.topology = topology.outofplane
+        self.selections = selections
+
+class outofplane_assignment_float_matrix(structure_assignment_float):
     __slots__ = "topology", "selections"
     def __init__(self, selections):
         self.topology = topology.outofplane
@@ -602,8 +641,8 @@ def smiles_assignment_geometry_distances(
     indices
 ) -> bond_assignment_float:
 
-    if indices is None:
-        indices = graphs.graph_bonds(pos.graph)
+    # if indices is None:
+    #     indices = graphs.graph_bonds(pos.graph)
 
     xyz = pos.selections
     selections = {}
[…]
