+++ besmarts-core/python/besmarts/mechanics/fits.py
@@ -71,71 +71,74 @@ class compute_config:
         self.method_vib_modes = "min"
         self.analytic = False
         self.analytic_use_gradients = True
+        self.dimer_start = 0
 
     def run(self) -> List[Dict[assignments.tid_t, assignments.graph_db_table]]:
 
+        return []
+
+    def prepare_systems(self, keyset=None):
+
+        gdb = self.GDB
         csys = self.csys
-        GDB = self.GDB
-        tbl_idx = assignments.ENERGY
-        tid = assignments.POSITIONS
+        psys = None
+        traj = []
+        if keyset is None:
+            keyset = {}
 
-        all_results = []
-        for keyset in self.keys:
-            results = []
-            results: List[Dict[assignments.tid_t, assignments.graph_db_table]]
+        for eid, gde in gdb.entries.items():
 
-            for eid, gdb in GDB.entries.items():
-                tbl = assignments.graph_db_table(topology.null)
-                for gid in gdb.graphs:
-                    rid = 0
-                    pos0 = assignments.graph_db_graph_to_graph_assignment(
-                        gdb,
-                        eid,
-                        tid,
-                        gid,
-                        rid
-                    )
+            system = assignments.graph_db_graph_to_graph_assignment(
+                gdb,
+                eid,
+                POSITIONS,
+            )
+            traj.append(system)
 
-                    psys = mm.chemical_system_to_physical_system(
-                        csys,
-                        [pos0],
-                        ref=self.psys,
-                        reuse=self.reuse
-                    )
-                    reapply = set()
-                    for k, v in keyset.items():
-                        mm.physical_system_set_value(psys, k, v)
-                        reapply.add(k[0])
-                    for m in reapply:
-                        procs = csys.models[m].procedures
-                        if len(procs) > 1:
-                            for _ in range(1, len(psys.models[m].values)):
-                                psys.models[m].values.pop()
-                                psys.models[m].labels.pop()
-                            for proc in procs[1:]:
-                                psys.models[m] = proc.assign(
-                                    csys.models[m],
-                                    psys.models[m],
-                                    overrides={
-                                        k[1:]: v for k, v in keyset.items()
-                                        if k[0] == m
-                                    }
-                                )
-                    ene = objectives.physical_system_energy(psys, csys)
+            if psys is None:
+                psys = mm.chemical_system_to_physical_system(
+                    csys,
+                    system,
+                    ref=self.psys[eid],
+                    reuse=self.reuse
+                )
+                reapply = set()
+                for k, v in keyset.items():
+                    mm.physical_system_set_value(psys, k, v)
+                    reapply.add(k[0])
+                for m in reapply:
+                    procs = csys.models[m].procedures
+                    if len(procs) > 1:
+                        for _ in range(1, len(psys.models[m].values)):
+                            psys.models[m].values.pop()
+                            psys.models[m].labels.pop()
+                        for proc in procs[1:]:
+                            psys.models[m] = proc.assign(
+                                csys.models[m],
+                                psys.models[m],
+                                overrides={
+                                    k[1:]: v for k, v in keyset.items()
+                                    if k[0] == m
+                                }
+                            )
 
-                    tbl[gid] = ene
-                r = {tbl_idx: tbl}
-                # print("Done calculating", self)
-                # print("Result is\n", r)
-                results.append(r)
-            all_results.append(results)
-        return all_results
+                if self.fc_reset_config:
+                    if self.fc_reset_config.get("unique", False):
+                        psys = reset_unique(
+                            self.fc_reset_config,
+                            gde,
+                            psys
+                        )
[…]

[…]
