+++ besmarts-core/python/besmarts/core/clusters.py
@@ -249,12 +249,12 @@ def perform_operations(
     nodes = {}
     for cnd_i, key in keys.items():
         (S, Sj, step, _, _, _, _) = candidates[key]
-        (edits, _, p_j) = key
+        (edits, _, p_j, oper) = key
         param_name = "p."
         sma = ""
         added = False
 
-        if step.operation == strategy.SPLIT:
+        if oper == strategy.SPLIT:
             # param_name = prefix + str(group_number)
             i = max(hidx.index.nodes) + 1
             name = f"{prefix}{i}"
@@ -287,7 +287,7 @@ def perform_operations(
             # print(datetime.datetime.now(), '*** 4')
 
             #####
-        elif step.operation == strategy.MERGE:
+        elif oper == strategy.MERGE:
             # if (S.name not in groups) or (Sj.name not in groups):
             #     continue
 
@@ -791,6 +791,7 @@ def smarts_clustering_optimize(
                         return_matches = config.splitter.return_matches
                         config.splitter.return_matches = True
 
+                        print("PRIMITIVES IS", config.splitter.primitives)
                         ret = splits.split_structures_distributed(
                             config.splitter,
                             S0,
@@ -808,7 +809,7 @@ def smarts_clustering_optimize(
                         )
 
                         new_candidates = clustering_collect_split_candidates_serial(
-                            S, ret, step
+                            S, ret, step, strategy.SPLIT
                         )
                         union_cache[(S.index, S.name, strategy.cursor)] = (Q, new_candidates)
 
@@ -817,7 +818,7 @@ def smarts_clustering_optimize(
                 if candidates:
                     p_j_max = max(x[2] for x in candidates) + 1
                 for k, v in new_candidates.items():
-                    k = (k[0], k[1], k[2]+p_j_max)
+                    k = (k[0], k[1], k[2]+p_j_max, step.operation)
                     candidates[k] = v
                 new_candidates = None
 
@@ -825,7 +826,7 @@ def smarts_clustering_optimize(
                 if candidates:
                     p_j_max = max(x[2] for x in candidates) + 1
                 for k, v in new_candidates_direct.items():
-                    k = (k[0], k[1], k[2]+p_j_max)
+                    k = (k[0], k[1], k[2]+p_j_max, step.operation)
                     candidates[k] = v
                 new_candidates_direct = None
 
@@ -835,7 +836,7 @@ def smarts_clustering_optimize(
                     J = cst.hierarchy.index.nodes[jidx]
 
                     for overlap in step.overlap:
-                        key = (overlap, macro.cursor, p_j)
+                        key = (overlap, macro.cursor, p_j, strategy.MERGE)
                         cnd = (S, J, step, None, None, None, None)
                         candidates[key] = cnd
 
@@ -852,14 +853,14 @@ def smarts_clustering_optimize(
         Sj_sma = []
 
         with multiprocessing.pool.Pool(configs.processors) as pool:
-            if step.operation == strategy.SPLIT:
-                # Sj_lst = [candidates[x[1]][1] for x in pq]
-                Sj_lst = [graphs.subgraph_as_structure(x[1], topo) for x in candidates.values()]
-            elif step.operation == strategy.MERGE:
-                Sj_lst = [
-                    graphs.subgraph_as_structure(cst.hierarchy.subgraphs[x[1].index], topo)
-                    for x in candidates.values()
-                ]
+            Sj_lst = []
+            for (_,_,_, oper), x in candidates.items():
+                if oper == strategy.SPLIT:
+                    Sj_lst.append(graphs.subgraph_as_structure(x[1], topo))
[…]
