+++ besmarts-core/python/besmarts/mechanics/force_pairwise.py
@@ -10,54 +10,139 @@ import shutil
 
 from besmarts.core import topology
 from besmarts.core import assignments
-from besmarts.core import trees
+# from besmarts.core import trees
 from besmarts.core import codecs
-from besmarts.core import hierarchies
+# from besmarts.core import hierarchies
 from besmarts.core import primitives
-from besmarts.core import graphs
-from besmarts.core import perception
+# from besmarts.core import graphs
+# from besmarts.core import perception
 
 from besmarts.mechanics import molecular_models as mm
 
-
+# import math
+INF = math.inf
 
 # electrostatics
 def energy_function_coulomb_mix(*, eps, c, s, qq, x) -> float:
-    return [[s[0]*eps[0]*q/xi if xi < c[0] else 0.0 for q in qq] for xi in x[0]]
+    return [[
+        INF if xi == 0.0 else
+        s[0]*eps[0]*q/xi if xi < c[0] else
+        0.0 for q in qq
+    ] for x0 in x for xi in x0]
 
 def force_function_coulomb_mix(*, eps, c, s, qq, x) -> float:
-    return [[s[0]*eps[0]*q/(xi*xi) if xi < c[0] else 0.0 for q in qq] for xi in x[0]]
+    return [[
+        INF if xi == 0.0 else
+        s[0]*eps[0]*q/(xi*xi) if xi < c[0] else
+        0.0 for q in qq
+    ] for x0 in x for xi in x0]
 
 def force_gradient_function_coulomb_mix(*, eps, c, s, qq, x) -> float:
-    return [[(q*-s[0]*eps[0]*2/(xi*xi*xi)) if xi < c[0] else 0.0 for q in qq] for xi in x[0]]
+    return [[
+        INF if xi == 0.0 else
+        (q*-s[0]*eps[0]*2/(xi*xi*xi)) if xi < c[0] else
+        0.0 for q in qq
+    ] for x0 in x for xi in x0]
 
 def force_system_coulomb_mix(*, eps, c, s, qq, x) -> float:
-    return [[(q, s[0]*eps[0]/(xi*xi)) if xi < c[0] else 0.0 for q in qq] for xi in x[0]]
+    return [[
+        (q, INF) if xi == 0.0 else
+        (q, s[0]*eps[0]/(xi*xi)) if xi < c[0] else
+        0.0 for q in qq
+    ] for x0 in x for xi in x0]
 
 def force_gradient_system_coulomb_mix(*, eps, c, s, qq, x) -> float:
-    return [[(q, -s[0]*eps[0]*2/(xi*xi*xi)) if xi < c[0] else 0.0 for q in qq] for xi in x[0]]
-
+    return [
+        [
+            (q, INF) if xi == 0.0 else
+            (q, -s[0]*eps[0]*2/(xi*xi*xi)) if xi < c[0] else
+            0.0 for q in qq
+        ]
+        for xi in x[0]
+    ]
 
 # vdW
 def energy_function_lennard_jones_combined(*, s, c, ee, rr, x) -> float:
-    xx = [[math.pow(ri/xi, 6) if xi < c[0] else 0.0 for ri in rr] for xi in x[0]]
-    return [[4.0*s[0]*ei*(xi*xi - xi) for ei, xi in zip(ee, xxi)] for xxi in xx]
+    xx = [
+        [
+            INF if xi == 0.0 else
+            math.pow(ri/xi, 6) if xi < c[0] else
+            0.0 for ri in rr
+        ] for x0 in x for xi in x0
+    ]
+    return [
+        [4.0*s[0]*ei*(xi*xi - xi) for ei, xi in zip(ee, xxi)]
[…]
