+++ besmarts-openmm/python/besmarts/mechanics/optimizers_openmm.py
@@ -28,29 +28,32 @@ PRECISION = configs.precision
 
 
 def physical_system_to_openmm_system(psys):
-    pos = None
-    for m in psys.models:
-        if m is None:
-            continue
-        pos = m.positions[0]
-        break
+
     system = openmm.openmm.System()
     atom_map = {}
 
-    for i, node in sorted(pos.graph.nodes.items(), key=lambda x: x[0]):
-        n = node.primitives['element'].on()[0]
-        m = openmm.app.element.Element.getByAtomicNumber(n)
-        atom_map[i] = system.addParticle(m.mass)
+    pos = psys.models[0].positions
+    for pi, posi in enumerate(pos):
+        for i, node in sorted(posi.graph.nodes.items(), key=lambda x: x[0]):
+            n = node.primitives['element'].on()[0]
+            m = openmm.app.element.Element.getByAtomicNumber(n)
+            for c, conf in enumerate(posi.selections[i,]):
+                atom_map[(pi, i, c)] = system.addParticle(m.mass)
+                # print(f"Ptl: {atom_map[(pi, i, c)]} {(pi, i, c)}")
 
+    # print("Bonds...")
     frc, b_map = assign_bonds(psys, atom_map)
     system.addForce(frc)
 
+    # print("Angles...")
     frc, a_map = assign_angles(psys, atom_map)
     system.addForce(frc)
 
+    # print("Torsions...")
     frc, t_map = assign_torsions(psys, atom_map)
     system.addForce(frc)
 
+    # print("OOP...")
     frc, i_map = assign_outofplanes(psys, atom_map)
     system.addForce(frc)
 
@@ -58,14 +61,18 @@ def physical_system_to_openmm_system(psys):
     # e_map = assign_scales_nonbonded(psys, atom_map, frc)
     # frc.setNonbondedMethod(frc.NoCutoff)
 
+    # print("Electostatics...")
     frc, q_map = assign_elec(psys, atom_map)
     frc.setNonbondedMethod(frc.NoCutoff)
-    eq_map = assign_scales_elec(psys, atom_map, frc)
+    # print("Electostatic exceptions...")
+    frc, eq_map = assign_scales_elec(psys, atom_map, frc)
     system.addForce(frc)
 
+    # print("LJ...")
     frc, lj_map = assign_lj(psys, atom_map)
     frc.setNonbondedMethod(frc.NoCutoff)
-    elj_map = assign_scales_lj(psys, atom_map, frc)
+    # print("LJ Exceptions...")
+    frc, elj_map = assign_scales_lj(psys, atom_map, frc)
     system.addForce(frc)
 
     for i, f in enumerate(system.getForces()):
@@ -75,14 +82,15 @@ def physical_system_to_openmm_system(psys):
     chain = topo.addChain()
     res = topo.addResidue("MOL", chain)
 
+    # print("Bonding...")
     atoms = {}
     for i, j in sorted(atom_map.items(), key=lambda x: x[1]):
-        n = pos.graph.nodes[i].primitives['element'].on_first()
+        n = pos[i[0]].graph.nodes[i[1]].primitives['element'].on_first()
         elem = openmm.app.element.Element.getByAtomicNumber(n)
         atoms[j] = topo.addAtom(elem.symbol, elem, res, id=j)
 
     for i, j in sorted(b_map.items(), key=lambda x: x[1]):
-        ic = atoms[atom_map[i[0]]], atoms[atom_map[i[1]]]
+        ic = atoms[atom_map[(i[0], i[1][0], i[2])]], atoms[atom_map[(i[0], i[1][1], i[2])]]
         topo.addBond(*ic)
 
     integ = openmm.openmm.VerletIntegrator(1.0)
[…]
