+++ besmarts-core/python/examples/11_aicp.py
@@ -33,7 +33,7 @@ from besmarts.codecs import codec_rdkit
 from besmarts.assign import hierarchy_assign_rdkit
 
 configs.compute_task_chunksize = 1000
-configs.processors = 16
+configs.processors = 1
 configs.remote_compute_enable = False
 # To use additional workers, from a terminal run (on the separate work machine)
 # python -m besmarts.worker <IP> 55555 np
@@ -68,7 +68,7 @@ def mystrategy():
     # bits, and do not look at neighboring atoms.
 
     splitter = configs.smarts_splitter_config(
-        1, 1, 0, 0, 0, 0, True, True, 0, False, True, True, True
+        1, 1, 0, 0, 0, 0, True, True, 0, False, True, True, True, ["element"]
     )
 
     # Disable general patterns like [!#6][!#8]. In general these will lead
@@ -132,9 +132,11 @@ def configure_tiers(objs, penalties, fit_models, fit_symbols):
     initial.anneal = 0
     initial.step_limit = 200
     initial.method = 'L-BFGS-B'
+    # initial.method = 'trust-ncg'
+    initial.minstep = 1e-5
 
     # line search in LBFGS
-    initial.maxls = 20
+    initial.maxls = 100
     initial.fit_models = fit_models
     initial.fit_symbols = fit_symbols
 
@@ -171,23 +173,28 @@ def configure_tiers(objs, penalties, fit_models, fit_symbols):
 
     # There will be 10 (serial) iterations of full fits, each with a randomly
     # kicked set of parameters.
-    final.anneal = 10
-    final.maxls = 20
+    final.anneal = 0
+    final.maxls = 200
+    final.minstep = 1e-3
 
     final.bounds = initial.bounds
     final.priors = initial.priors
 
     final.method = "L-BFGS-B"
+    # final.method = 'trust-ncg'
 
     tier = fits.objective_tier()
     tier.objectives = final.objectives
 
     # Only perform two optimization steps before scoring
     tier.step_limit = 2
+    tier.minstep = 1e-2
 
     # Pass the best 5 candidates
     tier.accept = 5
-    tier.maxls = 10
+    tier.maxls = 20
+
+    tier.method = "L-BFGS-B"
     tier.priors = final.priors
     tier.bounds = initial.bounds
     tier.fit_models = fit_models
@@ -197,7 +204,8 @@ def configure_tiers(objs, penalties, fit_models, fit_symbols):
     tier.penalties.extend(penalties)
     final.penalties.extend(penalties)
 
-    tiers = [tier]
+    #tiers = [tier]
+    tiers = []
 
     return initial, tiers, final
 
@@ -210,14 +218,17 @@ def configure_chem_perception_strat(csys, split_models):
     strat.enable_split = 1
     strat.enable_merge = 1
 
+    for m in split_models:
+        strat.bounds[m].splitter.primitives = ['element']
+
     # Don't scan periodicity modifications since we estimate them each fit
     strat.enable_modify = 0
 
     # When scoring, reset the torsions
-    strat.enable_dihedral_periodicity_reset = 1
+    strat.enable_dihedral_periodicity_reset = 0
     strat.dihedral_periodicity_reset_alpha = -.5
     strat.dihedral_periodicity_reset_max_n = 3
-    strat.dihedral_periodicity_reset_max_k = 5
+    strat.dihedral_periodicity_reset_max_k = 10
     strat.dihedral_periodicity_reset_min_k = 1e-3
 
     # For each iteration, accept only 1 split
@@ -258,10 +269,13 @@ def configure_objectives(opt_mols):
                 scale=1e-3,
                 include=True
         )
-        o.batch_size = 1
+        o.batch_size = None
         # o.freq_range = (500, 600)
         o.method_vib_modes = "qm"
[…]
