### 2025-03-05 #6653 @trevorgokey
After a number of hurdles, I believe everything works by installing from PyPI. The main README.md has been updated with instructions. Ambertools is only available via conda, so the best way to install everything is
```
conda create -n besmarts -c conda-forge python ambertools
conda activate besmarts
pip install besmarts[rdkit,scipy,openmm]
```
As an implementation detail, I separated out `besmarts.mechanics` into a separate `besmarts-mechanics` package since parts of that depend on Ambertools, whereas `besmarts.core` does not.
