### 2025-03-10 #6653 @lucydot
Hi @trevorgokey. I've ran through the install and done some testing:

- The installation instructions on the README page and RTD (https://besmarts.readthedocs.io/en/latest/installation.html) are different. I guess you should update the latter so that it is a pip install from PyPI?
- When I run the install instructions as written in zsh I get `zsh: no matches found: besmarts[rdkit,scipy,openmm]`. I need to put it into quotes to work: `pip install "besmarts[rdkit,scipy,openmm]"`
- When running through the tutorials I could not find a way to copy and paste code blocks. There are extensions available for this (https://github.com/choldgraf/sphinx-copybutton). It may be particularly useful for people who want to check their local install is working as expected.

The last point does not need to be addressed before publication, but is a suggested nice-to-have.
