### 2025-03-12 211222b
Structural fixes for docstrings
