### 2025-03-12 136b5f8
Structural fixes for docstrings
