### 2025-03-12 f6b015a
Structural fixes for docstrings
