+++ besmarts-mechanics/python/besmarts/mechanics/fits.py
@@ -758,26 +758,16 @@ class objective_config:
     Configuration of an objective function for force field fitting.
 
     The objective has two parts. The first part is how the properties are
-    calculated and is configured with a :py:class:`compute_config` as returned
-    by the :py:meth:`get_task` method. The second is how the
+    calculated and is configured with a ``compute_config`` as returned
+    by the ``get_task`` method. The second is how the
     objective is computed from the properties and is configured here.
 
     This is a base class and should generally not be instantiated directly;
-    instead, see the following classes for commonly used objectives:
-
-    :py:class:`objective_config_position`
-        Performs a geometry optimization and calculates the sum of squared
-        error (SSE). The root of the mean SSE is the RMSD.
-    :py:class:`objective_config_gradient`
-        The geometry is held fixed at the reference, and calculates the
-        objective as the SSE of the difference in QM/MM forces.
-
-    .. todo: More of the above, write a Parameters section
+    instead, see the following classes for commonly used objectives.
 
     See Also
     --------
-    objective_config_energy, objective_config_hessian, objective_config_penalty,
-    objective_config_position, objective_config_gradient
+    compute_config, get_task
 
     """
 
@@ -960,7 +950,10 @@ class objective_config:
 
 
 class objective_config_gradient(objective_config):
-
+    """
+    The geometry is held fixed at the reference, and calculates the
+    objective as the SSE of the difference in QM/MM forces.
+    """
     def __init__(self, *args, **kwds):
         super().__init__(*args, **kwds)
         self.enable_minimization = False
@@ -1425,7 +1418,7 @@ class objective_config_hessian(objective_config):
     #         #         g, hess_mm, grad_mm, DL2, ics, B, B2)
     #         #     )
     #         #     x0.extend([0]*(len(x0) - len(x1)))
-                
+
 
     #         # keep this for when I want to try internal hess
     #         # x0 = array_numpy.dlcmatrix_project_gradients(pos, x0)
@@ -1949,7 +1942,7 @@ class objective_config_energy_total(objective_config):
             dxdp, d2xdp2 = finite_difference_forward_1(f, h)
         elif self.grad_mode == "c2":
             dxdp, d2xdp2 = finite_difference_central_2(f, h)
-            
+
         return dxdp, d2xdp2
 
 
@@ -2037,7 +2030,7 @@ class objective_config_energy_total(objective_config):
         #     tbl = tbls[tid]
         #     x1 = tbl.values
         #     x1 = arrays.array_translate(x1, -min_ene)
-            
+
         #     obj.extend(arrays.array_difference(x1, x0))
             # ene_dict = tbl.values[0][0]
             # tot_dict = tbl.values[0][1]
@@ -2114,6 +2107,10 @@ class objective_config_energy_total(objective_config):
 
 
 class objective_config_position(objective_config):
+    """
+    Performs a geometry optimization and calculates the sum of squared
+    error (SSE). The root of the mean SSE is the RMSD.
+    """
 
     def get_task(
         self,
@@ -2240,14 +2237,12 @@ class objective_tier:
     objectives: Dict[int, objective_config]
         The objective functions that should be optimized in this tier. Note that
         each molecule in the graph database requires its own entry in this dict.
-    fit_models: list[int] | None
+    fit_models: List[int] | None
         Which models in the chemical system to fit with this tier. Fits all
-        models if ``None``. See the ``models`` attribute of
-        :py:class:`besmarts.mechanics.molecular_models.chemical_system`.
-    fit_symbols: list[str] | None
[…]
