### 2025-03-14 3c87408
Add docstring of generate_candidates
