### 2025-03-14 898c2db
Add docstring of generate_candidates
