+++ doc/conf.py
@@ -42,6 +42,7 @@ extensions = [
    'sphinx.ext.autodoc',
    'sphinx.ext.autosummary',
    'sphinx.ext.napoleon',
+   'sphinx_copybutton',
 #   'IPython.sphinxext.ipython_console_highlighting',
 #   'IPython.sphinxext.ipython_directive',
 ]
