+++ doc/usage.rst
@@ -8,14 +8,6 @@ Usage/Tutorial
    graph_io.rst
    labeling.rst
    splitting.rst
-   mapping.rst
    bitwise_operations.rst
-   internal_coordinates.rst
-   assignments.rst
-   conformation_loading.rst
-   energy.rst
-   force.rst
    clustering.rst
-   smirnoff.rst
    fffit.rst
-   compute.rst

[…]
