### 2025-03-25 3fea701
better key filtering
