+++ besmarts-mechanics/python/besmarts/mechanics/fits.py
@@ -619,7 +619,7 @@ class compute_config_hessian(compute_config):
                     HESSIANS: tbl_hess,
                     GRADIENTS: tbl_grad
                 }
-                if not self.enable_minimization:
+                if not (self.enable_minimization or self.method_vib_modes == "min"):
                     r.pop(POSITIONS)
 
                 results.append(r)
@@ -1067,10 +1067,11 @@ class objective_config_gradient(objective_config):
                         x1 = [arrays.array_magnitude(x1[i:i+3]) for i in range(0, len(x1), 3)]
 
                 x = arrays.array_difference(x1, x0)
+
                 xcart = arrays.array_difference(x1cart, x0cart)
                 nlines = len(out)
                 if self.verbose > 0:
-                    mode = "magnitude" if self.force_mode == "magnitude" else ""
+                    mode = self.force_mode
                     if self.verbose > 1:
                         out.append(f" EID {eid} TID {tid} Gradient {mode} objective:")
                     if mode == "magnitude":
@@ -1093,9 +1094,10 @@ class objective_config_gradient(objective_config):
                             totalxx = arrays.array_add(totalxx, xx)
                             xyz1 = " ".join([f"{xi:12.4e}" for xi in x1cart[3*i:3*i+3]])
                             xyz0 = " ".join([f"{xi:12.4e}" for xi in x0cart[3*i:3*i+3]])
-                            dxyz = " ".join([f"{xi:12.4e}" for xi in xx])
+                            dxyz = " ".join([f"{xi:12.4e}" for xi in x[3*i:3*i+3]])
+                            mag = f"{arrays.array_magnitude(x[3*i:3*i+3]):12.4e}" 
                             if self.verbose > 1:
-                                out.append(f"   {i+1:4d} MM: {xyz1} QM: {xyz0} dG: {dxyz}")
+                                out.append(f"   {i+1:4d} MM: {xyz1} QM: {xyz0} dG: {dxyz} mag: ")
                             rmse.append(arrays.array_inner_product(xx, xx))
                     # if self.verbose > 1:
                     #     xyz1 = " ".join([f"{xi:12.4e}" for xi in totalx1])
@@ -1224,9 +1226,16 @@ class objective_config_hessian(objective_config):
 
 
                 xyz = []
-                for pi, posi in enumerate(pos):
-                    for xyzi in posi.selections.values():
-                        xyz.extend(xyzi)
+
+                if self.method_vib_modes in ["min"]:
+                    for pi, posi in enumerate(etbls[POSITIONS].values):
+                        for xyzi in posi.selections.values():
+                            xyz.extend(xyzi)
+
+                else:
+                    for pi, posi in enumerate(pos):
+                        for xyzi in posi.selections.values():
+                            xyz.extend(xyzi)
 
                 sym = [s for posi in pos for s in graphs.graph_symbols(posi.graph).values()]
                 mass = np.array([[vibration.mass_table[s]]*3 for s in sym])
@@ -1234,6 +1243,7 @@ class objective_config_hessian(objective_config):
                 hess_mm = etbls[HESSIANS].values
 
                 if self.method_vib_modes in ["min", "mm", "mm+"]:
+
 
                     remove = 0
                     if self.method_vib_modes == "mm":
@@ -1521,7 +1531,7 @@ class objective_config_hessian(objective_config):
                         f"FC reset mode is {self.fc_reset_config}"
                     )
 
-                if self.method_vib_modes in ["min", "mm", "mm+"]:
+                if self.method_vib_modes in ["min", "mm", "mm+", "min+"]:
[…]
