### 2025-03-26 fce4b84
Comments for addition to forcefield_optimization_strategy docstring
