### 2025-04-08 93742d9
chemical_objective and so forth
