### 2025-04-09 28eebd1
improve hessian projections
