+++ besmarts-mechanics/python/besmarts/mechanics/hessians.py
@@ -5,6 +5,7 @@ besmarts.mechanics.hessians
 import copy
 import numpy as np
 import pprint
+import time
 
 from besmarts.core import assignments
 from besmarts.core import graphs
@@ -259,6 +260,8 @@ def hessian_project_onto_ics(
 ) -> dict:
 
     pos = psys.models[0].positions
+    g = pos[0].graph
+    posi = 0
 
     xyz = np.vstack([x[0] for posi in pos for x in posi.selections.values()], dtype=float)
     xyz = xyz.round(PRECISION)
@@ -270,8 +273,8 @@ def hessian_project_onto_ics(
     torsions = True
     outofplanes = True
     pairs = False
-
-    hess_qm_freq, hess_qm_modes = vibration.hessian_modes(
+    
+    hess_qm_freq, hess_qm_modes, DL = vibration.hessian_modes(
         hess_qm,
         sym,
         xyz,
@@ -279,7 +282,8 @@ def hessian_project_onto_ics(
         0,
         remove=0,
         stdtr=True,
-        verbose=False
+        verbose=False,
+        return_DL=True
     )
     omega = np.round(hess_qm_freq, PRECISION)
 
@@ -299,27 +303,514 @@ def hessian_project_onto_ics(
         ics, B = B
         B = np.array(B)
 
+    # icnb, bnb = assignments.bmatrix(
+    #     pos,
+    #     torsions=False,
+    #     outofplanes=False,
+    #     pairs=True,
+    #     remove1_3=True
+    # )
     ic_qm_fcs = {}
     if len(B.shape) > 1 and B.shape[0] > 0 and B.shape[1] > 0:
-        hess_qm_ic = project_ics(B, hess_qm)
 
+        # psys_hq = physical_system_force_gradient_internal(psys, csys)
+        hess_qm_ic = project_ics(B, hess_qm)
         hess_qm_ic = np.array(np.diag(hess_qm_ic))
+        # hess_qm_nb = np.array(np.diag(project_ics(bnb, hess_qm)))
+        A = []
+        C = []
+        version = 2
+
+        if version in [1,2]:
+            qic = [tuple((posi, x) for x in y) for y in graphs.graph_bonds(g)]
+            hqic = [x for ic, x in zip(ics, hess_qm_ic) if ic in qic]
+            bqic = np.array([x for ic, x in zip(ics, B) if ic in qic])
+            hx_bonds1 = np.dot(bqic.T, np.dot(np.diag(hqic), bqic)).tolist()
+            # hx_bonds0 = np.dot(bqic.T, bqic)
+
+            qic = graphs.graph_angles(g)
+            qic = [tuple((posi, x) for x in y) for y in graphs.graph_angles(g)]
+            hqic = [x for ic, x in zip(ics, hess_qm_ic) if ic in qic]
+            bqic = np.array([x for ic, x in zip(ics, B) if ic in qic])
+            hx_angles1 = np.dot(bqic.T, np.dot(np.diag(hqic), bqic)).tolist()
+            # hx_angles0 = np.dot(bqic.T,  bqic)
+
+            qic = [tuple((posi, x) for x in y) for y in graphs.graph_torsions(g)]
+            hqic = [x for ic, x in zip(ics, hess_qm_ic) if ic in qic]
+            bqic = np.array([x for ic, x in zip(ics, B) if ic in qic])
+            hx_torsions1 = np.dot(bqic.T, np.dot(np.diag(hqic), bqic)).tolist()
+            # hx_torsions0 = np.dot(bqic.T,  bqic)
+
+            qic = [tuple((posi, x) for x in y) for y in graphs.graph_outofplanes(g)]
[…]
