+++ besmarts-mechanics/python/besmarts/mechanics/fits.py
@@ -2273,6 +2273,7 @@ class objective_tier:
         self.minstep = 10**(-PRECISION)
         self.ftol = 1e-5
         self.gtol = 1e-5
+        self.score_mode = "objective" # "max_gradient"
 
     def key_filter(self, x):
 
@@ -2468,6 +2469,8 @@ def objective_tier_run(
         minstep
     )
 
+    
+    force_grad = (ot.score_mode == "max_gradient" and ot.step_limit == 0)
     ret = optimizers_scipy.optimize_forcefield_gdb_scipy(
         x0,
         args,
@@ -2476,19 +2479,43 @@ def objective_tier_run(
         maxls=ot.maxls,
         ftol=ot.ftol,
         gtol=ot.gtol,
-        anneal=ot.anneal
+        anneal=ot.anneal,
+        force_grad=force_grad
     )
 
-    result, y0, y1, gx = ret.value
+    result, y0, y1, g0, g1 = ret.value
 
-    ret.out.append(f">>> Initial Objective {y0:10.5g}")
-    ret.out.append(f">>> Final Objective   {y1:10.5g}")
-    if y0 > 0.0:
-        ret.out.append(f">>> Percent change    {(y1-y0)/y0*100:10.5g}%")
+    if ot.score_mode == "objective":
+        ret.out.append(f">>> Initial Objective {y0:10.5g}")
+        ret.out.append(f">>> Final Objective   {y1:10.5g}")
+        if y0 > 0.0:
+            ret.out.append(f">>> Percent change    {(y1-y0)/y0*100:10.5g}%")
 
-    if verbose:
-        for i in [-3, -2, -1]:
-            print(ret.out[i])
+        if verbose:
+            for i in [-3, -2, -1]:
+                print(ret.out[i])
+
+    elif ot.score_mode == "max_gradient":
+
+        if ot.step_limit == 0:
+            i0 = 0
+            i1 = 0
+        else:
+            i0 = arrays.argmin(arrays.array_abs(g0))[-1]
+            i1 = arrays.argmin(arrays.array_abs(g1))[-1]
+        y0 = abs(g0[i0])
+        y1 = abs(g0[i1])
+        ret.out.append(f">>> Initial max gradient {y0:10.5g} parameter {i0}")
+        ret.out.append(f">>> Final max gradient   {y1:10.5g} parameter {i1}")
+        if y0 > 0.0:
+            ret.out.append(f">>> Percent change    {(y1-y0)/y0*100:10.5g}%")
+
+        if verbose:
+            for i in [-3, -2, -1]:
+                print(ret.out[i])
+        y0 = -abs(g0[i0])
+        y1 = -abs(g0[i1])
+
 
     if ws:
         if wq:
@@ -2498,7 +2525,7 @@ def objective_tier_run(
 
     kv = {k: v*p[1] + p[0] for k, v, p in zip(keys, result, priors)}
 
-    return returns.success((kv, y0, y1, gx), out=ret.out, err=[])
+    return returns.success((kv, y0, y1, g1), out=ret.out, err=[])
 
 
 def objective_run_distributed(obj, shm=None):
@@ -3017,7 +3044,7 @@ def ff_optimize(
             n1 = s.branch_limit
             # probably print the models
             print(
-                f"{cur} {ma_i:3d}:{micro.index:02d}. op={micro.operation:2d} m={m} a={a} b={b0}->{b1} d={d0}->{d1} n={n0}->{n1}"
+                f"{cur} {ma_i:3d}:{micro.index:02d}. op={micro.operation:2d} m={m} a={a} b={b0}->{b1} d={d0}->{d1} n={n0}->{n1} p={s.primitives}"
[…]
