### 2025-04-09 #6653 @trevorgokey
Hi @lucydot apologies for the delay, these notifications slipped through. I've gone through the list. Here are the details:
- Version is 0.1.0
- All authors, IDs, and licenses are consistent
- Zenodo DOI: [10.5281/zenodo.15184074](https://doi.org/10.5281/zenodo.15184074)
