### 2025-04-09 #6653 @lucydot
@editorial set v0.1.0 as version
