### 2025-04-09 #6653 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
