+++ besmarts-core/python/besmarts/core/configs.py
@@ -62,6 +62,11 @@ class mapper_config:
 
 
 class smarts_extender_config:
+    """Configures how many atoms in the reference data should we consider when
+    writing our SMARTS strings
+
+    Determines `structure.select`.
+    """
     __slots__ = ("depth_min", "depth_max", "include_hydrogen")
 
     def __init__(self, depth_min: int, depth_max: int, include_hydrogen: bool):
@@ -76,6 +81,73 @@ class smarts_extender_config:
 
 
 class smarts_splitter_config:
+    """
+    Configuration object for splitting SMARTS patterns.
+
+    Parameters beginning ``bit_search`` choose the number of bits that may be
+    considered in the primitives of a SMARTS pattern ($k$ in eq. 11 of `Gokey
+    and Mobley 2023`_). This is the number of decisions made in each split. When
+    making a pattern more specific, this is the number of bits that should be
+    flipped.
+
+    Parameters beginning ``branch`` configure extending the SMARTS pattern by
+    adding new wildcard atoms that can then be specialized via bit search. Atoms
+    can be added by extending the graph out to ``branch_depth`` from the SMARTS
+    pattern's mapped atoms, which index the atoms that form the internal
+    coordinate. The mapped atoms have a depth of 0, neighbouring atoms have
+    1, and so forth. Once this search space is defined, the actual number of
+    atoms to keep is set by ``branch_min`` and ``branch_limit``. The bit search
+    is spread across all atoms, so extending the graph by more atoms can dilute
+    this search.
+
+    Attributes
+    ==========
+    bit_search_min
+        Minimum number of decisions made when splitting a SMARTS string.
+    bit_search_limit
+        Maximum number of decisions made when splitting a SMARTS string.
+    branch_min
+        Minimum number of new atoms to add in each split.
+    branch_limit
+        Maximum number of new atoms to add in each split.
+    self.branch_depth_min
+        Minimum ``branch_depth`` of new atoms to consider
+    branch_depth_limit
+        Maximum ``branch_depth`` of new atoms to consider
+    unique
+        If ``True``, do not return multiple splits that produce the same
+        partitioning of the reference structures; otherwise, do not perform this
+        pruning.
+    return_matches
+        Perform substructure matching on each split, or just return them?
+        ``True`` permits an optimization.
+    max_splits
+        How many SMARTS patterns are returned. Low values allow fast returns,
+        but not necessarily the best split.
+    split_general
+        If ``True``, return splits that tend to produce broad hierarchies by
+        turning on bits (using $\hat{b_i}$), otherwise do
+        not return these splits.
+    split_specific
+        If ``True``, return splits that tend to produce deep hierarchies by
+        turning off bits (using the partial inverse of $\hat{b_i}$), otherwise
+        do not return these splits.
+    unique_complements
+        If ``true``, do not return multiple splits that produce complementary
+        partitions of the reference structures; otherwise, do not perform this
+        pruning. Complementary patterns are those that produce the same
+        partition by matching the opposite reference structures.
+    unique_complements_prefer_min
+        If ``True`` and ``unique_complements == True``, keep the complement
+        where the SMARTS pattern matches the fewest structures. If ``False``
+        and ``unique_complements == True``, keep the complement where the SMARTS
+        pattern matches the most structures.
+    primitives
+        Collection of SMARTS primitives that should be split on.
+
+    .. _Gokey and Mobley 2023:
+        https://doi.org/10.26434/chemrxiv-2023-v969f-v3
+    """
     __slots__ = (
         "bit_search_min",
         "bit_search_limit",
@@ -110,21 +182,55 @@ class smarts_splitter_config:
         unique_complements_prefer_min=True,
         primitives=None,
     ):
[…]
