### 2025-04-10 a6da48e
tidy up docstrings in fits.py
