+++ besmarts-mechanics/python/besmarts/mechanics/fits.py
@@ -2297,9 +2297,6 @@ class objective_tier:
         Pass the best this many candidates on to the next tier. Passes all
         candidates on if 0.
 
-
-    .. TODO: Why is the objectives attribute a dict rather than a list?
-
     See Also
     --------
     besmarts.mechanics.molecular_models.chemical_system,
@@ -2647,12 +2644,150 @@ def forcefield_optimization_strategy_default(csys, models=None):
 class forcefield_optimization_strategy(optimization.optimization_strategy):
     """
     Configuration for fitting a force field using BESMARTS (this package).
+
+    Attributes
+    ==========
+
+    bounds
+        Mapping from model numbers (eg, 0 for bonds, 1 for angles, etc) to
+        splitter/extender config
+    direct_enable = False
+        Turn on analytic searching for SMARTS patterns if numerical search finds
+        too few splits
+    direct_limit = 10
+        Do a direct search if numerical search finds fewer than this many splits
+
+    iterative_enable = True
+        Enable numerical search for SMARTS patterns
+
+    enable_merge
+        Enable the MERGE operation during candidate parameter generation.
+    enable_split
+        Enable the SPLIT operation during candidate parameter generation.
+    enable_modify
+        Enable the MODIFY operation during candidate parameter generation.
+
+    modify_outofplane_frequency_limit
+        Set the largest out-of-plane frequency (also known as improper
+        torsion periodicity) that can be generated by the MODIFY operation
+        during parameter generation.
+    modify_torsion_frequency_limit
+        Set the largest torsion frequency (also known as proper torsion
+        periodicity) that can be generated by the MODIFY operation during
+        parameter generation.
+
+    tree_iterator
+        Change this to visit clusters of SMARTS in a particular order, filter
+        nodes, etc.
+
+    macro_accept_max_total
+        How many candidates to pass from one macro step to the next. ``1`` will
+        be fastest, but produce less optimal force fields. ``0`` means
+        unlimited.
+
+    micro_accept_max_total
+        How many candidates to pass from one micro step to the next. ``1`` will
+        be fastest, but produce less optimal force fields. ``0`` means
+        unlimited. Less relevance in force field fitting than other
+        ``accept_max`` options.
+
+    macro_accept_max_per_cluster
+        How many candidates of a given parameter to pass from one macro step to
+        the next. Bounded above by ``macro_accept_max_total``. ``1`` will be
+        fastest, but produce less optimal force fields. ``0`` means limited only
+        by ``macro_accept_max_total``.
+
+    micro_accept_max_per_cluster
+        How many candidates of a given parameter to pass from one micro step to
+        the next. Bounded above by ``micro_accept_max_total``. ``1`` will be
+        fastest, but produce less optimal force fields. ``0`` means limited only
+        by ``micro_accept_max_total``.
+
+    prune_empty
+        If we accept too many operations, some can match nothing due to
+        unexpected occlusion. With this enabled, merging is shortcutted to
+        prevent zero-matching SMARTS from being added.
+
+    enable_reset_bond_lengths
+        If ``True``, initialize bond lengths from the reference data. If
+        ``False``, initialize bond lengths to the parameter value from the
+        parent force field.
+    enable_reset_angle_lengths
+        If ``True``, initialize bond angles from the reference data. If
+        ``False``, initialize bond angles to the parameter value from the
+        parent force field.
+
+    enable_reset_bond_stiffness
+        If ``True``, initialize bond stiffness from the reference data
+        hessians. If ``False``, initialize bond stiffness to the parameter value
+        from the parent force field. Requires hessians in reference data.
+    enable_reset_angle_stiffness
+        If ``True``, initialize angle stiffness from the reference data
[…]
